Aggregation {mlr}	R Documentation

An aggregation method reduces the performance values of the test (and possibly the training sets) to a single value. To see all possible, implemented aggregations look at aggregations.

Description

The aggregation can access all relevant information of the result after resampling and combine them into a single value. Though usually something very simple like taking the mean of the test set performances is done.

Details

Object members:

id [character(1)]

Name of aggregation method.

fun [function(task, perf.test, perf.train, measure, group, pred)]

Aggregation function.

[Package mlr version 1.2 Index]