BenchmarkResult {mlr}	R Documentation

Result of a benchmark run.

Description

Container for results of benchmarked experiments using benchmark. The structure of the object itself is rather complicated, it is recommended to retrive required information via link{getAggrPerformances}, link{getPredictions}, getPerformances, getFeatSelResult, getTuneResult or getFilterResult. Alternatively, you can convert the object using as.data.frame

See Also

Other benchmark: FilterResult, getFilterResult; benchmark; getAggrPerformances; getFeatSelResult; getPerformances; getPredictions; getTuneResult

[Package mlr version 2.0 Index]