Query properties of learners.

Properties can be accessed with getLearnerProperties(learner), which returns a character vector.

The learner properties are defined as follows:

numerics, factors, ordered

Can numeric, factor or ordered factor features be handled?

functionals

Can an arbitrary number of functional features be handled?

single.functional

Can exactly one functional feature be handled?

missings

Can missing values in features be handled?

weights

Can observations be weighted during fitting?

oneclas, twoclass, multiclass

Only for classif: Can one-class, two-class or multi-class classification problems be handled?

class.weights

Only for classif: Can class weights be handled?

rcens, lcens, icens

Only for surv: Can right, left, or interval censored data be handled?

prob

For classif, cluster, multilabel, surv: Can probabilites be predicted?

se

Only for regr: Can standard errors be predicted?

oobpreds

Only for classif, regr and surv: Can out of bag predictions be extracted from the trained model?

featimp

For classif, regr, surv: Does the model support extracting information on feature importance?

getLearnerProperties(learner)

hasLearnerProperties(learner, props)

Arguments

learner
[Learner | character(1)] The learner. If you pass a string the learner will be created via makeLearner.

props
[character] Vector of properties to query.

Value

getLearnerProperties returns a character vector with learner properties. hasLearnerProperties returns a logical vector of the same length as props.

See also

Other learner: getClassWeightParam, getHyperPars, getLearnerId, getLearnerPackages, getLearnerParVals, getLearnerParamSet, getLearnerPredictType, getLearnerShortName, getLearnerType, getParamSet, helpLearnerParam, helpLearner, makeLearners, makeLearner, removeHyperPars, setHyperPars, setId, setLearnerId, setPredictThreshold, setPredictType