makeClassifTask {mlr}	R Documentation

Create a classification / regression task for a given data set.

Description

The task encapsulates the data and specifies - through its subclasses - the type of the task (either classification or regression), and contains a description object detailing further aspects of the data.

Usage

  makeClassifTask(id, data, target, weight, blocking,
    positive, check.data = TRUE)

  makeRegrTask(id, data, target, weight, blocking,
    check.data = TRUE)

Arguments

id
[character(1)]
Id string for object. Default is the name of R variable passed to data.

data
[data.frame]
A data frame containing the features and target variable.

target
[character(1)]
Name of the target variable.

weight
[character(1)]
Optional, name of the column containing the non-negative case weight vector to be used during fitting.

blocking
[factor]
An optional factor of the same length as the number of observations. Observations with the same blocking level “belong together”. Specifically, they are either put all in the training or the test set during a resampling iteration. Default is no blocking.

positive
[character(1)]
Positive class for binary classification. Default is the first factor level of the target attribute.

check.data
[logical(1)]
Should sanity of data be checked initially at task creation? You should have good reasons to turn this off. Default is TRUE

Details

Useful operators are: getTaskFormula, getTaskFormulaAsString, getTaskFeatureNames, getTaskData, getTaskTargets, subsetTask.

Object members:

env [environment]

Environment where data for the task are stored. Use getTaskData in order to access it.

blocking [factor]

See argument above. factor(0) if not present.

task.desc [TaskDesc]

Encapsulates further information about the task.

Value

[SupervisedTask].

Examples

library(mlbench)
data(BostonHousing)
RegrTask <- makeRegrTask(data = BostonHousing, target = "medv")

ClassifTask <- makeClassifTask(data = iris, target = "Species")

## an example of a classification task with more than those standard arguments:
library(mlbench)
data(Ionosphere)
blocks <- factor(c(rep(1, 51), rep(2, 300)))
ClassifTask <- makeClassifTask(id = "myIonosphere", data = Ionosphere, target = "Class",
                               positive = "good", blocking = blocks)

[Package mlr version 1.2 Index]