makeTuneControlCMAES {mlr}	R Documentation

Create control structures for tuning.

Description

The following tuners are available:

makeTuneControlGrid

Grid search. All kinds of parameter types can be handled, but you have discretize them yourself by always using makeDiscreteParam in the par.set passed to tuneParams.

makeTuneControlRandom

Random search. All kinds of parameter types can be handled.

makeTuneControlOptim

Tuning with optim. Can handle numeric(vector) and integer(vector) hyperparameters. For integers the internally proposed numeric values are automatically rounded.

makeTuneControlCMAES

CMA Evolution Strategy with method cma_es. Can handle numeric(vector) and integer(vector) hyperparameters. For integers the internally proposed numeric values are automatically rounded.

makeTuneControlIrace

Tuning with iterated F-Racing with method irace. All kinds of parameter types can be handled.

Usage

  makeTuneControlCMAES(same.resampling.instance = TRUE,
    start, ...)

  makeTuneControlGrid(same.resampling.instance = TRUE)

  makeTuneControlIrace(n.instances = 100L,
    show.irace.output = FALSE, ...)

  makeTuneControlOptim(same.resampling.instance = TRUE,
    start, ...)

  makeTuneControlRandom(same.resampling.instance = TRUE,
    maxit = 100L)

Arguments

same.resampling.instance
[logical(1)]
Should the same resampling instance be used for all evaluations to reduce variance? Default is TRUE.

start
[numeric]
Named list of initial parameter values.

n.instances
[integer(1)]
Number of random resampling instances for irace, see details. Default is 100.

maxit
[integer(1)]
Number of iterations for random search. Default is 100.

show.irace.output
[logical(1)]
Show console output of irace while tuning? Default is FALSE.

...
[any]
Further control parameters passed to the control argument of optim, the control argument of cma_es, tunerConfig argument of irace.

Details

Dependent parameters can currently only be handled by random search and irace.

Some notes on irace: For resampling you have to pass a ResampleDesc, not a ResampleInstance. The resampling strategy is randomly instantiated n.instances times and these are the instances in the sense of irace (instances element of tunerConfig in irace). Also note that irace will always store its tuning results in a file on disk, see the package documentation for details on this and how to change the file path.

Value

[TuneControl]. The specific subclass is one of TuneControlGrid, TuneControlRandom, TuneControlOptim, TuneControlCMAES, TuneControlIrace.

[Package mlr version 1.2 Index]