makeTuneMultiCritControlGrid {mlr}	R Documentation

Create control structures for multi-criteria tuning.

Description

The following tuners are available:

makeTuneControlGrid

Grid search. All kinds of parameter types can be handled. You can either use their correct param type and resolution, or discretize them yourself by always using makeDiscreteParam in the par.set passed to tuneParams.

makeTuneMultiCritControlRandom

Random search. All kinds of parameter types can be handled.

makeTuneMultiCritControlNSGA2

Evolutionary method nsga2. Can handle numeric(vector) and integer(vector) hyperparameters, but no dependencies. For integers the internally proposed numeric values are automatically rounded.

Usage

makeTuneMultiCritControlGrid(same.resampling.instance = TRUE,
  resolution = 10L)

makeTuneMultiCritControlNSGA2(same.resampling.instance = TRUE,
  impute.val = NULL, ...)

makeTuneMultiCritControlRandom(same.resampling.instance = TRUE,
  maxit = 100L)

Arguments

maxit
[integer(1)]
Number of iterations for random search. Default is 100.

same.resampling.instance
[logical(1)]
Should the same resampling instance be used for all evaluations to reduce variance? Default is TRUE.

resolution
[integer]
Resolution of the grid for each numeric/integer parameter in par.set. For vector parameters, it is the resolution per dimension. Either pass one resolution for all parameters, or a named vector. See generateGridDesign. Default is 10.

impute.val
[numeric]
If something goes wrong during optimization (e.g, the learner crashes), this value is fed back to the tuner, so the tuning algorithm does not abort. It is not stored in the optimization path, an NA and a corresponding error message are logged instead. Note that this value is later multiplied by -1 for maximization measures internally, so you need to enter a larger positive value for maximization here as well. Default is the worst obtainable value of the performance measure you optimize for when you aggregate by mean value, or Inf instead. For multi-criteria optimization pass a vector of imputation values, one for each of your measures, in the same order as your measures.

...
[any]
Further control parameters passed to the control argument of cma_es and tunerConfig argument of irace.

Value

[TuneMultiCritControl]. The specific subclass is one of TuneMultiCritControlGrid, TuneMultiCritControlRandom, TuneMultiCritControlNSGA2.

See Also

Other tune_multicrit: plotTuneMultiCritResult; tuneParamsMultiCrit

[Package mlr version 2.1 Index]