TuneMultiCritResult {mlr}	R Documentation

Result of multi-criteria tuning.

Description

Container for results of hyperparameter tuning. Contains the obtained pareto set and front and the optimization path which lead there.

Object members:

learner [Learner]

Learner that was optimized.

control [TuneControl]

Control object from tuning.

x [list]

List of lists of non-dominated hyperparameter settings in pareto set. Note that when you have trafos on some of your params, x will always be on the TRANSFORMED scale so you directly use it.

y [matrix]

Pareto front for x.

opt.path [OptPath]

Optimization path which lead to x. Note that when you have trafos on some of your params, the opt.path always contains the UNTRANSFORMED values on the original scale. You can simply call trafoOptPath(opt.path) to transform them, or, as.data.frame{trafoOptPath(opt.path)}

[Package mlr version 2.1 Index]