aggregations {mlr}	R Documentation

Aggregation functions.

Description

Aggregation functions. You can use any function in R which reduces a numerical vector to simple real number. Refer to the function either by using an R function object or by using a string denoting the built-in function. If you choose to pass an R function, you should set the attribute 'id' to a string to give the function a name. You can use multiple aggregation functions if you pass a list or vector of the former.

[Package mlr version 0.3.1095 Index]