analyzeFeatSelResult {mlr}	R Documentation

Show and visualize the steps of feature selection.

Description

This function prints the steps selectFeatures took to find its optimal set of features and the reason why it stopped. It can also print information about all calculations done in each intermediate step.

Currently only implemented for sequential feature selection.

Usage

analyzeFeatSelResult(res, reduce = TRUE)

Arguments

res
[FeatSelResult]
The result of of selectFeatures.

reduce
[logical(1)]
Per iteration: Print only the selected feature (or all features that were evaluated)? Default is TRUE.

Value

[invisible(NULL)].

See Also

Other featsel: FeatSelControl, FeatSelControlExhaustive, FeatSelControlGA, FeatSelControlRandom, FeatSelControlSequential, makeFeatSelControlExhaustive, makeFeatSelControlGA, makeFeatSelControlRandom, makeFeatSelControlSequential; makeFeatSelWrapper; selectFeatures

[Package mlr version 1.2 Index]