args.to.control {mlr}	R Documentation

Split arguments into 'control' and 'other' arguments.

Description

Split arguments into 'control' and 'other' arguments.

Usage

args.to.control(control, arg.names, args)

Arguments

control 	 [function]
Function to apply to the elements of args named in arg.names.

arg.names 	 [character]
List of argument names to extract from args.

args 	 [list]
List of named arguments to be split into control and other arguments.

Details

Find all elements in list args whose name is contained in arg.names and call function control on these. The result of this is returned as the control element of the list returned. All remaining elements in args are returned as the args element of the return list.

Value

List with elements control and args.

[Package mlr version 0.3.1095 Index]