base.wrapper-class {mlr}	R Documentation

Abstract base class to wrap something around a learner.

Description

Abstract base class to wrap something around a learner.

Extends

learner

Methods

[

train.learner

pred.learner

[Package mlr version 0.3.777 Index]