bench.result-class {mlr}	R Documentation

bench-result

Description

Container for the results of a benchmark experiment.

Details

bench.result-class: Getter.
The following getters all return list of lists of objects: perf, prediction, opt.result. The first list iterates the tasks, the second one the learnes, both are named by respective ids. You can reduce these list by using the optional arguments 'task' and 'learner'. 'Drop' is by default TRUE, which means that the list structures are simplified as much as possible, if you don't want this set 'drop' to FALSE.

learners [character]

Ids of learners used in experiment.

tasks [character]

Ids of tasks used in experiment.

measures [character]

Names of measures recorded in experiment.

iters [numeric]

Named numerical vector which lists the number of iterations for every task. Names are ids of task.

prediction [see above]

List of list of predictions for every task/learner.

conf.mat [see above]

List of list of confusion matrices for every task/learner.

opt.result [see above]

List of list of opt.result for every task/learner. Entry is NULL if no optimization was done.

opt.perf [see above]

List of list of performance vectors of optimals settings for every task/learner. Note that this performance refers to the inner resampling! Entry is NULL if no optimization was done.

opt.par [see above]

List of list of opt.result for every task/learner. Entry is NULL if no optimization was done.

opt.path [see above]

List of list of optimization paths for every task/learner. Entry is NULL if no optimization was done.

tuned.par [see above]

List of list of optimal hyperparameters for every task/learner. Entry is NULL if no tuning was done.

sel.var [see above]

List of list of optimal features for every task/learner. Entry is NULL if no feature selection was done.

perf [list]

Lists for every data set and for every performance measure the performances of the different learners in each iteration.

Extends

object

Methods

[

to.string

See Also

bench.exp

[Package mlr version 0.3.732 Index]