benchmark {mlr}	R Documentation

Benchmark experiment for multiple learners and tasks.

Description

Complete benchmark experiment to compare different learning algorithms across one or more tasks w.r.t. a given resampling strategy. Experiments are paired, meaning always the same training / test sets are used for the different learners. Furhtermore, your learners can be automatically tuned using makeTuneWrapper.

Usage

benchmark(learners, tasks, resamplings, measures,
  show.info = getMlrOption("show.info"))

Arguments

learners
[(list of) Learner]
Learning algorithms which should be compared.

tasks
[(list of) Task]
Tasks that learners should be run on.

resamplings
[(list of) ResampleDesc | ResampleInstance]
Resampling strategy for each tasks. If only one is provided, it will be replicated to match the number of tasks. If missing, a 10-fold cross validation is used.

measures
[(list of) Measure]
Performance measures for all tasks. If missing, the default measure of the first task is used.

show.info
[logical(1)]
Print verbose output on console? Default is set via configureMlr.

Value

[BenchmarkResult].

See Also

Other benchmark: BenchmarkResult; FilterResult, getFilterResult; getAggrPerformances; getFeatSelResult; getPerformances; getPredictions; getTuneResult

[Package mlr version 2.1 Index]