createDummyFeatures {mlr}	R Documentation

Generate dummy variables for factor features.

Description

Replace all factor features with their dummy variables. Internally model.matrix is used. Non factor features will be left untouched and passed to the result.

Usage

createDummyFeatures(obj, target = character(0L), method = "1-of-n",
  exclude = character(0L))

Arguments

obj
[data.frame | SupervisedTask]
Input data.

target
[character()]
Name of the column(s) specifying the response. Only used when obj is a data.frame, otherwise ignored.

method
[character(1)]
Available are:
“1-of-n”: For n factor levels there will be n dummy variables.
“reference”: There will be n-1 dummy variables leaving out the first factor level of each variable.

exclude
[character]
Names of the columns to exclude. The target does not have to be included here. Default is none.

Value

[data.frame | SupervisedTask]. Same type as obj.

See Also

model.matrix

[Package mlr version 2.0 Index]