crossover {mlr}	R Documentation

crossover

Description

Takes two bit strings and creates a new one of the same size by selecting the items from the first string or the second, based on a given rate (the probability of choosing an element from the first string).

Arguments

x
[logical]
First parent string.

y
[logical]
Second parent string.

rate
[numeric(1)]
A number representing the probability of selecting an element of the first string. Default is 0.5.

Value

[crossover].

[Package mlr version 2.1 Index]