estimateResidualVariance {mlr}	R Documentation

Estimate the residual variance

Description

Estimate the residual variance of a regression model on a given task. If a regression learner is provided instead of a model, the model is trained (see train) first.

Usage

estimateResidualVariance(x, task, data, target)

Arguments

x
[Learner or WrappedModel]
Learner or wrapped model.

task
[RegrTask]
Regression task. If missing, data and target must be supplied.

data
[data.frame]
A data frame containing the features and target variable. If missing, task must be supplied.

target
[character(1)]
Name of the target variable. If missing, task must be supplied.

[Package mlr version 2.1 Index]