Extract features from functional data.

Extract non-functional features from functional features using various methods. The function extractFDAFeatures performs the extraction for all functional features via the methods specified in feat.methods and transforms all mentioned functional matrix features into regular data.frame columns. Additionally, a “extractFDAFeatDesc” object which contains learned coefficients and other helpful data for extraction during the predict-phase is returned. This can be used with reextractFDAFeatures in order to extract features during the prediction phase.

extractFDAFeatures(obj, target = character(0L), feat.methods = list())

Arguments

obj
[Task | data.frame] Task or data.frame to extract functional features from. Must contain functional features as matrix columns.

target
[character] Task target column. Only neccessary for data.frames Default is character(0).

feat.methods
[named list] List of functional features along with the desired methods for each functional feature. “all” applies the extratFDAFeatures method to each functional feature. Names of feat.methods must match column names of functional features. Available feature extraction methods are available under family fda_featextractor. Default is list() which does nothing.

Value

[list]

data [data.frame | Task]

Extracted features, returns a data.frame when given a data.frame and a Task when given a Task.

desc [extracFDAFeatDesc]

Description object. See description for details.

Details

The description object contains these slots

target [character]

See argument.

coln [character]

colum names of data.

fd.cols [character]

Functional feature names.

extractFDAFeat [list]

Contains feature.methods and relevant parameters for reextraction

.

See also

Other fda: makeExtractFDAFeatMethod, makeExtractFDAFeatsWrapper