filterFeatures {mlr}	R Documentation

Filter features by using a numerical importance criterion.

Description

Calculates numerical importance values for all features. Thresholding of these values can be used to select “useful” features. Look at package FSelector for details on the filter algorithms.

Usage

filterFeatures(task, method = "random.forest.importance")

Arguments

task
[SupervisedTask]
The task.

method
[character(1)]
Filter method. Available are:
“linear.correlation”: Pearson's correlation for regression with continous data
“rank.correlation”: Spearman's correlation for regression with continous data
“information.gain”: Entropy-based information gain for classification and regression of mixed feature sets
“gain.ratio”: Entropy-based gain ratio for classification and regression of mixed feature sets
“symmetrical.uncertainty”: Entropy-based symmetrical uncertainty for classification and regression of mixed feature sets
“chi.squared”: Finds weights of mixed feature sets using chi-squared test for classification and regression
“random.forest.importance”: Finds weights of mixed feature sets using RandomForest algorithm for classification and regression
“relief”: Finds weights of mixed feature sets using distances between instances for classification and regression
“oneR”: Finds weights of mixed feature sets using association rules for classification and regression
Default is “random.forest.importance”.

Value

[numeric]. A named numeric vector that contains an importance value for each feature.

See Also

makeFilterWrapper

[Package mlr version 1.2 Index]