Extract the test performance values from a benchmark result.

Either a list of lists of “measure.test” data.frames, as returned by resample, or these objects are rbind-ed with extra columns “task.id” and “learner.id”.

getBMRPerformances(bmr, task.ids = NULL, learner.ids = NULL,
  as.df = FALSE, drop = FALSE)

Arguments

bmr
(BenchmarkResult)
Benchmark result.

task.ids
(character(1))
Restrict result to certain tasks. Default is all.

learner.ids
(character(1))
Restrict result to certain learners. Default is all.

as.df
(character(1))
Return one data.frame as result - or a list of lists of objects?. Default is FALSE.

drop
(logical(1))
If drop is FALSE (the default), a nested list with the following structure is returned:
res[task.ids][learner.ids].
If drop is set to TRUE it is checked if the list structure can be simplified.
If only one learner was passed, a list with entries for each task is returned.
If only one task was passed, the entries are named after the corresponding learner.
For an experiment with both one task and learner, the whole list structure is removed.
Note that the name of the task/learner will be dropped from the return object.

Value

(list | data.frame). See above.

See also