getFailureModelMsg {mlr}	R Documentation

Return error message of FailureModel.

Description

Such a model is created when one sets the corresponding option in configureMlr. If no failure occurred, NA is returned.

For complex wrappers this getter returns the first error essage encountered in ANY model that failed.

Usage

getFailureModelMsg(model)

Arguments

model
[WrappedModel]
The model.

Value

[character(1)].

[Package mlr version 2.1 Index]