getFeatSelResult {mlr}	R Documentation

Returns the selected feature set and optimization path after training or benchmarking.

Description

Returns the selected feature set and optimization path after training or benchmarking.

Usage

getFeatSelResult(object)

Arguments

object
[WrappedModel | BenchmarkResult]
Trained Model created with makeFeatSelWrapper or benchmark result created with benchmark.

Value

[FeatSelResult or list of FeatSelResults]. NULL, if no feature selection was performed.

See Also

Other benchmark: BenchmarkResult; FilterResult, getFilterResult; benchmark; getAggrPerformances; getPerformances; getPredictions; getTuneResult

Other featsel: FeatSelControl, FeatSelControlExhaustive, FeatSelControlGA, FeatSelControlRandom, FeatSelControlSequential, makeFeatSelControlExhaustive, makeFeatSelControlGA, makeFeatSelControlRandom, makeFeatSelControlSequential; analyzeFeatSelResult; makeFeatSelWrapper; selectFeatures

[Package mlr version 2.1 Index]