getHyperPars {mlr}	R Documentation

Get current parameter settings for a learner.

Description

Get current parameter settings for a learner.

Usage

getHyperPars(learner, for.fun = c("train", "predict", "both"))

Arguments

learner
[Learner | character(1)]
The learner. If you pass a string the learner will be created via makeLearner.

for.fun
[character(1)]
Restrict the returned settings to hyperparameters corresponding to when the are used (see LearnerParam). Must be a subset of: “train”, “predict” or “both”. Default is c("train", "predict", "both").

Value

[list]. A named list of values.

See Also

Other learner: getParamSet; removeHyperPars; setHyperPars; setId; setPredictType; showHyperPars

[Package mlr version 2.1 Index]