getLearnerModel {mlr}	R Documentation

Get underlying R model of learner integrated into mlr.

Description

Get underlying R model of learner integrated into mlr.

Usage

  getLearnerModel(model)

Arguments

model
[WrappedModel]
The model, returned by e.g., train.

Value

[any]. A fitted model, depending the learner / wrapped package. E.g., a model of class rpart for learner “classif.rpart”.

[Package mlr version 1.2 Index]