Retrieve binary classification measures for multilabel classification predictions.

Measures the quality of each binary label prediction w.r.t. some binary classification performance measure.

getMultilabelBinaryPerformances(pred, measures)

Arguments

pred
[Prediction] Multilabel Prediction object.

measures
[Measure | list of Measure] Performance measure(s) to evaluate, must be applicable to binary classification performance. Default is mmce.

Value

[named matrix]. Performance value(s), column names are measure(s), row names are labels.

See also

Other multilabel: makeMultilabelBinaryRelevanceWrapper, makeMultilabelClassifierChainsWrapper, makeMultilabelDBRWrapper, makeMultilabelNestedStackingWrapper, makeMultilabelStackingWrapper

Examples

# see makeMultilabelBinaryRelevanceWrapper