getParamSet {mlr}	R Documentation

Get a description of all possible parameter settings for a learner.

Description

Get a description of all possible parameter settings for a learner.

Usage

getParamSet(learner)

Arguments

learner
[Learner]
The learner.

Value

[ParamSet].

[Package mlr version 1.2 Index]