Get probabilities for some classes.

Get probabilities for some classes.

getPredictionProbabilities(pred, cl)

Arguments

pred
[Prediction] Prediction object.

cl
[character] Names of classes. Default is either all classes for multi-class / multilabel problems or the positive class for binary classification.

Value

[data.frame] with numerical columns or a numerical vector if length of cl is 1. Order of columns is defined by cl.

See also

Other predict: asROCRPrediction, getPredictionResponse, getPredictionTaskDesc, plotViperCharts, predict.WrappedModel, setPredictThreshold, setPredictType

Examples

task = makeClassifTask(data = iris, target = "Species")
lrn = makeLearner("classif.lda", predict.type = "prob")
mod = train(lrn, task)

#> Error: Please use column names for `x`

# predict probabilities
pred = predict(mod, newdata = iris)

#> Error in predict(mod, newdata = iris): object 'mod' not found

# Get probabilities for all classes
head(getPredictionProbabilities(pred))

#> Error in checkClass(x, classes, ordered, null.ok): object 'pred' not found

# Get probabilities for a subset of classes
head(getPredictionProbabilities(pred, c("setosa", "virginica")))

#> Error in checkClass(x, classes, ordered, null.ok): object 'pred' not found