getProbabilities {mlr}	R Documentation

Get probabilities for some classes.

Description

Get probabilities for some classes.

Usage

  getProbabilities(pred, cl)

Arguments

pred
[Prediction]
Prediction object.

cl
[character]
Names of classes. Default is either all classes for multi-class problems or the positive class for binary classification.

Value

[data.frame] with numerical columns or a numerical vector if length of cl is 1. Order of columns is defined by cl.

See Also

predict.WrappedModel

Examples

task <- makeClassifTask(data = iris, target = "Species")
lrn <- makeLearner("classif.lda", predict.type = "prob")
mod <- train(lrn, task)
# predict probabilities
pred <- predict(mod, newdata = iris)
# Get probabilities for all classes
head(getProbabilities(pred))

##   setosa versicolor virginica
## 1      1  3.896e-22 2.611e-42
## 2      1  7.218e-18 5.042e-37
## 3      1  1.464e-19 4.676e-39
## 4      1  1.269e-16 3.567e-35
## 5      1  1.637e-22 1.083e-42
## 6      1  3.883e-21 4.567e-40

# Get probabilities for a subset of classes
head(getProbabilities(pred, c("setosa", "virginica")))

##   setosa virginica
## 1      1 2.611e-42
## 2      1 5.042e-37
## 3      1 4.676e-39
## 4      1 3.567e-35
## 5      1 1.083e-42
## 6      1 4.567e-40

[Package mlr version 1.2 Index]