getTuneResult {mlr}	R Documentation

Returns the optimal hyperparameters and optimization path after training or benchmarking.

Description

Returns the optimal hyperparameters and optimization path after training or benchmarking.

Usage

getTuneResult(object)

Arguments

object
[WrappedModel | BenchmarkResult]
Trained Model created with makeTuneWrapper or benchmark result created with benchmark.

Value

[TuneResult or list of TuneResults].

See Also

Other tune: ModelMultiplexer, makeModelMultiplexer; TuneControl, TuneControlCMAES, TuneControlGrid, TuneControlIrace, TuneControlOptim, TuneControlRandom, makeTuneControlCMAES, makeTuneControlGrid, makeTuneControlIrace, makeTuneControlOptim, makeTuneControlRandom; makeModelMultiplexerParamSet; makeTuneWrapper; tuneParams; tuneThreshold

[Package mlr version 1.2 Index]