grid.control {mlr}	R Documentation

Control for grid search tuning.

Description

Control structure for grid search tuning.

Usage

grid.control(ranges=list(), minimize=TRUE, tune.threshold=FALSE, thresholds=10,
    scale=identity)

Arguments

ranges 	 [list]
A list of named vectors/lists of possible values for each hyperparameter. You can also pass a list of such ranges by using [combine.ranges] in the rare case when it does not make sense to search a complete cross-product of range values.

minimize 	 [logical]
Minimize performance measure? Default is TRUE.

tune.threshold 	 [logical]
Perform empirical thresholding? Default is FALSE. Only supported for binary classification and you have to set predict.type to "prob" for this in make.learner.

thresholds 	 [numeric]
Number of thresholds to try in tuning. Predicted probabilities are sorted and divided into groups of equal size. Default is 10.

scale 	 [function]
A function to scale the hyperparameters. E.g. maybe you want to optimize in some log-space. Has to take a vector and return a scaled one. Default is identity function.

Value

Control structure for tuning.

[Package mlr version 0.3.732 Index]