Join some class existing levels to new, larger class levels for classification problems.

Join some class existing levels to new, larger class levels for classification problems.

joinClassLevels(task, new.levels)

Arguments

task
[Task] The task.

new.levels
[list of character] Element names specify the new class levels to create, while the corresponding element character vector specifies the existing class levels which will be joined to the new one.

Value

[Task].

Examples

joinClassLevels(iris.task, new.levels = list(foo = c("setosa", "virginica")))

#> Supervised task: iris-example
#> Type: classif
#> Target: Species
#> Observations: 150
#> Features:
#>    numerics     factors     ordered functionals 
#>           4           0           0           0 
#> Missings: FALSE
#> Has weights: FALSE
#> Has blocking: FALSE
#> Is spatial: FALSE
#> Classes: 2
#>        foo versicolor 
#>        100         50 
#> Positive class: foo