learn.task-class {mlr}	R Documentation

Base class for learning tasks.

Description

A learning task is a general description object for a machine learning experiment. It wraps the data source and specifies - through its subclasses - the type of the task (e.g. classification or regression), the target variable and other details of the problem. As this is just an abstract base class, you should not instantiate it directly but use the inheriting classes and their factory methods.

Details

learn.task-class: Getter.

id [string]

Id string of task.

label [string]

Label string of task.

is.classif [boolean]

Is this a classification task?

is.regr [boolean]

Is this a regression task?

data [data.frame]. Optional parameters: row, col

The data.frame is returned, possibly indexed by row/col. If col is missing, only columns which were not excluded are returned.

size [integer]

Number of cases.

target.name [string]

The name of the target variable.

input.names [character]

The names of the input variables (without excluded variables).

excluded [character]

Names of excluded variables.

targets [character]. Optional parameters: row

If row is missing all target values are returned. Otherwise they are indexed by row.

weights [numeric]. Optional parameters: row

If row is missing all case weights are returned. Otherwise they are indexed by row. NULL if no weights were set.

Extends

object

Methods

resample.fit

make.res.instance

make.res.instance

[

train

See Also

make.task

[Package mlr version 0.3.777 Index]