listMeasures {mlr}	R Documentation

Find matching measures.

Description

Returns the matching measures which have specific characteristics, e.g. whether they supports classification or regression.

listMeasuresForTask returns all learners that are in principle applicable for a given task.

Usage

listMeasures(properties = character(0L))

listMeasuresForTask(task, properties = character(0L))

Arguments

properties
[character)]
Set of required properties to filter for. Default is character(0). See Measure for some standardized properties.

task
[SupervisedTask]
The task.

Value

[character]. Class names of matching learners.

[Package mlr version 1.2 Index]