make.multiclass.wrapper {mlr}	R Documentation

Fuse learner with multiclass method.

Description

Fuses a base learner with a multi-class method. Creates a learner object, which can be used like any other learner object. This way learners which can only handle binary classification will be able to handle multi-class problems too.

Usage

make.multiclass.wrapper(learner, id=as.character(NA), label=as.character(NA), method,
    codematrix, ...)

Arguments

learner 	 [learner or string]
Learning algorithm. See learners.

id 	 [string]
Id for resulting learner object. If missing, id of "learner" argument is used.

label 	 [string]
Label for resulting learner object. If missing, label of "learner" argument is used.

method 	 [string]
Currently unsupported.

codematrix 	 [matrix]
ECOC codematrix with entries +1,-1,0. Columns define new binary problems, rows correspond to classes.

... 	 [any]
Optional parameters. Not used currently.

Value

learner.

[Package mlr version 0.3.1095 Index]