make.preproc.wrapper {mlr}	R Documentation

Fuse learner with preprocessing.

Description

Fuses a base learner with a preprocessing method. Creates a learner object, which can be used like any other learner object, but which internally preprocesses the data as requested. If the train or predict function is called on it, the preprocessing is always invoked before.

Usage

make.preproc.wrapper(learner, fun, ...)

Arguments

learner 	 [learner or string]
Learning algorithm. See learners.

fun 	 [function]
Function to preprocess a data.frame. First argument must be called 'data', which will be preprocessed and subsequently returned.

... 	 [any]
Optional parameters to control the preprocessing. Passed to fun.

Value

learner.

[Package mlr version 0.3.732 Index]