make.task {mlr}	R Documentation

Construct learning task.

Description

Defines a learning task for a data set and is the starting point for further steps like training, predicting new data, resampling and tuning and benchmarking. The type (classification or regression) is automatically inferred from the target variable. It might perform some data conversions in the data.frame, like coverting integer features to numerics, but will generally warn about this.

Usage

make.task(id, label, data, target, formula, excluded, weights, costs, positive)

Arguments

id 	 [string]
Id string for object. Used to select the object from a named list, etc.

label 	 [string]
Label string for object. Used in plots, etc.

data 	 [data.frame]
A data frame containing the variables for the modelling.

target 	 [string]
Name of the target variable.

formula 	 [formula]
Instead of specifying the target, you can use the formula interface. If you are using just a subset of the variables or transformations, this will built a new internal data frame by calling model.frame.

excluded 	 [character] Names of inputs, which should be generally disregarded, e.g. IDs, etc. Default is zero-length vector.

weights 	 [numeric]
An optional vector of weights to be used in the fitting process. Default is not to use weights.

costs 	 [matrix]
An optional matrix of misclassification costs to be used in the fitting process. Ignored for regression.

positive 	 [string]
Positive class for binary classification. Default is the first factor level of the target attribute. Ignored for regression.

Value

learn.task.

[Package mlr version 0.3.528 Index]