make.tune.wrapper {mlr}	R Documentation

Fuse learner with tuning.

Description

Fuses a base learner with a search strategy to select its hyperparameters. Creates a learner object, which can be used like any other learner object, but which internally uses tune. If the train function is called on it, the search strategy and resampling are invoked to select an optimal set of hyperparameter values. Finally, a model is fitted on the complete training data with these optimal hyperparameters and returned.

Usage

make.tune.wrapper(learner, resampling, method="grid", control, measures, aggr)

Arguments

learner 	 [learner or string]
Learning algorithm. See learners.

resampling 	 [resample.instance] or [resample.desc]
Resampling strategy to evaluate points in hyperparameter space.

method 	 [character]
Search method. Currently supported are grid search "grid", pattern search "pattern", CMA-ES "cmaes" and Nelder-Mead "neldermead".

control 	 Control object for search method.

measures 	 [see measures] Performance measures.

aggr 	 [see aggregations] Aggregation functions.

Value

learner.

See Also

tune, grid.control, cmaes.control, neldermead.control

[Package mlr version 0.3.732 Index]