make.varsel.wrapper {mlr}	R Documentation

Fuse learner with variable selection.

Description

Fuses a base learner with a search strategy to select variables. Creates a learner object, which can be used like any other learner object, but which internally uses varsel. If the train function is called on it, the search strategy and resampling are invoked to select an optimal set variables. Finally, a model is fitted on the complete training data with these variables and returned.

Usage

make.varsel.wrapper(learner, id, label, resampling, measures, aggr, control)

Arguments

learner 	 [learner or string]
Learning algorithm. See learners.

id 	 [string]
Id string for object. Used to select the object from a named list, etc.

label 	 [string]
Label string for object. Used in plots, etc.

resampling 	 [resample.instance] or [resample.desc]
Resampling strategy to evaluate points in hyperparameter space.

control 	 Control object for search method. Also selects the optimization algorithm for feature selection.

measures 	 [see measures] Performance measures.

aggr 	 [see aggregations] Aggregation functions.

Value

learner.

See Also

varsel, varsel.control

[Package mlr version 0.3.819 Index]