Specify your own aggregation of measures.

This is an advanced feature of mlr. It gives access to some inner workings so the result might not be compatible with everything!

makeAggregation(id, name = id, properties, fun)

Arguments

id
[character(1)] Name of the aggregation method (preferably the same name as the generated function).

name
[character(1)] Long name of the aggregation method. Default is id.

properties
[character] Set of aggregation properties.

req.train

Are prediction or train sets required to calculate the aggregation?

req.test

Are prediction or test sets required to calculate the aggregation?

fun
[function(task, perf.test, perf.train, measure, group, pred)] Calculates the aggregated performance. In most cases you will only need the performances perf.test and optionally perf.train on the test and training data sets.

task [Task]

The task.

perf.test [numeric]

performance results on the test data sets.

perf.train [numeric]

performance results on the training data sets.

measure [Measure]

Performance measure.

group [factor]

Grouping of resampling iterations. This encodes whether specific iterations 'belong together' (e.g. repeated CV).

pred [Prediction]

Prediction object.

Value

[Aggregation].

See also

aggregations, setAggregation

Examples

# NOT RUN {
# computes the interquartile range on all performance values
test.iqr = makeAggregation(id = "test.iqr", name = "Test set interquartile range",
  properties = "req.test",
  fun = function (task, perf.test, perf.train, measure, group, pred) IQR(perf.test))
# }