makeBaggingWrapper {mlr}	R Documentation

Fuse learner with the bagging technique.

Description

Fuses a learner with the bagging method (i.e., similar to what a randomForest does). Creates a learner object, which can be used like any other learner object. Models can easily be accessed via getBaggingModels.

Usage

  makeBaggingWrapper(learner, bag.iters = 10L,
    bag.replace = TRUE, bag.size, bag.feats = 1,
    predict.type = "response")

Arguments

learner
[Learner]
The learner. Prediction type of the basic learner must be “response”.

bag.iters
[integer(1)]
Iterations = number of fitted models in bagging. Default is 10.

bag.replace
[logical(1)]
Sample bags with replacement (bootstrapping)? Default is TRUE.

bag.size
[numeric(1)]
Percentage size of sampled bags. Default is 1 for bootstrapping and 0.632 for subampling.

bag.feats
[numeric(1)]
Percentage size of randomly selected features in bags. Default is 1.

predict.type
[character(1)]
Classification: “response” (= labels) or “prob” (= probabilities and labels by selecting the ones with maximal probability). Regression: “response” (= mean response) or “se” (= standard errors and mean response). Default is “response”.

Details

Bagging is implemented as follows: For each iteration a random data subset is sampled (with or without replacement) and potentially the number of features is also restricted to a random subset. Note that this is usually handled in a slightly different way in the random forest where features are sampled at each tree split).

Prediction works as follows: For classification we do majority voting to create a discrete label and probabilites are predicted by considering the proportions of all predicted labels. For regression the mean value and the standard devations accross predictions is computed.

Value

[Learner].

[Package mlr version 1.2 Index]