makeCostSensClassifWrapper {mlr}	R Documentation

Wraps a classification learner for use in cost-sensitive learning.

Description

Creates a wrapper, which can be used like any other learner object. The classification model can easily be accessed via getCostSensClassifModel.

This is a very naive learner, where the costs are transformed into classification labels - the label for each case is the name of class with minimal costs. (If ties occur, the label which is better on average w.r.t. costs over all training data is preferred.) Then the classifier is fitted to that data and subsequently used for prediction.

Usage

makeCostSensClassifWrapper(learner)

Arguments

learner
[Learner | character(1)]
The classification learner. If you pass a string the learner will be created via makeLearner.

Value

[Learner].

[Package mlr version 2.0 Index]