makeCostSensRegrWrapper {mlr}	R Documentation

Wraps a regression learner for use in cost-sensitive learning.

Description

Creates a wrapper, which can be used like any other learner object. Models can easily be accessed via getCostSensRegrModels.

For each class in the task, an individual regression model is fitted for the costs of that class. During prediction, the class with the lowest predicted costs is selected.

Usage

makeCostSensRegrWrapper(learner)

Arguments

learner
[Learner | character(1)]
The regression learner. If you pass a string the learner will be created via makeLearner.

Value

[Learner].

See Also

Other costsens: ClassifTask, ClusterTask, CostSensTask, RegrTask, SurvTask, Task, makeClassifTask, makeClusterTask, makeCostSensTask, makeRegrTask, makeSurvTask; CostSensClassifModel, CostSensClassifWrapper, makeCostSensClassifWrapper; CostSensWeightedPairsModel, CostSensWeightedPairsWrapper, makeCostSensWeightedPairsWrapper

[Package mlr version 2.1 Index]