makeCustomResampledMeasure {mlr}	R Documentation

Construct your own resampled performance measure.

Description

Construct your own performance measure, used after resampling. Note that individual training / test set performance values will be set to NA, you only calculate an aggregated value. If you can define a function that makes sense for every single training / test set, implement your own Measure.

Usage

makeCustomResampledMeasure(id, minimize = TRUE, properties = character(0L),
  allowed.pred.types = character(0L), fun, extra.args = list(),
  best = NULL, worst = NULL)

Arguments

id
[character(1)]
Name of aggregated measure.

fun
[function(task, pred, group, pred, extra.args)]
Calculates performance value from ResamplePrediction object. For rare case you can also use the task, the grouping or the extra arguments extra.args.

extra.args
[list]
List of extra arguments which will always be passed to fun. Default is empty list.

minimize
[logical(1)]
Should the measure be minimized? Default is in TRUE.

properties
[character]
Set of measure properties. Some standard property names include:

classif

Is the measure applicable for classification?

classif.multi

Is the measure applicable for multi-class classification?

regr

Is the measure applicable for regression?

surv

Is the measure applicable for survival?

costsens

Is the measure applicable for cost-sensitve learning?

Default is character(0).

allowed.pred.types
[character]
Which prediction types are allowed for this measure? Subset of “response”,“prob”, “se”. Default is character(0).

best
[numeric(1)]
Best obtainable value for measure. Default is -Inf or Inf, depending on minimize.

worst
[numeric(1)]
Worst obtainable value for measure. Default is Inf or -Inf, depending on minimize.

Value

[Measure].

See Also

Other performance: Measure, makeMeasure; acc, auc, ber, cindex, f1, fdr, featperc, fn, fnr, fp, fpr, gmean, gpr, mae, mcc, mcp, meancosts, measures, medae, medse, mmce, mse, multiclass.auc, npv, ppv, rmse, sae, sse, timeboth, timepredict, timetrain, tn, tnr, tp, tpr; makeCostMeasure; performance

[Package mlr version 2.0 Index]