makeDownsampleWrapper {mlr}	R Documentation

Fuse learner with simple downsampling (subsampling).

Description

Creates a learner object, which can be used like any other learner object. It will only be trained on a subset of the original data to save computational time.

Usage

makeDownsampleWrapper(learner, dw.perc = 1, dw.stratify = FALSE)

Arguments

learner
[Learner | character(1)]
The learner. If you pass a string the learner will be created via makeLearner.

dw.perc
[numeric(1)]
See downsample.

dw.stratify
[logical(1)]
See downsample.

Value

[Learner].

See Also

Other downsample: downsample

[Package mlr version 2.0 Index]