Constructor for FDA feature extraction methods.

This can be used to implement custom feature FDA extraction.

makeExtractFDAFeatMethod(learn, reextract, args = list())

Arguments

learn
[function(data, target, col, ...)] Function to learn and extract information on functional column col. Arguments are:

data [data.frame] Data.frame with one row per observation of a single functional feature or time series and one column per measurement time point. All entries need to be numeric.

data [data.frame] Data.frame containing matricies with one row per observation of a single functional or time series and one column per measurement time point. All entries need to be numeric.

target [character] Name of the target variable. Default: “NULL”. The variable is only set to be consistent with the API.

col [character | numeric] column names or indices, the extraction should be performed on. The function has to return a named list of values.

reextract
[function(data, target, col, ...)] Function used for reextracting data in predict phase. Can be equal to learn.

args
[list] Named list of arguments to pass to learn via ....

See also

Other fda: extractFDAFeatures, makeExtractFDAFeatsWrapper