makeFeatSelWrapper {mlr}	R Documentation

Fuse learner with feature selection.

Description

Fuses a base learner with a search strategy to select variables. Creates a learner object, which can be used like any other learner object, but which internally uses selectFeatures. If the train function is called on it, the search strategy and resampling are invoked to select an optimal set of variables. Finally, a model is fitted on the complete training data with these variables and returned. See selectFeatures for more details.

After training, the optimal hyperparameters (and other related information) can be retrieved with getFeatSelResult.

Usage

makeFeatSelWrapper(learner, resampling, measures, bit.names, bits.to.features,
  control, show.info = TRUE)

Arguments

learner
[Learner]
The learner.

resampling
[ResampleDesc | ResampleInstance]
Resampling strategy to evaluate feature sets. If you pass a description, it is instantiated once at the beginning by default, so all feature sets are evaluated on the same training/test sets. If you want to change that behaviour, look at TuneControl.

measures
[Measure | list of Measure]
Performance measures to evaluate. The first measure, aggregated by the first aggregation function is optimized during feature selection, others are simply evaluated.

bit.names
[character]
Names of bits encoding the solutions. Also defines the total number of bits in the encoding. Per default these are the feature names of the task.

bits.to.features
[function(x, task)]
Function which transforms an integer-0-1 vector into a character vector of selected features. Per default a value of 1 in the ith bit selects the ith feature to be in the candidate solution.

control
[FeatSelControl]
Control object for search method. Also selects the optimization algorithm for feature selection.

show.info
[logical(1)]
Show info message after each feature set evaluation? Default is TRUE.

Value

[Learner].

[Package mlr version 1.2 Index]