makeFilterWrapper {mlr}	R Documentation

Fuse learner with a feature filter method.

Description

Fuses a base learner with a filter method. Creates a learner object, which can be used like any other learner object. Internally uses filterFeatures before every model fit.

After training, the selected features can be retrieved with getFilteredFeatures.

Note that observation weights do not influence the filtering and are simply passed down to the next learner.

Usage

makeFilterWrapper(learner, fw.method = "random.forest.importance",
  fw.select = "perc", fw.val)

Arguments

learner
[Learner | character(1)]
The learner. If you pass a string the learner will be created via makeLearner.

fw.method
[character(1)]
See getFilterValues. Default is “random.forest.importance”.

fw.select
[character(1)]
See filterFeatures. Default is “perc”.

fw.val
[numeric(1)]
See filterFeatures.

Value

[Learner].

See Also

Other filter: FilterResult, getFilterResult; FilterValues; filterFeatures; getFilterMethods; getFilterValues; getFilteredFeatures

Examples

task = makeClassifTask(data = iris, target = "Species")
lrn = makeLearner("classif.lda")
inner = makeResampleDesc("Holdout")
outer = makeResampleDesc("CV", iters = 2)
lrn = makeFilterWrapper(lrn, fw.val = 0.5)
mod = train(lrn, task)

## Loading required package: FSelector

print(getFilteredFeatures(mod))

## [1] "Petal.Length" "Petal.Width"

# now nested resampling, where we extract the features that the filter method selected
r = resample(lrn, task, outer, extract = function(model) {
  getFilteredFeatures(model)
})

## [Resample] cross-validation iter: 1
## [Resample] cross-validation iter: 2
## [Resample] Result: mmce.test.mean=0.0467

print(r$extract)

## [[1]]
## [1] "Petal.Width"  "Petal.Length"
## 
## [[2]]
## [1] "Petal.Length" "Petal.Width"

[Package mlr version 1.2 Index]