makeImputeMethod {mlr}	R Documentation

Create a custom imputation method.

Description

This is a constructor to create your own imputation methods.

Usage

makeImputeMethod(learn, impute, args = list())

Arguments

learn
[function(data, target, col, ...)]
Function to learn and extract information on column col out of data frame data. Argument target specifies the target column of the learning task. The function has to return a named list of values.

impute
[function(data, target, col, ...)]
Function to impute missing values in col using information returned by learn on the same column. All list elements of the return values o learn are passed to this function into ....

args
[list]
Named list of arguments to pass to learn via ....

See Also

Other impute: imputations, imputeConstant, imputeHist, imputeLearner, imputeMax, imputeMean, imputeMedian, imputeMin, imputeMode, imputeNormal, imputeUniform; impute; makeImputeWrapper; reimpute

[Package mlr version 2.0 Index]