makeLearner {mlr}	R Documentation

Create learner object.

Description

For a classification learner the predict.type can be set to “prob” to predict probabilities and the maximum value selects the label. The threshold used to assign the label can later be changed using the setThreshold function.

Usage

makeLearner(cl, id = cl, predict.type = "response", ...,
  par.vals = list())

Arguments

cl
[character(1)]
Class of learner. By convention, all classification learners start with “classif.”, all regression learners with “regr.” and all survival learners start with “surv.”. A list of all learners is available on the learners help page.

id
[character(1)]
Id string for object. Used to display object. Default is cl.

predict.type
[character(1)]
Classification: “response” (= labels) or “prob” (= probabilities and labels by selecting the ones with maximal probability). Regression: “response” (= mean response) or “se” (= standard errors and mean response). Survival: “response” (= some sort of orderable risk) or “prob” (= time dependent probabilities). Default is “response”.

...
[any]
Optional named (hyper)parameters. Alternatively these can be given using the par.vals argument.

par.vals
[list]
Optional list of named (hyper)parameters. The arguments in ... take precedence over values in this list. We strongly encourage you to use one or the other to pass (hyper)parameters to the learner but not both.

Value

[Learner].

See Also

[resample], [predict.WrappedModel]

Examples

makeLearner("classif.rpart")

## Loading required package: rpart

## Learner classif.rpart from package rpart
## Type: classif
## Class: classif.rpart
## Properties: twoclass,multiclass,missings,numerics,factors,prob,weights
## Predict-Type: response
## Hyperparameters: xval=0

makeLearner("classif.lda", predict.type = "prob")

## Learner classif.lda from package MASS
## Type: classif
## Class: classif.lda
## Properties: twoclass,multiclass,numerics,factors,prob
## Predict-Type: prob
## Hyperparameters:

lrn = makeLearner("classif.lda", method = "t", nu = 10)
print(lrn$par.vals)

## $method
## [1] "t"
## 
## $nu
## [1] 10

[Package mlr version 1.2 Index]