makeModelMultiplexer {mlr}	R Documentation

Create model multiplexer for model selection to tune over multiple possible models.

Description

Combines multiple base learners by dispatching on the hyperparameter “selected.learner” to a specific model class. This allows to tune not only the model class (SVM, random forest, etc) but also their hyperparameters in one go. Combine this with tuneParams and makeTuneControlIrace for a very powerful approach, see example below.

The parameter set is the union of all (unique) base learners. In order to avoid name clashes all parameter names are prefixed with the base learner id, i.e. “[learner.id].[parameter.name]”.

Usage

makeModelMultiplexer(base.learners, id = "ModelMultiplexer")

Arguments

base.learners
[list of Learner]
List of Learners with unique IDs.

id
[character(1)]
Identifier for constructed ModelMultiplexer. Default is “ModelMultiplexer”.

Value

[ModelMultiplexer]. A Learner specialized as ModelMultiplexer.

Note

Note that logging output during tuning is somewhat shortened to make it more readable. I.e., the artificial prefix before parameter names is suppressed.

See Also

Other multiplexer: makeModelMultiplexerParamSet

Other tune: TuneControl, TuneControlCMAES, TuneControlGrid, TuneControlIrace, TuneControlOptim, TuneControlRandom, makeTuneControlCMAES, makeTuneControlGrid, makeTuneControlIrace, makeTuneControlOptim, makeTuneControlRandom; getTuneResult; makeModelMultiplexerParamSet; makeTuneWrapper; tuneParams; tuneThreshold

Examples

bls = list(
  makeLearner("classif.ksvm"),
  makeLearner("classif.randomForest")
)

## Loading required package: randomForest
## randomForest 4.6-7
## Type rfNews() to see new features/changes/bug fixes.

lrn = makeModelMultiplexer(bls)
# simple way to contruct param set for tuning
# parameter names are prefixed automatically and the 'requires'
# element is set, too, to make all paramaters subordinate to 'selected.learner'
ps = makeModelMultiplexerParamSet(lrn,
  makeNumericParam("sigma", lower=-10, upper = 10, trafo = function(x) 2^x),
  makeIntegerParam("ntree", lower = 1L, upper = 500L)
)
print(ps)

##                                Type len Def
## selected.learner           discrete   -   -
## classif.ksvm.sigma          numeric   -   -
## classif.randomForest.ntree  integer   -   -
##                                                       Constr Req Trafo
## selected.learner           classif.ksvm,classif.randomForest   -     -
## classif.ksvm.sigma                                 -10 to 10   Y     Y
## classif.randomForest.ntree                          1 to 500   Y     -

rdesc = makeResampleDesc("CV", iters = 2L)
# to save some time we use random search. but you probably want something like this:
# ctrl = makeTuneControlIrace(maxExperiments = 500L)
ctrl = makeTuneControlRandom(maxit = 4L)
res = tuneParams(lrn, iris.task, rdesc, par.set = ps, control = ctrl)

## [Tune] Started tuning learner ModelMultiplexer for parameter set:
##                                Type len Def
## selected.learner           discrete   -   -
## classif.ksvm.sigma          numeric   -   -
## classif.randomForest.ntree  integer   -   -
##                                                       Constr Req Trafo
## selected.learner           classif.ksvm,classif.randomForest   -     -
## classif.ksvm.sigma                                 -10 to 10   Y     Y
## classif.randomForest.ntree                          1 to 500   Y     -
## With control class: TuneControlRandom
## [Tune] 1: selected.learner=classif.randomForest; ntree=22 : mmce.test.mean=0.04
## [Tune] 2: selected.learner=classif.randomForest; ntree=211 : mmce.test.mean=0.0467
## [Tune] 3: selected.learner=classif.ksvm; sigma=0.00239 : mmce.test.mean=0.427
## [Tune] 4: selected.learner=classif.ksvm; sigma=0.216 : mmce.test.mean=0.0533
## [Tune] Result: selected.learner=classif.randomForest; classif.randomForest.ntree=22 : mmce.test.mean=0.04

print(res)

## Tune result:
## Op. pars: selected.learner=classif.randomForest; classif.randomForest.ntree=22
## mmce.test.mean=0.04

print(head(as.data.frame(res$opt.path)))

##       selected.learner classif.ksvm.sigma classif.randomForest.ntree
## 1 classif.randomForest                 NA                         22
## 2 classif.randomForest                 NA                        211
## 3         classif.ksvm           0.002391                         NA
## 4         classif.ksvm           0.215967                         NA
##   mmce.test.mean dob eol
## 1        0.04000   1  NA
## 2        0.04667   2  NA
## 3        0.42667   3  NA
## 4        0.05333   4  NA

# more unique and reliable way to construct the param set
ps = makeModelMultiplexerParamSet(lrn,
  classif.ksvm = makeParamSet(
    makeNumericParam("sigma", lower=-10, upper = 10, trafo = function(x) 2^x)
  ),
  classif.randomForest = makeParamSet(
    makeIntegerParam("ntree", lower = 1L, upper = 500L)
  )
)

# this is how you would construct the param set manually, works too
ps = makeParamSet(
  makeDiscreteParam("selected.learner", values = extractSubList(bls, "id")),
  makeNumericParam("classif.ksvm.sigma", lower=-10, upper = 10, trafo = function(x) 2^x,
    requires = quote(selected.learner == "classif.ksvm")),
  makeIntegerParam("classif.randomForest.ntree", lower = 1L, upper = 500L,
    requires = quote(selected.learner == "classif.randomForst"))
)

# all three ps-objects are exactly the same internally.

[Package mlr version 1.2 Index]