makeOverBaggingWrapper {mlr}	R Documentation

Fuse learner with the bagging technique and oversampling for imbalancy correction.

Description

Fuses a classification learner for binary classification with an over-bagging method for imbalancy correction when we have strongly unequal class sizes. Creates a learner object, which can be used like any other learner object. Models can easily be accessed via getBaggingModels.

OverBagging is implemented as follows: For each iteration a random data subset is sampled. Minority class examples are oversampled with replacement with a given rate. Majority class examples are either simply copied into each bag, or bootstrapped with replacement until we have as many majority class examples as in the original training data. Features are currently not changed or sampled.

Prediction works as follows: For classification we do majority voting to create a discrete label and probabilities are predicted by considering the proportions of all predicted labels.

Usage

makeOverBaggingWrapper(learner, obw.iters = 10L, obw.rate = 1,
  obw.maxcl = "boot")

Arguments

learner
[Learner | character(1)]
The learner. If you pass a string the learner will be created via makeLearner.

obw.iters
[integer(1)]
Number of fitted models in bagging. Default is 10.

obw.rate
[numeric(1)]
Factor to upsample the smaller class in each bag. Must be between 1 and Inf, where 1 means no oversampling and 2 would mean doubling the class size. Default is 1.

obw.maxcl
[character(1)]
character value that controls how to sample majority class. “all” means every instance of the majority class gets in each bag, “boot” means the majority class instances are bootstrapped in each iteration. Default is “boot”.

Value

[Learner].

See Also

Other imbalancy: makeOversampleWrapper, makeUndersampleWrapper; oversample, undersample; smote

[Package mlr version 2.1 Index]