makePreprocWrapper {mlr}	R Documentation

Fuse learner with preprocessing.

Description

Fuses a base learner with a preprocessing method. Creates a learner object, which can be used like any other learner object, but which internally preprocesses the data as requested. If the train or predict function is called on data / a task, the preprocessing is always performed automatically.

Usage

makePreprocWrapper(learner, train, predict, par.set = makeParamSet(),
  par.vals = list())

Arguments

learner
[Learner | character(1)]
The learner. If you pass a string the learner will be created via makeLearner.

train
[function(data, target, args)]
Function to preprocess the data before training. target is a string and denotes the target variable in data. args is a list of further arguments and parameters to influence the preprocessing. Must return a list(data, control), where data is the preprocessed data and control stores all information necessary to do the preprocessing before predictions.

predict
[function(data, target, args, control)]
Function to preprocess the data before prediction. target is a string and denotes the target variable in data. args are the args that were passed to train. control is the object you returned in train.

par.set
[ParamSet]
Parameter set of LearnerParam objects to describe the parameters in args. Default is empty set.

par.vals
[list]
Named list of default values for params in args repectively par.set. Default is empty list.

Value

[Learner].

[Package mlr version 2.1 Index]