makeTuneWrapper {mlr}	R Documentation

Fuse learner with tuning.

Description

Fuses a base learner with a search strategy to select its hyperparameters. Creates a learner object, which can be used like any other learner object, but which internally uses tune. If the train function is called on it, the search strategy and resampling are invoked to select an optimal set of hyperparameter values. Finally, a model is fitted on the complete training data with these optimal hyperparameters and returned. See tuneParams for more details.

Usage

  makeTuneWrapper(learner, resampling, measures, par.set,
    control, show.info = TRUE)

Arguments

learner
[Learner or string]
Learning algorithm.

resampling
[ResampleDesc | ResampleInstance]
Resampling strategy to evaluate points in hyperparameter space.

measures
[list of Measure]
Performance measures to evaluate. The first measure, aggregated by the first aggregation function is optimized during tuning, others are simply evaluated.

par.set
[ParamSet]
Collection of parameters and their constraints for optimization.

control
[TuneControl]
Control object for search method. Also selects the optimization algorithm for tuning.

show.info
[logical(1)]
Show info message after each hyperparameter evaluation? Default is TRUE.

Details

After training, the optimal hyperparameters (and other related information) can be retrieved with getTuneResult.

Value

[Learner].

Examples

task = makeClassifTask(data=iris, target="Species")
lrn = makeLearner("classif.ksvm")

## Lade nötiges Paket: kernlab

# stupid mini grid
ps = makeParamSet(
  makeDiscreteParam("C", values = 1:2),
  makeDiscreteParam("sigma", values = 1:2)
)
ctrl = makeTuneControlGrid()
inner = makeResampleDesc("Holdout")
outer = makeResampleDesc("CV", iters = 2)
lrn = makeTuneWrapper(lrn, resampling = inner, par.set = ps, control = ctrl)
mod = train(lrn, task)

## [Tune] Started tuning learner classif.ksvm for parameter set:
## Disc param 'C'. Vals: 1,2. Trafo: FALSE. Requires: FALSE
## Disc param 'sigma'. Vals: 1,2. Trafo: FALSE. Requires: FALSE
## With control class: TuneControlGrid
## [Tune] 1: C=1,sigma=1 : mmce.test.mean=0.02
## [Tune] 1: C=2,sigma=1 : mmce.test.mean=0.02
## [Tune] 1: C=1,sigma=2 : mmce.test.mean=0.02
## [Tune] 1: C=2,sigma=2 : mmce.test.mean=0.04
## [Tune] Result: C=1,sigma=2 : =0.02

print(getTuneResult(mod))

## Tune result:
## Op. pars: C=1,sigma=2
## =0.02

# nested resampling for evaluation
# we also extract tuned hyper pars in each iteration
r = resample(lrn, task, outer, extract = function(model) {
  getTuneResult(model)$x
})

## [Resample] cross-validation iter: 1
## [Tune] Started tuning learner classif.ksvm for parameter set:
## Disc param 'C'. Vals: 1,2. Trafo: FALSE. Requires: FALSE
## Disc param 'sigma'. Vals: 1,2. Trafo: FALSE. Requires: FALSE
## With control class: TuneControlGrid
## [Tune] 1: C=1,sigma=1 : mmce.test.mean=0.04
## [Tune] 1: C=2,sigma=1 : mmce.test.mean=0.04
## [Tune] 1: C=1,sigma=2 : mmce.test.mean=0.04
## [Tune] 1: C=2,sigma=2 : mmce.test.mean=0.04
## [Tune] Result: C=2,sigma=2 : =0.04
## [Resample] cross-validation iter: 2
## [Tune] Started tuning learner classif.ksvm for parameter set:
## Disc param 'C'. Vals: 1,2. Trafo: FALSE. Requires: FALSE
## Disc param 'sigma'. Vals: 1,2. Trafo: FALSE. Requires: FALSE
## With control class: TuneControlGrid
## [Tune] 1: C=1,sigma=1 : mmce.test.mean=   0
## [Tune] 1: C=2,sigma=1 : mmce.test.mean=   0
## [Tune] 1: C=1,sigma=2 : mmce.test.mean=   0
## [Tune] 1: C=2,sigma=2 : mmce.test.mean=   0
## [Tune] Result: C=1,sigma=2 : =   0
## [Resample] Result: mmce.test.mean=0.0533

print(r$extract)

## [[1]]
## [[1]]$C
## [1] 2
## 
## [[1]]$sigma
## [1] 2
## 
## 
## [[2]]
## [[2]]$C
## [1] 1
## 
## [[2]]$sigma
## [1] 2

[Package mlr version 1.2 Index]