makeUndersampleWrapper {mlr}	R Documentation

Fuse learner with simple over/undersampling for binary classification.

Description

Creates a learner object, which can be used like any other learner object. Internally uses oversample or undersample before every model fit.

Usage

makeUndersampleWrapper(learner, usw.rate)

makeOversampleWrapper(learner, osw.rate)

Arguments

learner
[Learner]
The learner.

usw.rate
[numeric(1)]
Factor to downsample the bigger class. Must be between 0 and 1, where 1 means no downsampling, 0.5 implies reduction to 50 percent and 0 would imply reduction to 0 observations.

osw.rate
[numeric(1)]
Factor to oversample the smaller class. Must be between 1 and Inf, where 1 means no oversampling and 2 would mean doubling the class size.

Details

Note that observation weights do not influence the sampling and are simply passed down to the next learner.

Value

[Learner].

See Also

Other OverUndersample: oversample, undersample

[Package mlr version 1.2 Index]