makeWrappedModel {mlr}	R Documentation

Induced model of learner.

Description

Result from train.

It internally stores the underlying fitted model, the subset used for training, features used for training, levels of factors in the data set and computation time that was spent for training.

Object members: See arguments.

The constructor makeWrappedModel is mainly for internal use.

Usage

makeWrappedModel(learner, learner.model, task.desc, subset, features,
  factor.levels, time)

Arguments

learner
[Learner | character(1)]
The learner. If you pass a string the learner will be created via makeLearner.

learner.model
[any]
Underlying model.

task.desc
[TaskDesc]
Task description object.

subset
[integer]
Subset used for training.

features
[character]
Features used for training.

factor.levels
[named list of character]
Levels of factor variables (features and potentially target) in training data. Named by variable name, non-factors do not occur in the list.

time
[numeric(1)]
Computation time for model fit in seconds.

Value

[WrappedModel].

[Package mlr version 2.1 Index]