measures {mlr}	R Documentation

Performance measures.

Description

Performance measures can always be passed as a single string (name of a single measure), a character vector of multiple names of measures or a list containing string names of measures and your on performance measures as function objects. The latter ones should be named list elements.

Details

Classification:

mmce
Mean misclassification error

acc
Accuracy

costs
Misclassification costs according to cost matrix

tp
True positives

tpr, hit-rate, recall
True positive rate

fp, false-alarm
False positives

fpr, false-alarm-rate, fall-out
False positive rate

tn, correct-rejection
True negatives

tnr, specificity
True negative rate

fn, miss
False negatives

fnr
False negative rate

ppv, precision
Positive predictive value

npv
Negative predictive value

fdr
False discovery rate

f1
F1 measure

mcc
Matthews correlation coefficient

gmean
G-mean, geomentric mean of recall and specificity.

gpr
Geometric mean of precision and recall.

auc
Area under the curve.

time.train
Time of fitting the model

time.predict
Time of predicting test set

time
time.train + train.predict

Regression:

sse
Sum of squared errors

mse
Mean of squared errors

medse
Median of squared errors

sae
Sum of absolute errors

mae
Mean of absolute errors

medae
Median of absolute errors

time.train
Time of fitting the model

time.predict
Time of predicting test set

time
time.train + train.predict

[Package mlr version 0.3.1095 Index]