mlr: Machine Learning in R

Interface to a large number of classification and regression techniques, including machine-readable parameter descriptions. There is also an experimental extension for survival analysis, clustering and general, example-specific cost-sensitive learning. Generic resampling, including cross-validation, bootstrapping and subsampling. Hyperparameter tuning with modern optimization techniques, for single- and multi-objective problems. Filter and wrapper methods for feature selection. Extension of basic learners with additional operations common in machine learning, also allowing for easy nested resampling. Most operations can be parallelized.

See also

Author

Maintainer: Patrick Schratz patrick.schratz@gmail.com (ORCID)

Authors:

Bernd Bischl bernd_bischl@gmx.net (ORCID)

Michel Lang michellang@gmail.com (ORCID)

Lars Kotthoff larsko@uwyo.edu

Julia Schiffner schiffner@math.uni-duesseldorf.de

Jakob Richter code@jakob-r.de

Zachary Jones zmj@zmjones.com

Giuseppe Casalicchio giuseppe.casalicchio@stat.uni-muenchen.de (ORCID)

Mason Gallo masonagallo@gmail.com

Other contributors:

Jakob Bossek jakob.bossek@tu-dortmund.de (ORCID) [contributor]

Erich Studerus erich.studerus@upkbs.ch (ORCID) [contributor]

Leonard Judt leonard.judt@tu-dortmund.de [contributor]

Tobias Kuehn tobi.kuehn@gmx.de [contributor]

Pascal Kerschke kerschke@uni-muenster.de (ORCID) [contributor]

Florian Fendt flo_fendt@gmx.de [contributor]

Philipp Probst philipp_probst@gmx.de (ORCID) [contributor]

Xudong Sun xudong.sun@stat.uni-muenchen.de (ORCID) [contributor]

Janek Thomas janek.thomas@stat.uni-muenchen.de (ORCID) [contributor]

Bruno Vieira bruno.hebling.vieira@usp.br [contributor]

Laura Beggel laura.beggel@web.de (ORCID) [contributor]

Quay Au quay.au@stat.uni-muenchen.de (ORCID) [contributor]

Martin Binder ma.binder@campus.lmu.de [contributor]

Florian Pfisterer pfistererf@googlemail.com [contributor]

Stefan Coors stefan.coors@gmx.net [contributor]

Steve Bronder sab2287@columbia.edu [contributor]

Alexander Engelhardt alexander.w.engelhardt@gmail.com [contributor]

Christoph Molnar christoph.molnar@stat.uni-muenchen.de [contributor]

Annette Spooner a.spooner@unsw.edu.au [contributor]