opt.result-class {mlr}	R Documentation

Optimization result.

Description

Container for results of hyperparameter tuning or variable selection. Contains the obtained optimal parameter vector, its performance values and the optimization path which lead there. It might also optionally contain a wrapped.model, which was fitted by using the optimal parameters.

Extends

object

Methods

[

signature( = "opt.result"): Getter.

to.string

See Also

tune, varsel

[Package mlr version 0.3.528 Index]