opt.wrapper-class {mlr}	R Documentation

Abstract base class to wrap an optimization algorithm around a learner.

Description

Abstract base class to wrap an optimization algorithm around a learner.

Extends

base.wrapper

Methods

[

train.learner

[Package mlr version 0.3.1095 Index]