oversample {mlr}	R Documentation

Over- or undersample binary classification task to handle class imbalancy.

Description

Oversampling: From the smaller class, observations are randomly drawn with repetitions.

Undersampling: From the larger class, observations are randomly drawn without repetitions.

Usage

oversample(task, rate)

undersample(task, rate)

Arguments

task
[Task]
The task.

rate
[numeric(1)]
Factor to upsample the smaller or downsample the bigger class. For undersampling: Must be between 0 and 1, where 1 means no downsampling, 0.5 implies reduction to 50 percent and 0 would imply reduction to 0 observations. For oversampling: Must be between 1 and Inf, where 1 means no oversampling and 2 would mean doubling the class size.

Value

[Task].

See Also

Other imbalancy: makeOverBaggingWrapper; makeOversampleWrapper, makeUndersampleWrapper; smote

[Package mlr version 2.1 Index]