parallel.setup {mlr}	R Documentation

Parallelization setup.

Description

Defines the technical mode and level of parallelization when executing code.

Usage

parallel.setup(mode="local", parallel.type, cpus, level="resample", ...)

Arguments

mode 	 [string]
Which parallel mode should be used: local, multicore, snowfall, sfCluster.

parallel.type 	 [string]
Currently this is only used for sfInit, where it is passed to the 'type' argument. Default is 'MPI'.

cpus 	 [numeric]
Number of requested cpus. Default is mpi.universe.size() for snowfall/MPI, ignored for for sfCluster and 1 otherwise.

level 	 [string]
What is parallelized / what is a job. resample: resample.fit is parallelized and a job is train / test. tune: tune is parallelized and a job is a resampled evaluation of one hyperparameter setting. varsel: varsel is parallelized and a job is a resampled evaluation of a feature set. bench: bennch.exp is parallelized and a job is completely evaluating one learner on one data set.

... 	 [any]
Optional parameters, only passed to sfInit currently.

[Package mlr version 0.3.732 Index]