Measure performance of prediction.

Measures the quality of a prediction w.r.t. some performance measure.

performance(pred, measures, task = NULL, model = NULL, feats = NULL)

Arguments

pred
[Prediction] Prediction object.

measures
[Measure | list of Measure] Performance measure(s) to evaluate. Default is the default measure for the task, see here getDefaultMeasure.

task
[Task] Learning task, might be requested by performance measure, usually not needed except for clustering or survival.

model
[WrappedModel] Model built on training data, might be requested by performance measure, usually not needed except for survival.

feats
[data.frame] Features of predicted data, usually not needed except for clustering. If the prediction was generated from a task, you can also pass this instead and the features are extracted from it.

Value

[named numeric]. Performance value(s), named by measure(s).

See also

Other performance: ConfusionMatrix, calculateConfusionMatrix, calculateROCMeasures, estimateRelativeOverfitting, makeCostMeasure, makeCustomResampledMeasure, makeMeasure, measures, setAggregation, setMeasurePars

Examples

training.set = seq(1, nrow(iris), by = 2)
test.set = seq(2, nrow(iris), by = 2)

task = makeClassifTask(data = iris, target = "Species")
lrn = makeLearner("classif.lda")
mod = train(lrn, task, subset = training.set)
pred = predict(mod, newdata = iris[test.set, ])
performance(pred, measures = mmce)

#> mmce 
#> 0.04 

# Compute multiple performance measures at once
ms = list("mmce" = mmce, "acc" = acc, "timetrain" = timetrain)
performance(pred, measures = ms, task, mod)

#>      mmce       acc timetrain 
#>     0.040     0.960     0.004