Create box or violin plots for a BenchmarkResult.

Plots box or violin plots for a selected measure across all iterations of the resampling strategy, faceted by the task.id.

Usage,

plotBMRBoxplots(
  bmr,
  measure = NULL,
  style = "box",
  order.lrns = NULL,
  order.tsks = NULL,
  pretty.names = TRUE,
  facet.wrap.nrow = NULL,
  facet.wrap.ncol = NULL
)

Arguments

bmr

(BenchmarkResult)
Benchmark result.

measure

(Measure)
Performance measure. Default is the first measure used in the benchmark experiment.

style

(character(1))
Type of plot, can be “box” for a boxplot or “violin” for a violin plot. Default is “box”.

order.lrns

(character(n.learners))
Character vector with learner.ids in new order.

order.tsks

(character(n.tasks))
Character vector with task.ids in new order.

pretty.names

(logical(1))
Whether to use the Measure name and the Learner short name instead of the id. Default is TRUE.

facet.wrap.nrow, facet.wrap.ncol

(integer)
Number of rows and columns for facetting. Default for both is NULL. In this case ggplot's facet_wrap will choose the layout itself.

Value

ggplot2 plot object.

See also

Examples

# see benchmark