Plot threshold vs. performance(s) for 2-class classification using ggplot2.

Plots threshold vs. performance(s) data that has been generated with generateThreshVsPerfData.

plotThreshVsPerf(obj, measures = obj$measures, facet = "measure",
  mark.th = NA_real_, pretty.names = TRUE, facet.wrap.nrow = NULL,
  facet.wrap.ncol = NULL)

Arguments

obj
[ThreshVsPerfData] Result of generateThreshVsPerfData.

measures
[Measure | list of Measure] Performance measure(s) to plot. Must be a subset of those used in generateThreshVsPerfData. Default is all the measures stored in obj generated by generateThreshVsPerfData.

facet
[character(1)] Selects “measure” or “learner” to be the facetting variable. The variable mapped to facet must have more than one unique value, otherwise it will be ignored. The variable not chosen is mapped to color if it has more than one unique value. The default is “measure”.

mark.th
[numeric(1)] Mark given threshold with vertical line? Default is NA which means not to do it.

pretty.names
[logical(1)] Whether to use the Measure name instead of the id in the plot. Default is TRUE.

facet.wrap.nrow, facet.wrap.ncol
[integer()] Number of rows and columns for facetting. Default for both is NULL. In this case ggplot's facet_wrap will choose the layout itself.

Value

ggplot2 plot object.

See also

Other plot: plotBMRBoxplots, plotBMRRanksAsBarChart, plotBMRSummary, plotCalibration, plotCritDifferences, plotFilterValuesGGVIS, plotLearningCurveGGVIS, plotLearningCurve, plotPartialDependenceGGVIS, plotPartialDependence, plotROCCurves, plotResiduals, plotThreshVsPerfGGVIS

Other thresh_vs_perf: generateThreshVsPerfData, plotROCCurves, plotThreshVsPerfGGVIS

Examples

lrn = makeLearner("classif.rpart", predict.type = "prob")
mod = train(lrn, sonar.task)

#> Error: Please use column names for `x`

pred = predict(mod, sonar.task)

#> Error in predict(mod, sonar.task): object 'mod' not found

pvs = generateThreshVsPerfData(pred, list(acc, setAggregation(acc, train.mean)))

#> Error in generateThreshVsPerfData(pred, list(acc, setAggregation(acc,     train.mean))): object 'pred' not found

plotThreshVsPerf(pvs)

#> Error in checkClass(x, classes, ordered, null.ok): object 'pvs' not found