Plots multi-criteria results after tuning using ggplot2.

Visualizes the pareto front and possibly the dominated points.

plotTuneMultiCritResult(res, path = TRUE, col = NULL, shape = NULL,
  pointsize = 2, pretty.names = TRUE)

Arguments

res
[TuneMultiCritResult] Result of tuneParamsMultiCrit.

path
[logical(1)] Visualize all evaluated points (or only the non-dominated pareto front)? For the full path, the size of the points on the front is slightly increased. Default is TRUE.

col
[character(1)] Which column of res$opt.path should be mapped to ggplot2 color? Default is NULL, which means none.

shape
[character(1)] Which column of res$opt.path should be mapped to ggplot2 shape? Default is NULL, which means none.

pointsize
[numeric(1)] Point size for ggplot2 geom_point for data points. Default is 2.

pretty.names
[logical{1}] Whether to use the ID of the measures instead of their name in labels. Defaults to TRUE.

Value

ggplot2 plot object.

See also

Other tune_multicrit: TuneMultiCritControl, plotTuneMultiCritResultGGVIS, tuneParamsMultiCrit

Examples

# see tuneParamsMultiCrit