Visualize binary classification predictions via ViperCharts system.

This includes ROC, lift charts, cost curves, and so on. Please got to http://viper.ijs.si for further info.

For resampled learners, the predictions from different iterations are combined into one. That is, for example for cross-validation, the predictions appear on a single line even though they were made by different models. There is currently no facility to separate the predictions for different resampling iterations.

plotViperCharts(obj, chart = "rocc", browse = TRUE, auth.key = NULL,
  task.id = NULL)

Arguments

obj
[(list of) Prediction | (list of) ResampleResult | BenchmarkResult] Single prediction object, list of them, single resample result, list of them, or a benchmark result. In case of a list probably produced by different learners you want to compare, then name the list with the names you want to see in the plots, probably learner shortnames or ids.

chart
[character(1)] First chart to display in focus in browser. All other charts can be displayed by clicking on the browser page menu. Default is “rocc”.

browse
[logical(1)] Open ViperCharts plot in web browser? If not you simple get the URL returned. Calls browseURL. Default is TRUE.

auth.key
[character(1)] API key to use for call to Viper charts website. Only required if you want the chart to be private. Default is NULL.

task.id
[character(1)] Selected task in BenchmarkResult to do plots for, ignored otherwise. Default is first task.

Value

[character(1)]. Invisibly returns the ViperCharts URL.

References

Sluban and Lavrač - ViperCharts: Visual Performance Evaluation Platform, ECML PKDD 2013, pp. 650-653, LNCS 8190, Springer, 2013.

See also

Other roc: asROCRPrediction, calculateROCMeasures

Other predict: asROCRPrediction, getPredictionProbabilities, getPredictionResponse, getPredictionTaskDesc, predict.WrappedModel, setPredictThreshold, setPredictType

Examples

# NOT RUN {
lrn1 = makeLearner("classif.logreg", predict.type = "prob")
lrn2 = makeLearner("classif.rpart", predict.type = "prob")
b = benchmark(list(lrn1, lrn2), pid.task)
z = plotViperCharts(b, chart = "lift", browse = TRUE)
# }