predict {mlr}	R Documentation

Predict new data.

Description

Predict the target variable of new data using a fitted model.

Arguments

object 	 [wrapped.model]
Wrapped model, trained from a learn task.

task 	 [learn.task]
Specifies learning task. If this is passed, data from this task is predicted.

subset 	 [integer]
Index vector to subset the data in the task to use for prediction.

newdata 	 [data.frame]
New observations which should be predicted. Alternatively pass this instead of task.

type 	 [character]
Classification: vector of "response" | "prob" | "decision", specifying the types to predict. Default is "response". Ignored for regression.

Value

prediction.

See Also

train

[Package mlr version 0.3.528 Index]