prediction-class {mlr}	R Documentation

Prediction.

Description

Result from predict. Use as.data.frame to access all information in a convenient format.

Details

prediction-class: Getter.

id [numeric]

Vector of index numbers of predicted cases from the task.

response [numeric | factor]

Predicted response values.

truth [numeric | factor]

True target values.

prob [numeric | matrix] Optional parameters: class

Predicted probabilities. For binary class. only the probs for the pos. class are returned.

decision [matrix]

Predicted decision values.

type [string]

Type set in predict function: "response", "prob", or "decision".

threshold [numeric]

Threshold set in predict function.

Extends

object

Methods

performance

[

signature( = "prediction"): Getter.

as.data.frame

signature( = "prediction"): Convert to data.frame

to.string

See Also

performance

[Package mlr version 0.3.732 Index]