reimpute {mlr}	R Documentation

Re-impute a data set

Description

This function accepts a data frame and a imputation description as returned by impute to perform the following actions:

Restore dropped columns, setting them to NA

Add dummy variables for columns as specified in impute

Optionally check factors for new levels to treat them as NAs

Reorder factor levels to ensure identical integer representation as before

Impute missing values using previously collected data

Usage

reimpute(x, desc)

Arguments

x
[data.frame]
Object to reimpute. Currently only data frames are supported.

desc
[ImputationDesc]
Imputation description as returned by impute.

Value

Imputated x.

[Package mlr version 1.2 Index]