removeHyperPars {mlr}	R Documentation

Remove hyperparameters settings of a learner.

Description

Remove settings (previously set through mlr) for some parameters. Which means that the default behavior for that param will now be used.

Usage

removeHyperPars(learner, ids = character(0L))

Arguments

learner
[Learner | character(1)]
The learner. If you pass a string the learner will be created via makeLearner.

ids
[character]
Parameter names to remove settings for. Default is character(0L).

Value

[Learner].

See Also

Other learner: getHyperPars; getParamSet; setHyperPars; setId; setPredictType; showHyperPars

[Package mlr version 2.1 Index]