resample.desc-class {mlr}	R Documentation

resample.desc

Description

Base class for description of resampling algorithms. A description of a resampling algorithm contains all necessary information to provide a resampling.instance, when given the size of the data set.

Arguments

x 	 resample.desc object

i

instance.class

S4 class name of the corresponding resample.instance

name

Name of this resampling algorithm

iters

Number of iterations

Extends

object

Methods

[

signature( = "resample.desc"): Getter.

to.string

resample.fit

make.res.instance

make.res.instance

[Package mlr version 0.3.528 Index]