resample.instance-class {mlr}	R Documentation

Resampling instance.

Description

Base class for specific resampling draws like cross-validation or bootstrapping. This class encapsulates training and test sets generated from the data set for a number of iterations. It mainly stores a set of integer vectors indicating the training examples for each iteration. Don't create objects from this class directly but use the corresponding subclasses. For construction simply use the factory method make.res.instance.

Arguments

x 	 resample.instance object

i 	 [character]

size

Number of observations.

name

The name of the resample description object, i.e. the type of resampling.

iters

The number of resampling iterations.

train.inds

If j is a single integer, the vector of training indices for the jth iteration is returned. If j is an integer vector, the list of training indices for the given iterations is returned. If j is missing, all indices are returned.

test.inds

If j is a single integer, the vector of test indices for the jth iteration is returned. If j is an integer vector, the list of test indices for the given iterations is returned. If j is missing, all indices are returned.

j 	 [integer]
See above, i == "train.inds" or i == "test.inds".

Extends

object

Methods

resample.fit

to.string

[

signature( = "resample.instance"): Getter.

Note

If you want to add another resampling strategy, have a look at the web documentation.

See Also

resample.desc, make.res.instance, resample.fit

[Package mlr version 0.3.528 Index]