seq.control {mlr}	R Documentation

seq.control-class

Description

Control structure for sequential variable selection.

Arguments

minimize 	 [logical]
Minimize performance measure? Default is TRUE.

maxit 	 [integer]
Maximal number of variable sets to evaluate. Default is 100.

max.vars 	 [integer]
Maximal number of allowed variables in the final set. Default is max. integer.

alpha 	 [numeric]
sfs, sffs: In a forward step, minimal improvement of performance measure. Can be negative.

beta 	 [numeric]
sbs, sfbs: In a backward step, minimal improvement of performance measure. Can be negative.

tune.threshold 	 [logical]
Perform empirical thresholding? Default is FALSE. Only supported for binary classification and you have to set predict.type to "prob" for this in make.learner.

thresholds 	 [numeric]
Number of thresholds to try in tuning. Predicted probabilities are sorted and divided into groups of equal size. Default is 10.

Value

seq.control-methods: Control structure.

Extends

varsel.control

Methods

initialize

[Package mlr version 0.3.732 Index]