setHyperPars {mlr}	R Documentation

Set the hyperparameters of a learner object.

Description

Set the hyperparameters of a learner object.

Usage

setHyperPars(learner, ..., par.vals)

Arguments

learner
[Learner]
The learner.

...
[any]
Named (hyper)parameters with new setting. Alternatively these can be passed using the par.vals argument.

par.vals
[list]
Optional list of named (hyper)parameter settings. The arguments in ... take precedence over values in this list.

Value

[Learner] with changed hyperparameters.

See Also

See getHyperPars for a function to retrieve the currently set hyperparameters. To get a list of all hyperparameters of a learner, see the par.set slot of the Learner object.

Examples

cl1 <- makeLearner("classif.ksvm", sigma=1)
cl2 <- setHyperPars(cl1, sigma=10, par.vals=list(C=2))
print(cl1)

## Learner classif.ksvm from package kernlab
## Type: classif
## Class: classif.ksvm
## Predict-Type: response
## Hyperparameters: fit=FALSE,sigma=1.00
## 
## Supported features Numerics:TRUE Factors:TRUE
## Supports missings: FALSE
## Supports weights: FALSE
## Supports classes: two,multi
## Supports probabilities: TRUE

# note the now set and altered hyperparameters:
print(cl2)

## Learner classif.ksvm from package kernlab
## Type: classif
## Class: classif.ksvm
## Predict-Type: response
## Hyperparameters: fit=FALSE,sigma=10.00,C=2.00
## 
## Supported features Numerics:TRUE Factors:TRUE
## Supports missings: FALSE
## Supports weights: FALSE
## Supports classes: two,multi
## Supports probabilities: TRUE

[Package mlr version 1.2 Index]