Set threshold of prediction object.

Set threshold of prediction object for classification or multilabel classification. Creates corresponding discrete class response for the newly set threshold. For binary classification: The positive class is predicted if the probability value exceeds the threshold. For multiclass: Probabilities are divided by corresponding thresholds and the class with maximum resulting value is selected. The result of both are equivalent if in the multi-threshold case the values are greater than 0 and sum to 1. For multilabel classification: A label is predicted (with entry TRUE) if a probability matrix entry exceeds the threshold of the corresponding label.

setThreshold(pred, threshold)

Arguments

pred
[Prediction] Prediction object.

threshold
[numeric] Threshold to produce class labels. Has to be a named vector, where names correspond to class labels. Only for binary classification it can be a single numerical threshold for the positive class.

Value

[Prediction] with changed threshold and corresponding response.

See also

predict.WrappedModel

Examples

# create task and train learner (LDA)
task = makeClassifTask(data = iris, target = "Species")
lrn = makeLearner("classif.lda", predict.type = "prob")
mod = train(lrn, task)

# predict probabilities and compute performance
pred = predict(mod, newdata = iris)
performance(pred, measures = mmce)

#> mmce 
#> 0.02 

head(as.data.frame(pred))

#>    truth prob.setosa prob.versicolor prob.virginica response
#> 1 setosa           1    3.896358e-22   2.611168e-42   setosa
#> 2 setosa           1    7.217970e-18   5.042143e-37   setosa
#> 3 setosa           1    1.463849e-19   4.675932e-39   setosa
#> 4 setosa           1    1.268536e-16   3.566610e-35   setosa
#> 5 setosa           1    1.637387e-22   1.082605e-42   setosa
#> 6 setosa           1    3.883282e-21   4.566540e-40   setosa

# adjust threshold and predict probabilities again
threshold = c(setosa = 0.4, versicolor = 0.3, virginica = 0.3)
pred = setThreshold(pred, threshold = threshold)
performance(pred, measures = mmce)

#> mmce 
#> 0.02 

head(as.data.frame(pred))

#>    truth prob.setosa prob.versicolor prob.virginica response
#> 1 setosa           1    3.896358e-22   2.611168e-42   setosa
#> 2 setosa           1    7.217970e-18   5.042143e-37   setosa
#> 3 setosa           1    1.463849e-19   4.675932e-39   setosa
#> 4 setosa           1    1.268536e-16   3.566610e-35   setosa
#> 5 setosa           1    1.637387e-22   1.082605e-42   setosa
#> 6 setosa           1    3.883282e-21   4.566540e-40   setosa