showHyperPars {mlr}	R Documentation

Display all possible hyperparameter settings for a learner that mlr knows.

Description

Useful for a quick overview, also does not foce you to create the learner.

Usage

showHyperPars(learner)

Arguments

learner
[Learner | character(1)]
The learner. If you pass a string the learner will be created via makeLearner.

Value

[invisible(NULL)].

[Package mlr version 1.2 Index]