subsetTask {mlr}	R Documentation

Subset data in task.

Description

Subset data in task.

Usage

subsetTask(task, subset, features)

Arguments

task
[SupervisedTask]
The task.

subset
[integer]
Selected cases. Default is all cases.

features
[character]
Selected inputs. Note that target feature is always included in the resulting task, you should not pass it here. Default is all features.

Value

[SupervisedTask]. Task with subsetted data.

Examples

task <- makeClassifTask(data = iris, target = "Species")
subsetTask(task, subset = 1:100)

## Supervised task: data
## Type: classif
## Target: Species
## Observations: 100
## Features:
## numerics  factors 
##        4        0 
## Missings: FALSE
## Has weights: FALSE
## Has blocking: FALSE
## Classes: 3
##     setosa versicolor  virginica 
##         50         50          0 
## Positive class: NA

[Package mlr version 1.2 Index]