train {mlr}	R Documentation

Train a learning algorithm.

Description

Given a learn.task, creates a model for the learning machine which can be used for predictions on new data.

Usage

train(learner, task, subset, parset, vars, type)

Arguments

learner 	 [learner or string]
Learning algorithm. See learners.

task 	 [learn.task]
Specifies learning task.

subset 	 [integer]
An index vector specifying the training cases to be used for fitting. By default the complete data set is used.

parset 	 [list]
Named list of hyperparameter values. Will overwrite the ones specified in the learner object. Default is empty list.

vars 	 [character]
Vector of variable names to use in training the model. Default is to use all variables, except the excluded in the task.

type 	 [string]
Classification: "response" | "prob" | "decision", specifying the type to predict later. Default is "response". Very rarely you have to set this during training as well, as the fitted models differ.

Value

train-methods: An object of class wrapped.model.

See Also

predict

[Package mlr version 0.3.819 Index]