trainLearner {mlr}	R Documentation

Train an R learner.

Description

Mainly for internal use. Trains a wrapped learner on a given training set. You have to implement this method if you want to add another learner to this package.

Usage

  trainLearner(.learner, .task, .subset, ...)

Arguments

.learner
[RLearner]
Wrapped learner.

.task
[SupervisedTask]
Task to train learner on.

.subset
[integer]
Subset of cases for training set, index the task with this. You probably want to use getTaskData for this purpose.

...
[any]
Additional (hyper)parameters, which need to be passed to the underlying train function.

Details

Your implementation must adhere to the following: The model must be fitted on the subset of .task given by .subset. All parameters must in ... must be passed to the underlying training function.

Value

[any]. Model of the underlying learner.

[Package mlr version 1.2 Index]