tune {mlr}	R Documentation

Hyperparameter tuning

Description

Optimizes the hyperparameters of a learner for a classification or regression problem. Allows for different optimization methods, commonly grid search is used but other search techniques are available as well. The specific details of the search algorithm are set by passing a control object.

Usage

tune(learner, task, resampling, method="grid", control, measures, aggr, model=F)

Arguments

learner 	 [wrapped.learner or string]
Learning algorithm. See learners.

task 	 [learn.task]
Learning task.

resampling 	 [resample.instance] or [resample.desc]
Resampling strategy to evaluate points in hyperparameter space.

method 	 [character]
Search method. Currently supported are grid search "grid", pattern search "pattern", CMA-ES "cmaes" and Nelder-Mead "nm".

control 	 Control object for search method.

measures 	 [see measures]
Performance measures.

aggr 	 [see aggregations]
Aggregation functions.

model 	 [logical]
Should a final model be fitted on the complete data with the best found hyperparameters?

Details

The first measure, aggregated by the first aggregation function is optimized, to find a set of optimal hyperparameters.

Value

opt.result.

See Also

grid.control, ps.control, cmaes.control, nm.control

[Package mlr version 0.3.528 Index]