tune.threshold {mlr}	R Documentation

Tune prediction threshold.

Description

Optimizes the threshold of prediction based on probabilities or decision values. Currently only implemented for probabilities and binary classification.

Usage

tune.threshold(pred, measures, aggr, task, minimize=T, thresholds=10)

Arguments

pred 	 [prediction]
Prediction object to use for tuning the treshold.

measures 	 [see measures] Performance measures.

aggr 	 [see aggregations] Aggregation functions. Ignored if not a resample.prediction

task 	 [learn.task]
Learning task. Rarely neeeded, only when required for the performance measure.

minimize 	 [logical]
Minimize performance measure? Default is TRUE.

thresholds 	 [integer]
Number of thresholds to try in tuning.

Value

A list with with the following components: "th" is the optimal threshold, pred a prediction object based on "th", "th.seq" a numerical vector of threhold values which were tried and "perf" their respective performance values.

See Also

tune

[Package mlr version 0.3.777 Index]