tuneThreshold {mlr}	R Documentation

Tune prediction threshold.

Description

Optimizes the threshold of prediction based on probabilities. Uses optimize for 2class problems and cma_es for multiclass problems.

Usage

tuneThreshold(pred, measure, task, model, control = list())

Arguments

pred
[Prediction]
Prediction object to use for tuning the treshold.

measure
[Measure]
Performance measure to optimize.

task
[SupervisedTask]
Learning task. Rarely neeeded, only when required for the performance measure.

model
[WrappedModel]
Fitted model. Rarely neeeded, only when required for the performance measure.

control
[list]
Control object for cma_es when used. Default is empty list.

Value

[list]. A named list with with the following components: th is the optimal threshold, perf the performance value.

[Package mlr version 1.2 Index]