varsel {mlr}	R Documentation

Variable selection.

Description

Optimizes the variables for a classification or regression problem by choosing a variable selection wrapper approach. Allows for different optimization methods. The specific details of the search algorithm are set by passing a control object.

Usage

varsel(learner, task, resampling, control, measures, aggr, model=F, path=F)

Arguments

learner 	 [learner or string]
Learning algorithm. See learners.

task 	 [learn.task]
Learning task.

resampling 	 [resample.instance] or [resample.desc]
Resampling strategy to evaluate points in hyperparameter space.

control 	 [see varsel.control] Control object for search method. Also selects the optimization algorithm for feature selection.

measures 	 [see measures]
Performance measures.

aggr 	 [see aggregations]
Aggregation functions.

model 	 [boolean]
Should a final model be fitted on the complete data with the best found features?

path 	 [boolean]
Should optimization path be saved?

Details

The first measure, aggregated by the first aggregation function is optimized, to find a set of optimal variables.

Value

opt.result.

See Also

varsel.control, make.varsel.wrapper

[Package mlr version 0.3.777 Index]