wrapped.learner-class {mlr}	R Documentation

Base class for inducers.

Description

Wraps an already implemented learning method from R to make it accesible to mlr. Also includes a properties object to describe the features of the learner.

Details

wrapped.learner-class: Getter.

id [string]

Id string of learner.

label [string]

Label string of learner.

pack [string]

Package were underlying learner is implemented.

train.fct.pars [list]

Paramters that will be passed to the underlying train function.

predict.fct.pars [list]

Paramters that will be passed to the underlying predict function.

props [learner.props]

Properties object to describe functionality of the learner.

Extends

object

Methods

resample.fit

resample.fit

[

signature( = "wrapped.learner"): Getter.

to.string

[Package mlr version 0.3.528 Index]