BenchmarkResult object.

Result of a benchmark experiment conducted by benchmark with the following members:

results (list of ResampleResult):

A nested list of resample results, first ordered by task id, then by learner id.

measures (list of Measure):

The performance measures used in the benchmark experiment.

learners (list of Learner):

The learning algorithms compared in the benchmark experiment.

The print method of this object shows aggregated performance values for all tasks and learners.

It is recommended to retrieve required information via the getBMR* getter functions. You can also convert the object using as.data.frame.

See also