makeFeatSelControlExhaustive {mlr}	R Documentation

Create control structures for feature selection.

Description

Feature selection method used by selectFeatures. The following methods are available:

FeatSelControlExhaustive

Exhaustive search. All feature sets (up to a certain number of features max.features) are searched.

FeatSelControlRandom

Random search. Features vectors are randomly drawn, up to a certain number of features max.features. A feature is included in the current set with probability prob. So we are basically drawing (0,1)-membership-vectors, where each element is Bernoulli(prob) distributed.

FeatSelControlSequential

Deterministic forward or backward search. That means extending (forward) or shrinking (backward) a feature set. Depending on the given method different approaches are taken.
sfs Sequential Forward Search: Starting from an empty model, in each step the feature increasing the performance measure the most is added to the model.
sbs Sequential Backward Search: Starting from a model with all features, in each step the feature decreasing the performance measure the least is removed from the model.
sffs Sequential Floating Forward Search: Starting from an empty model, in each step the algorithm chooses the best model from all models with one additional feature and from all models with one feature less.
sfbs Sequential Floating Backward Search: Similar to sffs but starting with a full model.

FeatSelControlGA

Search via genetic algorithm. The GA is a simple (mu, lambda) or (mu + lambda) algorithm, depending on the comma setting. A comma strategy selects a new population of size mu out of the lambda > mu offspring. A plus strategy uses the joint pool of mu parents and lambda offspring for selecting mu new candidates. Out of those mu features, the new lambda features are generated by randomly choosing pairs of parents. These are crossed over and crossover.rate represents the probability of choosing a feature from the first parent instead of the second parent. The resulting offspring is mutated, i.e., its bits are flipped with probability mutation.rate. If max.features is set, offspring are repeatedly generated until the setting is satisfied.

Usage

makeFeatSelControlExhaustive(same.resampling.instance = TRUE,
  impute.val = Inf, maxit = NA_integer_, max.features = NA_integer_)

makeFeatSelControlGA(same.resampling.instance = TRUE, impute.val = Inf,
  maxit = NA_integer_, max.features = NA_integer_, comma = FALSE,
  mu = 10L, lambda, crossover.rate = 0.5, mutation.rate = 0.05)

makeFeatSelControlRandom(same.resampling.instance = TRUE, impute.val = Inf,
  maxit = 100L, max.features = NA_integer_, prob = 0.5)

makeFeatSelControlSequential(same.resampling.instance = TRUE,
  impute.val = Inf, method, alpha = 0.01, beta = -0.001,
  maxit = NA_integer_, max.features = NA_integer_)

Arguments

same.resampling.instance
[logical(1)]
Should the same resampling instance be used for all evaluations to reduce variance? Default is TRUE.

impute.val
[numeric(1)]
If something goes wrong during optimization (e.g, the learner crashes), this value is fed back to the tuner, so the tuning algorithm does not abort. It is not stored in the optimization path, an NA and a corresponding error message are logged instead. Note that this value is later multiplied by -1 for maximization measures internally, so you need to enter a larger positive value for maximization here as well. Default is the worst obtainable value of the performance measure you optimize for when you aggregate by mean value, or Inf instead.

maxit
[integer(1)]
Maximal number of iterations. Note, that this is usually not equal to the number of function evaluations.

max.features
[integer(1)]
Maximal number of features.

prob
[numeric(1)]
Parameter of the random feature selection. Probability of choosing a feature.

method
[character(1)]
Parameter of the sequential feature selection. A character representing the method. Possible values are sfs (forward search), sbs (backward search), sffs (floating forward search) and sfbs (floating backward search).

alpha
[numeric(1)]
Parameter of the sequential feature selection. Minimal required value of improvement difference for a forward / adding step. Default is 0.01.

beta
[numeric(1)]
Parameter of the sequential feature selection. Minimal required value of improvement difference for a backward / removing step. Negative values imply that you allow a slight decrease for the removal of a feature. Default is -0.001.

mu
[integer(1)]
Parameter of the GA feature selection. Size of the parent population.

lambda
[integer(1)]
Parameter of the GA feature selection. Size of the children population (should be smaller or equal to mu).

crossover.rate
[numeric(1)]
Parameter of the GA feature selection. Probability of choosing a bit from the first parent within the crossover mutation.

mutation.rate
[numeric(1)]
Parameter of the GA feature selection. Probability of flipping a feature bit, i.e. switch between selecting / deselecting a feature.

comma
[logical(1)]
Parameter of the GA feature selection, indicating whether to use a (mu, lambda) or (mu + lambda) GA. The default is FALSE.

Value

[FeatSelControl]. The specific subclass is one of FeatSelControlExhaustive, FeatSelControlRandom, FeatSelControlSequential, FeatSelControlGA.

See Also

Other featsel: analyzeFeatSelResult; getFeatSelResult; makeFeatSelWrapper; selectFeatures

[Package mlr version 2.0 Index]