makeFeatSelControlExhaustive {mlr}	R Documentation

Create control structures for feature selection.

Description

The following methods are available:

FeatSelControlExhaustive

Exhaustive search. All feature sets (up to a certain number of features max.features) are searched.

FeatSelControlRandom

Random search. Features vectores are randomly drawn, up to a certain number of features max.features. A feature is included in the current set with probability prob. So we are basically drawing (0,1)-membership-vectors, where each element is Bernoulli(prob) distributed.

FeatSelControlSequential

Deterministic forward or backward search.

FeatSelControlGA

Search via genetic algorithm. The GA is a simple (mu, lambda) or (mu + lambda) algorithm, depending on the comma setting. A comma strategy selects a new population of size mu out of the lambda > mu offspring. A plus strategy uses the joint pool of mu parents and lambda offspring for selecting mu new candidates. Out of those mu features, the new lambda features are generated by randomly choosing pairs of parents. These are crossed over and crossover.rate represents the probability of choosing a feature from the first parent instead of the second parent. The resulting offspring is mutated, i.e., its bits are flipped with probability mutation.rate. If max.features is set, offspring are repeatedly generated until the setting is satisfied.

Usage

  makeFeatSelControlExhaustive(same.resampling.instance = TRUE,
    maxit = as.integer(NA), max.features = as.integer(NA))

  makeFeatSelControlGA(same.resampling.instance = TRUE,
    maxit = as.integer(NA), max.features = as.integer(NA),
    comma = FALSE, mu = 10, lambda, crossover.rate = 0.5,
    mutation.rate = 0.05)

  makeFeatSelControlRandom(same.resampling.instance = TRUE,
    maxit = 100L, max.features = as.integer(NA),
    prob = 0.5)

  makeFeatSelControlSequential(same.resampling.instance = TRUE,
    method, alpha = 0.01, beta = 0.01,
    maxit = as.integer(NA), max.features = as.integer(NA))

Arguments

same.resampling.instance
[logical(1)]
Should the same resampling instance be used for all evaluations to reduce variance? Default is TRUE.

maxit
[integer(1)]
Maximal number of iterations. Note, that this is usually not equal to the number of function evaluations.

max.features
[integer(1)]
Maximal number of features.

crossover.rate
[numeric(1)]
Parameter of the GA feature selection. Probability of choosing a bit from the first parent within the crossover mutation.

mutation.rate
[numeric(1)]
Parameter of the GA feature selection. Probability of flipping a feature bit, i.e. switch between selecting / deselecting a feature.

comma
[logical(1)]
Parameter of the GA feature selection, indicating whether to use a (mu, lambda) or (mu + lambda) GA. The default is FALSE.

mu
[integer(1)]
Parameter of the GA feature selection. Size of the parent population.

lambda
[integer(1)]
Parameter of the GA feature selection. Size of the children population (should be smaller or equal to mu).

prob
[numeric(1)]
Parameter of the random feature selection. Probability of choosing a feature.

method
[character(1)]
Parameter of the sequential feature selection. A character representing the method. Possible values are sfs (forward search), sbs (backward search), sffs (floating forward search) and sfbs (floating backward search).

alpha
[numeric(1)]
Parameter of the sequential feature selection. Minimal value of improvement.

beta
[numeric(1)]
Parameter of the sequential feature selection. Maximal value of setback.

Value

[FeatSelControl]. The specific subclass is one of FeatSelControlExhaustive, FeatSelControlRandom, FeatSelControlSequential, FeatSelControlGA.

[Package mlr version 1.2 Index]