LearnerProperties {mlr}	R Documentation

Set, add, remove or query properties of learners

Description

Set, add, remove or query properties of learners

Usage

setProperties(learner, props)

addProperties(learner, props)

removeProperties(learner, props)

hasProperties(learner, props)

Arguments

learner
[Learner | character(1)]
The learner. If you pass a string the learner will be created via makeLearner.

props
[character]
Vector of properties to set, add, remove or query.

Value

setProperties, addProperties and removeProperties return an updated Learner. hasProperties returns a logical vector of the same length of props.

[Package mlr version 1.2 Index]