RLearner {mlr}	R Documentation

Internal construction / wrapping of learner object.

Description

Wraps an already implemented learning method from R to make it accessible to mlr. Call this method in your constructor. You have to pass an id (name), the required package(s), a description object for all changeable parameters (you dont have to do this for the learner to work, but it is strongly recommended), and define what the learner can / cannot do.

Usage

  makeRLearner()

  makeRLearnerClassif(cl, package, par.set,
    numerics = FALSE, factors = FALSE, missings = FALSE,
    weights = FALSE, oneclass = FALSE, twoclass = FALSE,
    multiclass = FALSE, prob = FALSE, par.vals = list())

  makeRLearnerRegr(cl, package, par.set, numerics,
    factors = FALSE, missings = FALSE, weights = FALSE,
    se = FALSE, par.vals = list())

Arguments

cl
[character(1)]
Class name for learner to create. By convention, all classification learners start with “classif.” and all regression learners with “regr.”

package
[character]
Package(s) to load for the implementation of the learner.

par.set
[ParamSet]
Parameter set of (hyper)parameters and their constraints.

numerics
[logical(1)]
Can numeric features be handled? Default is FALSE.

factors
[logical(1)]
Can factor features be handled? Default is FALSE.

missings
[logical(1)]
Can missing values be handled? Default is FALSE.

weights
[logical(1)]
Can case weights be handled? Default is FALSE.

oneclass
[logical(1)]
Can one-class problems be handled? Default is FALSE.

twoclass
[logical(1)]
Can two-class problems be handled? Default is FALSE.

multiclass
[logical(1)]
Can multi-class problems be handled? Default is FALSE.

prob
[logical(1)]
Can probabilities be predicted? Default is FALSE.

se
[logical(1)]
Can standard errors be predicted?? Default is FALSE.

par.vals
[list]
Always set hyperparameters to these values when the object is constructed. Useful when default values are missing in the underlying function. The values can later be overwritten when the user sets hyperparameters. Default is empty list.

Value

[RLearnerClassif or RLearnerRegr].

[Package mlr version 1.2 Index]