RLearner {mlr}	R Documentation

Internal construction / wrapping of learner object.

Description

Wraps an already implemented learning method from R to make it accessible to mlr. Call this method in your constructor. You have to pass an id (name), the required package(s), a description object for all changeable parameters (you dont have to do this for the learner to work, but it is strongly recommended), and use property tags to define features of the learner.

Usage

makeRLearner()

makeRLearnerClassif(cl, package, par.set, par.vals = list(),
  properties = character(0L))

makeRLearnerRegr(cl, package, par.set, par.vals = list(),
  properties = character(0L))

makeRLearnerSurv(cl, package, par.set, par.vals = list(),
  properties = character(0L))

makeRLearnerCluster(cl, package, par.set, par.vals = list(),
  properties = character(0L))

Arguments

cl
[character(1)]
Class name for learner to create. By convention, all classification learners start with “classif.”, all regression learners with “regr.” and all survival learners start with “surv.”.

package
[character]
Package(s) to load for the implementation of the learner.

properties
[character(1)]
Set of learner properties. Some standard property names include:

numerics

Can numeric features be handled?

factors

Can factor features be handled?

missings

Can missing features be handled?

oneclas,twoclass,multiclass

Can one-class, two-class or multi-class classification problems be handled?

prob

Can probabilites be predicted?

se

Can standard errors be predicted?

Default is character(0).

par.set
[ParamSet]
Parameter set of (hyper)parameters and their constraints.

par.vals
[list]
Always set hyperparameters to these values when the object is constructed. Useful when default values are missing in the underlying function. The values can later be overwritten when the user sets hyperparameters. Default is empty list.

Value

[RLearnerClassif, RLearnerRegr or RLearnerSurv].

[Package mlr version 2.1 Index]