ResamplePrediction {mlr}	R Documentation

Prediction from resampling.

Description

Contains predictions from resampling, returned (among other stuff) by function resample. Can basically be used in the same way as Prediction, its super class. The main differences are: (a) The internal data.frame (member data) contains an additional column id, specifying the iteration of the resampling strategy, and and additional columns set, specifying whether the prediction was from an observation in the “train” or “test” set. (b) The prediction time is a numeric vector, its length equals the number of iterations.

See Also

Other resample: ResampleDesc, makeResampleDesc; ResampleInstance, makeResampleInstance; bootstrapB632, bootstrapB632plus, bootstrapOOB, crossval, holdout, repcv, resample, subsample

[Package mlr version 2.1 Index]