makeClassifTask {mlr}	R Documentation

Create a classification, regression, survival, cluster, or cost-sensitive classification task.

Description

The task encapsulates the data and specifies - through its subclasses - the type of the task. It also contains a description object detailing further aspects of the data.

Useful operators are: getTaskFormula, getTaskFormulaAsString, getTaskFeatureNames, getTaskData, getTaskTargets, subsetTask.

Object members:

env [environment]

Environment where data for the task are stored. Use getTaskData in order to access it.

weights [numeric]

See argument above. NULL if not present.

blocking [factor]

See argument above. NULL if not present.

task.desc [TaskDesc]

Encapsulates further information about the task.

Usage

makeClassifTask(id, data, target, weights = NULL, blocking = NULL, positive,
  fixup.data = "warn", check.data = TRUE)

makeClusterTask(id, data, weights = NULL, blocking = NULL,
  fixup.data = "warn", check.data = TRUE)

makeCostSensTask(id, data, costs, blocking = NULL, fixup.data = "warn",
  check.data = TRUE)

makeRegrTask(id, data, target, weights = NULL, blocking = NULL,
  fixup.data = "warn", check.data = TRUE)

makeSurvTask(id, data, target, surv.type = "right", weights = NULL,
  blocking = NULL, fixup.data = "warn", check.data = TRUE)

Arguments

surv.type
[character(1)]
Survival type. Allowed are “right” (default), “left” and “interval2”. See Surv for details.

id
[character(1)]
Id string for object. Default is the name of R variable passed to data.

data
[data.frame]
A data frame containing the features and target variable(s).

target
[character(1) | character(2)]
Name of the target variable. For survival analysis these are the names of the survival time and event columns, so it has length 2.

costs
[data.frame]
A numeric matrix or data frame containing the costs of misclassification. We assume the general case of observation specific costs. This means we have n rows, corresponding to the observations, in the same order as data. The columns correspond to classes and their names are the class labels (if unnamed we use y1 to yk as labels). Each entry (i,j) of the matrix specifies the cost of predicting class j for observation i.

weights
[numeric]
Optional, non-negative case weight vector to be used during fitting. Cannot be set for cost-sensitive learning. Default is NULL which means no (= equal) weights.

blocking
[factor]
An optional factor of the same length as the number of observations. Observations with the same blocking level “belong together”. Specifically, they are either put all in the training or the test set during a resampling iteration. Default is NULL which means no blocking.

positive
[character(1)]
Positive class for binary classification. Default is the first factor level of the target attribute.

fixup.data
[character(1)]
Should some basic cleaning up of data be performed? Currently this means removing empty factor levels for the columns. Possible coices are: “no” = Don't do it. “warn” = Do it but warn about it. “quiet” = Do it but keep silent. Default is “warn”.

check.data
[logical(1)]
Should sanity of data be checked initially at task creation? You should have good reasons to turn this off (one might be speed). Default is TRUE

Value

[Task].

See Also

Other costsens: CostSensClassifModel, CostSensClassifWrapper, makeCostSensClassifWrapper; CostSensRegrModel, CostSensRegrWrapper, makeCostSensRegrWrapper; CostSensWeightedPairsModel, CostSensWeightedPairsWrapper, makeCostSensWeightedPairsWrapper

Examples

library(mlbench)
data(BostonHousing)
data(Ionosphere)

makeClassifTask(data = iris, target = "Species")

## Supervised task: iris
## Type: classif
## Target: Species
## Observations: 150
## Features:
## numerics  factors 
##        4        0 
## Missings: FALSE
## Has weights: FALSE
## Has blocking: FALSE
## Classes: 3
##     setosa versicolor  virginica 
##         50         50         50 
## Positive class: NA

makeRegrTask(data = BostonHousing, target = "medv")

## Supervised task: BostonHousing
## Type: regr
## Target: medv
## Observations: 506
## Features:
## numerics  factors 
##       12        1 
## Missings: FALSE
## Has weights: FALSE
## Has blocking: FALSE

# an example of a classification task with more than those standard arguments:
blocking = factor(c(rep(1, 51), rep(2, 300)))
makeClassifTask(id = "myIonosphere", data = Ionosphere, target = "Class",
  positive = "good", blocking = blocking)

## Supervised task: myIonosphere
## Type: classif
## Target: Class
## Observations: 351
## Features:
## numerics  factors 
##       32        2 
## Missings: FALSE
## Has weights: FALSE
## Has blocking: TRUE
## Classes: 2
##  bad good 
##  126  225 
## Positive class: good

makeClusterTask(data = iris[, -5L])

## Unsupervised task: iris[, -5L]
## Type: cluster
## Observations: 150
## Features:
## numerics  factors 
##        4        0 
## Missings: FALSE
## Has weights: FALSE
## Has blocking: FALSE

[Package mlr version 2.1 Index]