TaskDesc {mlr}	R Documentation

Description object for task.

Description

Description object for task, encapsulates basic properties of the task without having to store the complete data set.

Details

Object members:

id [character(1)]

Id string of task.

type [character(1)]

Type of task, “classif” for classification, “regr” for regression, “surv” for survival, “costsens” for cost-sensitive classification.

target [character(0) | character(1) | character(2)]

Name of target variable. For “surv” these are the names of the survival time and event columns, so it has length 2. For “costsens” ist has length 0, as there is no target column, but a cost matrix instead.

size [integer(1)]

Number of cases in data set.

n.feat [integer(2)]

Number of features, named vector with entries: “numerics”, “factors”.

has.missings [logical(1)]

Are missing values present?

has.weights [logical(1)]

Are weights specified for each observation?

has.blocking [logical(1)]

Is a blocking factor for cases available in the task?

class.levels [character]

All possible classes. Only present for “classif” and “costsens”.

positive [character(1)]

Positive class label for binary classification. Only present for “classif”, NA for multiclass.

negative [character(1)]

Negative class label for binary classification. Only present for “classif”, NA for multiclass.

[Package mlr version 1.2 Index]