makeTuneControlCMAES {mlr}	R Documentation

Create control structures for tuning.

Description

The following tuners are available:

makeTuneControlGrid

Grid search. All kinds of parameter types can be handled. You can either use their correct param type and resolution, or discretize them yourself by always using makeDiscreteParam in the par.set passed to tuneParams.

makeTuneControlRandom

Random search. All kinds of parameter types can be handled.

makeTuneControlCMAES

CMA Evolution Strategy with method cma_es. Can handle numeric(vector) and integer(vector) hyperparameters, but no dependencies. For integers the internally proposed numeric values are automatically rounded. The sigma variance parameter is initialized to 1/4 of the span of box-constraints per parameter dimension.

makeTuneControlGenSA

Generalized simulated annealing with method GenSA. Can handle numeric(vector) and integer(vector) hyperparameters, but no dependencies. For integers the internally proposed numeric values are automatically rounded.

makeTuneControlIrace

Tuning with iterated F-Racing with method irace. All kinds of parameter types can be handled. We return the best of the final elite candidates found by irace in the last race. Its estimated performance is the mean of all evaluations ever done for that candidate.

Some notes on irace: For resampling you have to pass a ResampleDesc, not a ResampleInstance. The resampling strategy is randomly instantiated n.instances times and these are the instances in the sense of irace (instances element of tunerConfig in irace). Also note that irace will always store its tuning results in a file on disk, see the package documentation for details on this and how to change the file path.

Usage

makeTuneControlCMAES(same.resampling.instance = TRUE, impute.val = NULL,
  start = NULL, ...)

makeTuneControlGenSA(same.resampling.instance = TRUE, impute.val = NULL,
  start = NULL, ...)

makeTuneControlGrid(same.resampling.instance = TRUE, impute.val = NULL,
  resolution = 10L)

makeTuneControlIrace(impute.val = NULL, n.instances = 100L,
  show.irace.output = FALSE, ...)

makeTuneControlRandom(same.resampling.instance = TRUE, maxit = 100L)

Arguments

resolution
[integer]
Resolution of the grid for each numeric/integer parameter in par.set. For vector parameters, it is the resolution per dimension. Either pass one resolution for all parameters, or a named vector. See generateGridDesign. Default is 10.

n.instances
[integer(1)]
Number of random resampling instances for irace, see details. Default is 100.

show.irace.output
[logical(1)]
Show console output of irace while tuning? Default is FALSE.

same.resampling.instance
[logical(1)]
Should the same resampling instance be used for all evaluations to reduce variance? Default is TRUE.

impute.val
[numeric]
If something goes wrong during optimization (e.g, the learner crashes), this value is fed back to the tuner, so the tuning algorithm does not abort. It is not stored in the optimization path, an NA and a corresponding error message are logged instead. Note that this value is later multiplied by -1 for maximization measures internally, so you need to enter a larger positive value for maximization here as well. Default is the worst obtainable value of the performance measure you optimize for when you aggregate by mean value, or Inf instead. For multi-criteria optimization pass a vector of imputation values, one for each of your measures, in the same order as your measures.

start
[numeric]
Named list of initial parameter values.

...
[any]
Further control parameters passed to the control argument of cma_es and tunerConfig argument of irace.

maxit
[integer(1)]
Number of iterations for random search. Default is 100.

Value

[TuneControl]. The specific subclass is one of TuneControlGrid, TuneControlRandom, TuneControlCMAES, TuneControlGenSA, TuneControlIrace.

See Also

Other tune: ModelMultiplexer, makeModelMultiplexer; getTuneResult; makeModelMultiplexerParamSet; makeTuneWrapper; tuneParams; tuneThreshold

[Package mlr version 2.1 Index]