TuneResult {mlr}	R Documentation

Result of tuning.

Description

Container for results of hyperparameter tuning. Contains the obtained point in search space, its performance values and the optimization path which lead there.

Object members:

learner [Learner]

Learner that was optimized.

control [TuneControl]

Control object from tuning.

x [list]

Named list of hyperparameter values identified as optimal.

y [numeric]

Performance values for optimal x.

opt.path [OptPath]

Optimization path which lead to x.

[Package mlr version 2.0 Index]