Aggregation methods.

test.mean Mean of performance values on test sets.

test.sd Standard deviation of performance values on test sets.

test.median Median of performance values on test sets.

test.min Minimum of performance values on test sets.

test.max Maximum of performance values on test sets.

test.sum Sum of performance values on test sets.

train.mean Mean of performance values on training sets.

train.sd Standard deviation of performance values on training sets.

train.median Median of performance values on training sets.

train.min Minimum of performance values on training sets.

train.max Maximum of performance values on training sets.

train.sum Sum of performance values on training sets.

b632 Aggregation for B632 bootstrap.

b632plus Aggregation for B632+ bootstrap.

testgroup.mean Performance values on test sets are grouped according to resampling method. The mean for every group is calculated, then the mean of those means. Mainly used for repeated CV.

testgroup.sd Similar to testgroup.mean - after the mean for every group is calculated, the standard deviation of those means is obtained. Mainly used for repeated CV.

test.join Performance measure on joined test sets. This is especially useful for small sample sizes where unbalanced group sizes have a significant impact on the aggregation, especially for cross-validation test.join might make sense now. For the repeated CV, the performance is calculated on each repetition and then aggregated with the arithmetic mean.

test.mean

test.sd

test.median

test.min

test.max

test.sum

test.range

test.rmse

train.mean

train.sd

train.median

train.min

train.max

train.sum

train.range

train.rmse

b632

b632plus

testgroup.mean

testgroup.sd

test.join

Format

None

See also

Aggregation