aggregations {mlr}	R Documentation

Aggregation methods.

Description

test.mean
Mean of performance values on test sets.

test.sd
Standard deviation of performance values on test sets.

test.median
Median of performance values on test sets.

test.min
Minimum of performance values on test sets.

test.max
Maximum of performance values on test sets.

test.sum
Sum of performance values on test sets.

train.mean
Mean of performance values on training sets.

train.sd
Standard deviation of performance values on training sets.

train.median
Median of performance values on training sets.

train.min
Minimum of performance values on training sets.

train.max
Maximum of performance values on training sets.

train.sum
Sum of performance values on training sets.

b632
Aggregation for B632 bootstrap.

b632plus
Aggregation for B632+ bootstrap.

testgroup.mean
Performance values on test sets are grouped according to resampling method. The mean for very group is calculated, then the mean of those means. Mainly used for repeated CV.

Usage

  test.mean

  test.sd

  test.median

  test.min

  test.max

  test.sum

  test.sqrt.of.mean

  train.mean

  train.sd

  train.median

  train.min

  train.max

  train.sum

  b632

  b632plus

  testgroup.mean

Format

List of 2 $ id : chr "test.mean" $ fun:function (task, perf.test, perf.train, measure, group, pred) - attr(*, "class")= chr "Aggregation"

See Also

Aggregation

[Package mlr version 1.2 Index]