analyzeFeatSelResult {mlr}	R Documentation

Shows the detailed steps of the feature selection.

Description

This function prints the steps selectFeatures took to find it's optimal set of features and the reason why it stopped. It can also give information about all calculations done on each intermediate step.

Usage

analyzeFeatSelResult(fs.obj, reduce = TRUE)

Arguments

fs.obj
[FeatSelResult(1)]
The result of one feature selection generated by selectFeatures.

reduce
[logical(1)]
TRUE if the results of all calculations for each step should be printed.
FALSE only the calculation leading to the improving feature set is printed.
Default is FALSE.

Value

Nothing.

[Package mlr version 1.2 Index]