configureMlr {mlr}	R Documentation

Configures the behavior of the package.

Description

Configuration is done by setting custom options.

If you do not set an option here, its current value will be kept.

If you call this function with an empty argument list, everything is set to its defaults.

Usage

configureMlr(show.info, on.learner.error, on.learner.warning,
  on.par.without.desc, show.learner.output)

Arguments

show.info
[logical(1)]
Some methods of mlr support a show.info argument to enable verbose output on the console. This option sets the default value for these arguments. Setting the argument manually in one of these functions will overwrite the default value for that specific function call. Default is TRUE.

on.learner.error
[character(1)]
What should happen if an error in an underlying learning algorithm is caught:
“stop”: R exception is generated.
“warn”: A FailureModel will be created, which predicts only NAs and a warning will be generated.
“quiet”: Same as “warn” but without the warning.
Default is “stop”.

on.learner.warning
[character(1)]
What should happen if a warning in an underlying learning algorithm is generated:
“warn”: The warning is generated as usual.
“quiet”: The warning is suppressed.
Default is “warn”.

on.par.without.desc
[character(1)]
What should happen if a parameter of a learner is set to a value, but no parameter description object exists, indicating a possibly wrong name:
“stop”: R exception is generated.
“warn”: Warning, but parameter is still passed along to learner.
“quiet”: Same as “warn” but without the warning.
Default is “stop”.

show.learner.output
[logical(1)]
Should the output of the learning algorithm during training and prediction be shown or captured and suppressed? Default is TRUE.

Value

[invisible(NULL)].

See Also

Other configure: getMlrOptions

[Package mlr version 2.0 Index]