filterFeatures {mlr}	R Documentation

Filter features by using a numerical importance criterion. Calculates numerical importance values for all features. Thresholding of these values can be used to select “useful” features. Look at package FSelector for details on the filter algorithms.

Description

Filter features by using a numerical importance criterion. Calculates numerical importance values for all features. Thresholding of these values can be used to select “useful” features. Look at package FSelector for details on the filter algorithms.

Usage

  filterFeatures(task, method = "random.forest.importance")

Arguments

task
[SupervisedTask]
The task.

method
[character(1)]
Filter method. Available are: “linear.correlation”, “rank.correlation”, “information.gain”, “gain.ratio”, “symmetrical.uncertainty”, “chi.squared”, “random.forest.importance”, “relief”, “oneR” Default is “random.forest.importance”.

Value

[numeric]. A named numeric vector that contains an importance value for each feature.

[Package mlr version 1.2 Index]