Returns the selected feature set and optimization path after training.
Source: R/getFeatSelResult.R

Returns the selected feature set and optimization path after training.

getFeatSelResult(object)

Arguments

object
(WrappedModel)
Trained Model created with makeFeatSelWrapper.

Value

(FeatSelResult).

See also