getParamSet {mlr}	R Documentation

Get a description of all possible parameter settings for a learner.

Description

Get a description of all possible parameter settings for a learner.

Usage

getParamSet(learner)

Arguments

learner
[Learner | character(1)]
The learner. If you pass a string the learner will be created via makeLearner.

Value

[ParamSet].

See Also

Other learner: getHyperPars; setHyperPars; setId; setPredictType; showHyperPars

[Package mlr version 2.0 Index]