Get the resampling indices from a tuning or feature selection wrapper..

After you resampled a tuning or feature selection wrapper (see makeTuneWrapper) with resample(..., extract = getTuneResult) or resample(..., extract = getFeatSelResult) this helper returns a list with the resampling indices used for the respective method.

getResamplingIndices(object, inner = FALSE)

Arguments

object

(ResampleResult)
The result of resampling of a tuning or feature selection wrapper.

inner

(logical)
If TRUE, returns the inner indices of a nested resampling setting.

Value

(list). One list for each outer resampling fold.

See also

Examples

task = makeClassifTask(data = iris, target = "Species")
lrn = makeLearner("classif.rpart")
# stupid mini grid
ps = makeParamSet(
  makeDiscreteParam("cp", values = c(0.05, 0.1)),
  makeDiscreteParam("minsplit", values = c(10, 20))
)
ctrl = makeTuneControlGrid()
inner = makeResampleDesc("Holdout")
outer = makeResampleDesc("CV", iters = 2)
lrn = makeTuneWrapper(lrn, resampling = inner, par.set = ps, control = ctrl)
# nested resampling for evaluation
# we also extract tuned hyper pars in each iteration and by that the resampling indices
r = resample(lrn, task, outer, extract = getTuneResult)
#> Resampling: cross-validation
#> Measures:             mmce      
#> [Tune] Started tuning learner classif.rpart for parameter set:
#>              Type len Def   Constr Req Tunable Trafo
#> cp       discrete   -   - 0.05,0.1   -    TRUE     -
#> minsplit discrete   -   -    10,20   -    TRUE     -
#> With control class: TuneControlGrid
#> Imputation value: 1
#> [Tune-x] 1: cp=0.05; minsplit=10
#> [Tune-y] 1: mmce.test.mean=0.1600000; time: 0.0 min
#> [Tune-x] 2: cp=0.1; minsplit=10
#> [Tune-y] 2: mmce.test.mean=0.1600000; time: 0.0 min
#> [Tune-x] 3: cp=0.05; minsplit=20
#> [Tune-y] 3: mmce.test.mean=0.1600000; time: 0.0 min
#> [Tune-x] 4: cp=0.1; minsplit=20
#> [Tune-y] 4: mmce.test.mean=0.1600000; time: 0.0 min
#> [Tune] Result: cp=0.1; minsplit=10 : mmce.test.mean=0.1600000
#> [Resample] iter 1:    0.0400000 
#> [Tune] Started tuning learner classif.rpart for parameter set:
#>              Type len Def   Constr Req Tunable Trafo
#> cp       discrete   -   - 0.05,0.1   -    TRUE     -
#> minsplit discrete   -   -    10,20   -    TRUE     -
#> With control class: TuneControlGrid
#> Imputation value: 1
#> [Tune-x] 1: cp=0.05; minsplit=10
#> [Tune-y] 1: mmce.test.mean=0.0800000; time: 0.0 min
#> [Tune-x] 2: cp=0.1; minsplit=10
#> [Tune-y] 2: mmce.test.mean=0.0800000; time: 0.0 min
#> [Tune-x] 3: cp=0.05; minsplit=20
#> [Tune-y] 3: mmce.test.mean=0.0800000; time: 0.0 min
#> [Tune-x] 4: cp=0.1; minsplit=20
#> [Tune-y] 4: mmce.test.mean=0.0800000; time: 0.0 min
#> [Tune] Result: cp=0.1; minsplit=20 : mmce.test.mean=0.0800000
#> [Resample] iter 2:    0.0800000 
#> 
#> Aggregated Result: mmce.test.mean=0.0600000
#> 
# get tuning indices
getResamplingIndices(r, inner = TRUE)
#> [[1]]
#> [[1]]$train.inds
#> [[1]]$train.inds[[1]]
#>  [1]  96  11 132  58  16 144  27  30 108 114 145   7 129  45  91  18 127 103  92
#> [20]  48 117  25  46  56  98 111  14   3 143  59  37 110  21  94 119 122 105 101
#> [39]  51  50  57 113  74  44  41  86  19  23 118 149
#> 
#> 
#> [[1]]$test.inds
#> [[1]]$test.inds[[1]]
#>  [1] 130 100  90 138  39  84 124  93  78  28   2  71  35  31 106  20  88 134  85
#> [20]  22  36  38  53 125  80
#> 
#> 
#> 
#> [[2]]
#> [[2]]$train.inds
#> [[2]]$train.inds[[1]]
#>  [1]  72 142  99   6 148  42  33  65  15  79  82 120  87  68 109  76   1  69  32
#> [20]  26  66 131 146 135   5  62  89 137 116   4 133  13 123  10 140  70 121 136
#> [39]  60   9 141  77  43  24  64 107  29  61  73  17
#> 
#> 
#> [[2]]$test.inds
#> [[2]]$test.inds[[1]]
#>  [1]  95 112 147 102  47  83 139  55 128 115   8  34 126  97 104  40  54  52 150
#> [20]  75  12  63  67  49  81
#> 
#> 
#>