Get the resampling indices from a tuning or feature selection wrapper..

After you resampled a tuning or feature selection wrapper (see makeTuneWrapper) with resample(..., extract = getTuneResult) or resample(..., extract = getFeatSelResult) this helper returns a list with the resampling indices used for the respective method.

Usage,

getResamplingIndices(object, inner = FALSE)

Arguments

object

(ResampleResult)
The result of resampling of a tuning or feature selection wrapper.

inner

(logical)
If TRUE, returns the inner indices of a nested resampling setting.

Value

(list). One list for each outer resampling fold.

See also

Examples

task = makeClassifTask(data = iris, target = "Species")
lrn = makeLearner("classif.rpart")
# stupid mini grid
ps = makeParamSet(
  makeDiscreteParam("cp", values = c(0.05, 0.1)),
  makeDiscreteParam("minsplit", values = c(10, 20))
)
ctrl = makeTuneControlGrid()
inner = makeResampleDesc("Holdout")
outer = makeResampleDesc("CV", iters = 2)
lrn = makeTuneWrapper(lrn, resampling = inner, par.set = ps, control = ctrl)
# nested resampling for evaluation
# we also extract tuned hyper pars in each iteration and by that the resampling indices
r = resample(lrn, task, outer, extract = getTuneResult)
#> Resampling: cross-validation
#> Measures:             mmce      
#> [Tune] Started tuning learner classif.rpart for parameter set:
#>              Type len Def   Constr Req Tunable Trafo
#> cp       discrete   -   - 0.05,0.1   -    TRUE     -
#> minsplit discrete   -   -    10,20   -    TRUE     -
#> With control class: TuneControlGrid
#> Imputation value: 1
#> [Tune-x] 1: cp=0.05; minsplit=10
#> [Tune-y] 1: mmce.test.mean=0.0800000; time: 0.0 min
#> [Tune-x] 2: cp=0.1; minsplit=10
#> [Tune-y] 2: mmce.test.mean=0.0800000; time: 0.0 min
#> [Tune-x] 3: cp=0.05; minsplit=20
#> [Tune-y] 3: mmce.test.mean=0.0800000; time: 0.0 min
#> [Tune-x] 4: cp=0.1; minsplit=20
#> [Tune-y] 4: mmce.test.mean=0.0800000; time: 0.0 min
#> [Tune] Result: cp=0.05; minsplit=20 : mmce.test.mean=0.0800000
#> [Resample] iter 1:    0.0666667 
#> [Tune] Started tuning learner classif.rpart for parameter set:
#>              Type len Def   Constr Req Tunable Trafo
#> cp       discrete   -   - 0.05,0.1   -    TRUE     -
#> minsplit discrete   -   -    10,20   -    TRUE     -
#> With control class: TuneControlGrid
#> Imputation value: 1
#> [Tune-x] 1: cp=0.05; minsplit=10
#> [Tune-y] 1: mmce.test.mean=0.0000000; time: 0.0 min
#> [Tune-x] 2: cp=0.1; minsplit=10
#> [Tune-y] 2: mmce.test.mean=0.0000000; time: 0.0 min
#> [Tune-x] 3: cp=0.05; minsplit=20
#> [Tune-y] 3: mmce.test.mean=0.0000000; time: 0.0 min
#> [Tune-x] 4: cp=0.1; minsplit=20
#> [Tune-y] 4: mmce.test.mean=0.0000000; time: 0.0 min
#> [Tune] Result: cp=0.05; minsplit=20 : mmce.test.mean=0.0000000
#> [Resample] iter 2:    0.0666667 
#> 
#> Aggregated Result: mmce.test.mean=0.0666667
#> 
# get tuning indices
getResamplingIndices(r, inner = TRUE)
#> [[1]]
#> [[1]]$train.inds
#> [[1]]$train.inds[[1]]
#>  [1] 138 113  88  58 117  70  20  49  48 116 133  29  89 120 105  69  85   5 119
#> [20] 112  71  46 136  51 104 110  68 107 100  10  91  54  99  98  95  16  19  22
#> [39] 140   7  72  74  53  47  13  90  59  50  79  11
#> 
#> 
#> [[1]]$test.inds
#> [[1]]$test.inds[[1]]
#>  [1] 131  80 134 144  57 128  94 111  17 118  27 124   8  30  64 135 123 115  82
#> [20]  93  24  86 142  73 148
#> 
#> 
#> 
#> [[2]]
#> [[2]]$train.inds
#> [[2]]$train.inds[[1]]
#>  [1]  96  62  65  81  63  34  78 109  40  61  67  37  45  25  55  23 101  66 132
#> [20] 122  41 139 103 108  56   3  52  42 130  87 145 146   9  21  15 114  32 137
#> [39]  36 125   2  35  84 143  28  12 106 147  83  77
#> 
#> 
#> [[2]]$test.inds
#> [[2]]$test.inds[[1]]
#>  [1]  33  76  60  39  75 149  14   4   6 141  18 129  92  43  38  44  26  97 150
#> [20] 121   1  31 126 127 102
#> 
#> 
#>