getTaskData {mlr}	R Documentation

Extract data in task.

Description

Useful in trainLearner when you add a learning machine to the package.

Usage

getTaskData(task, subset, features, target.extra = FALSE,
  recode.target = "no")

Arguments

task
[Task]
The task.

subset
[integer]
Selected cases. Default is all cases.

features
[character]
Selected features. Default is all.

target.extra
[logical(1)]
Should target vector be returned separately? If not, a single data.frame including the target is returned, otherwise a list with the input data.frame and an extra vector for the targets. Default is FALSE.

recode.target
[character(1)]
Should target classes be recoded? Only for binary classification. Possible are “no” (do nothing), “01”, and “-1+1”. In the two latter cases the target vector is converted into a numeric vector. The positive class is coded as +1 and the negative class either as 0 or -1. Default is “no”.

Value

Either a data.frame or a list with data.frame data and vector target.

See Also

Other task: getTaskCosts; getTaskFeatureNames; getTaskFormula, getTaskFormulaAsString; getTaskNFeats; getTaskTargets; subsetTask

Examples

library("mlbench")
data(BreastCancer)

df = BreastCancer
df$Id = NULL
task = makeClassifTask(id = "BreastCancer", data = df, target = "Class", positive = "malignant")
head(getTaskData)

##                                                                                 
## 1 function (task, subset, features, target.extra = FALSE, recode.target = "no") 
## 2 {                                                                             
## 3     indexHelper = function(df, i, j, drop = TRUE) {                           
## 4         switch(2L * is.null(i) + is.null(j) + 1L, df[i, j, drop = drop],      
## 5             df[i, , drop = drop], df[, j, drop = drop], df)                   
## 6     }

head(getTaskData(task, features = c("Cell.size", "Cell.shape"), recode.target = "-1+1"))

##   Cell.size Cell.shape Class
## 1         1          1    -1
## 2         4          4    -1
## 3         1          1    -1
## 4         8          8    -1
## 5         1          1    -1
## 6        10         10     1

head(getTaskData(task, subset = 1:100, recode.target = "01"))

##   Cl.thickness Cell.size Cell.shape Marg.adhesion Epith.c.size Bare.nuclei
## 1            5         1          1             1            2           1
## 2            5         4          4             5            7          10
## 3            3         1          1             1            2           2
## 4            6         8          8             1            3           4
## 5            4         1          1             3            2           1
## 6            8        10         10             8            7          10
##   Bl.cromatin Normal.nucleoli Mitoses Class
## 1           3               1       1     0
## 2           3               2       1     0
## 3           3               1       1     0
## 4           3               7       1     0
## 5           3               1       1     0
## 6           9               7       1     1

[Package mlr version 2.1 Index]