imputations {mlr}	R Documentation

Built in imputation methods The built-ins are:

imputeConstant(const) for imputation using a constant value,

imputeMedian() for imputation using the median,

imputeMode() for imputation using the mode,

imputeMin(multiplier) for imputation using the minimum,

imputeMax(multiplier) for imputation using the maximum,

imputeNormal(mean, sd) for imputation using normally distributed random values. Mean and standard deviation will be calculated from the data if not provided.

imputeHist(breaks, use.mids) for imputation using random values with probabilities calculated using table or hist.

imputeLearner(learner, preimpute) for imputations using the response of a classification or regression learner.

Description

Built in imputation methods The built-ins are:

imputeConstant(const) for imputation using a constant value,

imputeMedian() for imputation using the median,

imputeMode() for imputation using the mode,

imputeMin(multiplier) for imputation using the minimum,

imputeMax(multiplier) for imputation using the maximum,

imputeNormal(mean, sd) for imputation using normally distributed random values. Mean and standard deviation will be calculated from the data if not provided.

imputeHist(breaks, use.mids) for imputation using random values with probabilities calculated using table or hist.

imputeLearner(learner, preimpute) for imputations using the response of a classification or regression learner.

Usage

imputeConstant(const)

imputeMedian()

imputeMean()

imputeMode()

imputeMin(multiplier = 1)

imputeMax(multiplier = 1)

imputeUniform(min, max)

imputeNormal(mu, sd)

imputeHist(breaks, use.mids = TRUE)

imputeLearner(learner, preimpute = list())

Arguments

const
[any]
Constant valued use for imputation.

multiplier
[numeric(1)]
Value that stored minimum or maximum is multiplied with when imputation is done.

min
[numeric(1)]
Lower bound for uniform distribution. If missing it will be estimated from the data.

max
[numeric(1)]
Upper bound for uniform distribution. If missing it will be estimated from the data.

mu
[numeric(1)]
Mean of normal distribution. If missing it will be estimated from the data.

sd
[numeric(1)]
Standard deviation of normal distribution. If missing it will be estimated from the data.

breaks
[numeric(1)]
Number of breaks to use in hist. If missing, defaults to auto-detection via “Sturges”.

use.mids
[logical(1)]
If x is numeric and a histogram is used, impute with bin mids (default) or instead draw uniformly distributed samples within bin range.

learner
[Learner]
Classification or regression learner. Its predictions will be used for imputations. Note that the target column is not available for this operation.

preimpute
[list]
Arguments for inner call to impute in scenarios where learner itself cannot handle missing values. Default is list().

[Package mlr version 1.2 Index]