impute {mlr}	R Documentation

Impute and re-impute data

Description

Allows imputation of missing feature values through various techniques. Note that you have the possibility to re-impute a data set in the same way as the imputation was performed during training. This especially comes in handy during resampling when one wants to perform the same imputation on the test set as on the training set.

The function impute performs the imputation on a data set and returns, alongside with the imputed data set, an “ImputationDesc” object which can contain “learned” coefficients and helpful data. It can then be passed together with a new data set to reimpute.

The imputation techniques can be specified for certain features or for feature classes, see function arguments.

You can either provide an arbitrary object, use a built-in imputation method listed under imputations or create one yourself using makeImputeMethod.

Usage

impute(data, target, classes = list(), cols = list(),
  dummies = character(0L), impute.new.levels = TRUE,
  recode.factor.levels = TRUE)

Arguments

data
[data.frame]
Input data.

target
[character]
Name of the column specifying the response.

classes
[named list]
Named list containing imputation techniques for classes of columns. E.g. list(numeric = imputeMedian()).

cols
[named list]
Named list containing names of the built-in imputation methods to impute missing values in the data column referenced by the list element's name. Overwrites imputation set via classes.

dummies
[character]
Column names for which dummy variables (binary missing indicator) should be created. Default is character(0).

impute.new.levels
[logical(1)]
If new, unencountered factor level occur during reimputation, should these be handled as NAs and then be imputed the same way? Default is TRUE.

recode.factor.levels
[logical(1)]
Recode factor levels after reimputation, so they match the respective element of lvls (in the description object) and therefore match the levels of the feature factor in the training data after imputation?. Default is TRUE.

Details

The description object contains these slots

target [character]

See argument.

features [character]

Feature names, these are the column names of data, excluding target.

lvls [named list]

Mapping of column names of factor features to their levels, including newly created ones during imputation.

impute [named list]

Mapping of column names to imputation functions.

dummies [named list]

Mapping of column names to imputation functions.

impute.new.levels [logical(1)]

See argument.

recode.factor.levels [logical(1)]

See argument.

Value

[list]

data data.frame
Imputed data.

desc ImputationDesc
Description object.

Examples

df = data.frame(x = c(1, 1, NA), y = factor(c("a", "a", "b")), z=1:3)
imputed = impute(df, target=character(0), cols=list(x = 99, y = imputeMode()))
print(imputed$data)

##    x y z
## 1  1 a 1
## 2  1 a 2
## 3 99 b 3

[Package mlr version 1.2 Index]