impute {mlr}	R Documentation

Impute and re-impute data

Description

These functions provide some convenience for imputation of missing values. A noteworthy feature is given by the possibility to re-impute a data set in the same way as learned previously which especially becomes handy in combination with resampling strategies.

Usage

  impute(data, target, classes = list(), cols = list(),
    dummies = character(0L), impute.newlevels = TRUE)

  imp.const(const)

  imp.median()

  imp.mode()

  imp.min(multiplier = 1)

  imp.max(multiplier = 1)

  imp.normal()

  imp.hist(breaks, use.mids = TRUE)

Arguments

data
[data.frame]
Input data.

target
[character]
Name of the column specifying the response.

classes
[named list]
Named list containing imputation techniques for classes of columns. E.g. list(integer = 0).

cols
[named list]
Named list containing names of the built-in imputation methods to impute missing values in the data column referenced by the list element's name. Overwrites imputation set via classes.

dummies
[character()]
Set of column names for which dummy variables (binary missing indicator) should be created. Default is character(0).

impute.newlevels
[logical(1)]
Impute new levels in factors the same way as missing values? Default is TRUE.

const
[any]
Any object to use for imputation.

multiplier
[numeric(1)]
Numeric value to multiply the minimum or maximum.

breaks
[numeric(1)]
Number of breaks to use in hist. If missing, defaults to auto-detection via “Sturges”.

use.mids
[logical(1)]
If x is numeric and a histogram is used, impute with bin mids (default) or instead draw uniformly distributed samples within bin range.

Details

The function impute performs the imputations on a data set and returns, alongside with the imputed data set, an “ImputationDesc” object which then can be passed together with a new data set to reimpute.

The imputation techniques can be specified for certain features or using the feature classes, see parameters. You can either provide an arbitrary object use a built-in imputation method or create one yourself using ... FIXME we should provide a constructor and document it. The built-ins are:

imp.const(const) for imputation using a constant value,

imp.median() for imputation using the median,

imp.mode() for imputation using the mode,

imp.min(multiplier) for imputation using the minimum,

imp.max(multiplier) for imputation using the maximum,

imp.normal() for imputation using normally distributed random values with mean and variance estimated from the data,

imp.hist(breaks, use.mids) for imputation using random values with probabilities calculated using table or hist.

Value

list with two named elements “data” and “desc”.

Examples

df = data.frame(x = c(1, 1, NA), y = factor(c("a", "a", "b")), z=1:3)
 imputed = impute(df, target=character(0), cols=list(x = 99, y = imp.mode()))
 print(imputed$data)

##    x y z
## 1  1 a 1
## 2  1 a 2
## 3 99 b 3

 reimpute(data.frame(x=NA), imputed$desc)

##    x y  z
## 1 99 a NA

[Package mlr version 1.2 Index]