learn.task-class {mlr}	R Documentation

Base class for learning tasks.

Description

A learning task is a general description object for a machine learning experiment. It wraps the data source and specifies - through its subclasses - the type of the task (e.g. classification or regression), the target variable and other details of the problem. As this is just an abstract base class, you should not instantiate it directly but use the inheriting classes and their factory methods.

Details

learn.task-class: Getter.
Note that all getters of task.desc and data.desc can also be used.

data [data.frame]. Optional parameters: row, col

The data.frame is returned, possibly indexed by row/col. If col is missing, only columns which were not excluded are returned.

input.names [character]

The names of the input variables (without excluded variables).

targets [character]. Optional parameters: row

If row is missing all target values are returned. Otherwise they are indexed by row.

weights [numeric]. Optional parameters: row

If row is missing all case weights are returned. Otherwise they are indexed by row. NULL if no weights were set.

rows.with.missing [integer]

Index vector for rows which contain missing values.

cols.with.missing [integer]

Index vector for columns which contain missing values.

rows.with.inf [integer]

Index vector for rows which contain infinite numerical values.

cols.with.inf [integer]

Index vector for columns which contain infinite numerical values.

Extends

object

Methods

get.learners

resample.update

resample.fit

make.res.instance

make.res.instance

[

train

See Also

make.task

[Package mlr version 0.3.1095 Index]