learner-class {mlr}	R Documentation

Base class for inducers.

Description

Abstract base class for learning algorithms.

Details

learner-class: Getter.

is.classif [boolean]

Is this learner for classification tasks?

is.regr [boolean]

Is this learner for regression tasks?

id [string]

Id string of learner.

label [string]

Label string of learner.

pack [string]

Package were underlying learner is implemented.

props [learner.props]

Properties object to describe functionality of the learner.

hyper.pars [list]

List of fixed hyperparameters and respective values for this learner.

hyper.names [character]

Names of currently used hyperparameters.

hyper.types [character]

For which step in the model building process are the respective hyperparameters used? Named character vector.

supports.numerics [boolean]

Can numeric inputs be processed?

supports.factors [boolean]

Can factor inputs be processed?

supports.characters [boolean]

Can character inputs be processed?

supports.missings [boolean]

Can missing values be processed?

supports.multiclass [boolean]

Can probabilities be predicted?

supports.costs [boolean]

Can misclassification costs be directly used during training?

supports.probs [boolean]

Can probabilities be predicted?

supports.decision [boolean]

Can probabilities be predicted?

supports.weights [boolean]

Can case weights be used?

Extends

object

Methods

[

signature( = "learner"): Getter.

to.string

resample.fit

train

[Package mlr version 0.3.732 Index]