learners {mlr}	R Documentation

Implemented inducers.

Description

List of supported learning algorithms.

Details

classif.kknn
k-Nearest Neighbor from kknn package: kknn

classif.naiveBayes
Naive Bayes from e1071 package: naiveBayes

classif.lda
Linear Discriminant Analysis from MASS package: lda

classif.qda
Quadratic Discriminant Analysis from MASS package: qda

classif.rda
Regularized Discriminant Analysis from klaR package: rda

classif.mda
Mixture Discriminant Analysis from mda package: mda

classif.logreg
Logistic Regression from stats package: lm

classif.rpart
Decision Tree from rpart package: rpart

classif.randomForest
Random Forest from randomForest package: randomForest

classif.ada
Boosting from ada package: ada

classif.adaboost.M1
Boosting from adabag package: adaboost.M1

classif.gbm
Gradient boosting machine from gbm package: gbm

classif.ksvm
Support Vector Machines from kernlab package: ksvm
Note that kernel parameters have to be passed directly and not by using the kpar list in ksvm.

classif.nnet
Neural Network from nnet package: nnet

regr.lm
Simple linear regression from stats package: lm

regr.ridge
Ridge regression from penalized package: penalized

regr.lasso
Lasso regression from penalized package: penalized

regr.kknn
K-Nearest-Neighbor regression from kknn package: kknn

regr.gbm
Gradient boosting machine from gbm package: gbm

regr.blackboost
Gradient boosting with regression trees from mboost package: blackboost

regr.ksvm
Support Vector Machines from kernlab package: ksvm
Note that kernel parameters have to be passed directly and not by using the kpar list in ksvm.

regr.nnet
Neural Network from nnet package: nnet

[Package mlr version 0.3.732 Index]