learners {mlr}	R Documentation

List of supported learning algorithms.

Description

classif.ada
Boosting from ada package: ada

classif.boosting
Boosting from adabag package: boosting
Note that xval has been set to 0 by default for speed.

classif.blackboost
Gradient boosting with regression trees from mboost package: blackboost

classif.ctree
Conditional Inference Trees from party package: ctree

classif.fnn
Fast k-Nearest Neighbor from FNN package: knn

classif.gbm
Gradient boosting machine from gbm package: gbm

classif.geoDA
Geometric Predictive Discriminant Analysis from DiscriMiner package: geoDA

classif.glmboost
Boosting for GLMs from mboost package: glmboost
Note that family has been set to Binomial() by default.

classif.IBk
K-nearest neighbours from RWeka package: IBk

classif.J48
J48 Decision Trees from RWeka package: J48 Note that NAs are directly passed to WEKA with na.action = na.pass.

classif.JRip
Propositional Rule Learner from RWeka package: JRip Note that NAs are directly passed to WEKA with na.action = na.pass.

classif.kknn
k-Nearest Neighbor from kknn package: kknn

classif.knn
k-Nearest Neighbor from class package: knn

classif.ksvm
Support Vector Machines from kernlab package: ksvm
Note that kernel parameters have to be passed directly and not by using the kpar list in ksvm.
Note that fit has been set to FALSE by default for speed.

classif.lda
Linear Discriminant Analysis from MASS package: lda

classif.LiblineaRBinary
Regularized Binary Linear Predictive Models Estimation from LiblineaR package: LiblineaR Note that this model subsumes the types 1,2,3,5

classif.LiblineaRLogReg
Regularized Logistic Regression from LiblineaR package: LiblineaR Note that this model subsumes type 0,6,7.

classif.LiblineaRMultiClass
Multi-class Support Vector Classification by Crammer and Singer from LiblineaR package: LiblineaR Note that this model is type 4.

classif.linDA
Linear Discriminant Analysis from DiscriMiner package: linDA

classif.logreg
Logistic Regression from stats package: glm

classif.lssvm
Least Squares Support Vector Machine from kernlab package: lssvm
Note that fitted has been set to FALSE by default for speed.

classif.lvq1
Learning Vector Quantization from class package: lvq1

classif.mda
Mixture Discriminant Analysis from mda package: mda
Note that keep.fitted has been set to FALSE by default for speed.

classif.multinom
Multinomial Regression from nnet package: multinom

classif.naiveBayes
Naive Bayes from e1071 package: naiveBayes

classif.nnet
Neural Network from nnet package: nnet
Note that size has been set to 3 by default.

classif.OneR
1-R classifier from RWeka package: OneR
Note that NAs are directly passed to WEKA with na.action = na.passi.

classif.PART
PART decision lists from RWeka package: PART
Note that NAs are directly passed to WEKA with na.action = na.pass.

classif.plr
Logistic regression with a L2 penalty from stepPlr package: plr
Note that AIC and BIC penalty types can be selected via the new parameter cp.type.

classif.plsDA
Partial Least Squares (PLS) Discriminant Analysis from DiscriMiner package: plsDA

classif.qda
Quadratic Discriminant Analysis from MASS package: qda

classif.quaDA
Quadratic Discriminant Analysis from DiscriMiner package: quaDA

classif.randomForest
Random Forest from randomForest package: randomForest.
The argument fix.factors restores the factor levels seen in the training data before prediction to circumvent randomForest's internal sanity checks. Default is FALSE.

classif.rda
Regularized Discriminant Analysis from klaR package: rda
Note that estimate.error has been set to FALSE by default for speed.

classif.rpart
Decision Tree from rpart package: rpart
Note that xval has been set to 0 by default for speed.

classif.svm
Support Vector Machines (libsvm) from e1071 package: svm

regr.blackboost
Gradient boosting with regression trees from mboost package: blackboost

regr.crs
Regression Splines from crs package: crs

regr.earth
Multivariate Adaptive Regression Splines from earth package: earth

regr.fnn
Fast k-Nearest Neighbor from FNN package: knn

regr.gbm
Gradient boosting machine from gbm package: gbm
Note that distribution has been set to “gaussian” by default.

regr.IBk
K-nearest neighbours from RWeka package: IBk

regr.kknn
K-Nearest-Neighbor regression from kknn package: kknn

regr.km
Kriging from DiceKriging package: km

regr.ksvm
Support Vector Machines from kernlab package: ksvm
Note that kernel parameters have to be passed directly and not by using the kpar list in ksvm.
Note that fit has been set to FALSE by default for speed.

regr.penalized.lasso
Lasso regression from penalized package: penalized

regr.lm
Simple linear regression from stats package: lm

regr.mars
Multivariate Adaptive Regression Splines from mda package: mars

regr.mob
Model-based recursive partitioning yielding a tree with fitted models associated with each terminal node from party package: mob

regr.nnet
Neural Network from nnet package: nnet
Note that size has been set to 3 by default.

regr.pcr
Principal component regression from pls package: pcr
Note that model has been set to FALSE by default for speed.

regr.randomForest
Random Forest from randomForest package: randomForest. The argument fix.factors restores the factor levels seen in the training data before prediction to circumvent randomForest's internal sanity checks. Default is FALSE.

regr.penalized.ridge
Ridge regression from penalized package: penalized

regr.rpart
Decision Tree from rpart package: rpart
Note that xval has been set to 0 by default for speed.

regr.rsm
Response surface regression from rsm package: rsm
Note that you select the order of the regression by using modelfun="FO" (first order), "TWI" (two-way interactions, this is with 1st oder terms!) and "SO" (full second order).

regr.rvm
Relevance Vector Machine from package kernlab: rvm
Note that kernel parameters have to be passed directly and not by using the kpar list in rvm.
Note that fit has been set to FALSE by default for speed.

surv.CoxBoost
Cox proportional hazards model with componentwise likelhood based boosting from CoxBoost package: CoxBoost

surv.coxph
Cox proportional hazard model from survival package: coxph

surv.glmnet
GLM with regularization from glmnet package: glmnet

surv.randomForestSRC
Random Forests for Survival from randomForestSRC package: randomForestSRC

[Package mlr version 1.2 Index]