listLearners {mlr}	R Documentation

Find matching learning algorithms.

Description

Returns the class names of learning algorithms which have specific characteristics, e.g. whether they supports missing values, case weights, etc.

listLearnersForTask returns all learners that are in principle applicable for a given task.

Note that the packages of all learners are loaded during the search.

Usage

listLearners(obj = NA_character_, properties = character(0L),
  quiet = TRUE, warn.missing.packages = TRUE)

## Default S3 method:
listLearners(obj, properties = character(0L),
  quiet = TRUE, warn.missing.packages = TRUE)

## S3 method for class 'character'
listLearners(obj, properties = character(0L),
  quiet = TRUE, warn.missing.packages = TRUE)

## S3 method for class 'SupervisedTask'
listLearners(obj, properties = character(0L),
  warn.missing.packages = TRUE)

Arguments

type
[character(1)]
Type of the learning algorithm, one of “classif”, “regr” or “surv”. Default is NA, matching all types.

properties
[character)]
Set of required properties to filter for. Default is character(0).

quiet
[logical(1)]
Construct learners quietly to check their properties, shows no package startup messages. Turn off if you suspect errors. Default is codeTRUE.

warn.missing.packages
[logical(1)]
If some learner cannot be constructed because its package is missing, should a warning be shown? Default is codeTRUE.

task
[SupervisedTask]
The task.

Value

[character]. Class names of matching learners.

[Package mlr version 1.2 Index]