listLearners {mlr}	R Documentation

Find matching learning algorithms.

Description

Returns the class names of learning algorithms which have specific characteristics, e.g. whether they supports missing values, case weights, etc.

Usage

listLearners(type = NA, numerics = NA, factors = NA, missings = NA,
  weights = NA, oneclass = NA, twoclass = NA, multiclass = NA,
  prob = NA, se = NA, quiet = TRUE, warn.missing.packages = TRUE)

listLearnersForTask(task, weights = NA, prob = NA, se = NA,
  warn.missing.packages = TRUE)

Arguments

type
[character(1)]
Type of the learning algorithm, one of “classif”, “regr” or “surv”.

numerics
[logical(1)]
Supports real-valued features?

factors
[logical(1)]
Supports factor features?

missings
[logical(1)]
Supports missing values in features?

weights
[logical(1)]
Supports case weights?

oneclass
[logical(1)]
Supports oneclass problems?

twoclass
[logical(1)]
Supports twoclass problems?

multiclass
[logical(1)]
Supports multiclass problems?

prob
[logical(1)]
Can predict probabilities (classification)?

se
[logical(1)]
Can predict standard errors (regression)?

quiet
[logical(1)]
Construct learners quietly to check their properties, shows no package startup messages. Turn off if you suspect errors. Default is codeTRUE.

warn.missing.packages
[logical(1)]
If some learner cannot be constructed because its package is missing, should a warning be shown? Default is codeTRUE.

task
[SupervisedTask]
The task. Learners are returned that are applicable.

Details

The default for all search parameters is NA, meaning: property is not required, do not care.

Note that the packages of all learners are loaded during the search.

Value

[character]. Class names of matching learners.

[Package mlr version 1.2 Index]