listLearners {mlr}	R Documentation

Find matching learning algorithms.

Description

Returns the class names of learning algorithms which have specific characteristics, e.g. whether they supports missing values, case weights, etc.

Note that the packages of all learners are loaded during the search.

Note that for general cost-sensitive learning, mlr currently supports mainly “wrapper” approaches like CostSensWeightedPairsWrapper, which are not listed, as they are not basic R learning algorithms.

Usage

listLearners(obj = NA_character_, properties = character(0L),
  quiet = TRUE, warn.missing.packages = TRUE, create = FALSE)

## Default S3 method:
listLearners(obj, properties = character(0L),
  quiet = TRUE, warn.missing.packages = TRUE, create = FALSE)

## S3 method for class 'character'
listLearners(obj, properties = character(0L),
  quiet = TRUE, warn.missing.packages = TRUE, create = FALSE)

## S3 method for class 'Task'
listLearners(obj, properties = character(0L), quiet = TRUE,
  warn.missing.packages = TRUE, create = FALSE)

Arguments

obj
[character(1) | Task]
Either a task or the type of the task, in the latter case one of: “classif”, “regr”, “surv”, “costsens”, “cluster”. Default is NA, matching all types.

properties
[character]
Set of required properties to filter for. Default is character(0).

quiet
[logical(1)]
Construct learners quietly to check their properties, shows no package startup messages. Turn off if you suspect errors. Default is TRUE.

warn.missing.packages
[logical(1)]
If some learner cannot be constructed because its package is missing, should a warning be shown? Default is TRUE.

create
[logical(1)]
Instantiate objects (or return strings)? Default is FALSE.

Value

[character | list of Learner]. Class names of matching learners or instantiated objects.

[Package mlr version 2.1 Index]