listMeasures {mlr}	R Documentation

Find matching measures.

Description

Returns the matching measures which have specific characteristics, e.g. whether they supports classification or regression.

Usage

listMeasures(obj, properties = character(0L), create = FALSE)

## Default S3 method:
listMeasures(obj, properties = character(0L),
  create = FALSE)

## S3 method for class 'character'
listMeasures(obj, properties = character(0L),
  create = FALSE)

## S3 method for class 'SupervisedTask'
listMeasures(obj, properties = character(0L),
  create = FALSE)

Arguments

obj
[character(1) | SupervisedTask]
Either a task or the type of the task, in the latter case one of: “classif”, “regr”, “surv”, dQuotecostsens. Default is NA, matching all types.

properties
[character]
Set of required properties to filter for. See Measure for some standardized properties. Default is character(0).

create
[logical(1)]
Instantiate objects (or return strings)? Default is FALSE.

Value

[character | list of Measure]. Class names of matching measures or instantiated objects.

[Package mlr version 2.0 Index]