make.learner {mlr}	R Documentation

Create learner object.

Description

Create learner object.

Usage

make.learner(class, id, label, predict.type="response",
    predict.threshold=numeric(0), hyper.types="train", ...)

Arguments

class 	 [string]
Class of learner to create.

id 	 [string]
Id string for object. Used to select the object from a named list, etc.

label 	 [string]
Label string for object. Used in plots, etc.

predict.type 	 [string]
Classification: "response" | "prob" | "decision", specifying the type to predict. Default is "response". "decision" is experimental. Ignored for regression.

hyper.types 	 [character]
Character vector. If named, names have to correpond to the hyper pars in the ... arguments. If a single string is passed it is used as the type for all hyper parameters. Specifies in which stage of the model fitting the hyper parameters are used. Currently supported at user level are "train" and "predict". Default is "train".

predict.threshold 	 [numeric]
Threshold to produce class labels if type is not "response". Currently only supported for binary classification and type="prob", where it represents the required predicted probability for the positive class, so that a positive class is predicted as "response". Default is 0.5 for type="prob". Ignored for regression.

... 	 [any]
Optional named (hyper)parameters.

Value

learner.

[Package mlr version 0.3.732 Index]