make.multiclass.wrapper {mlr}	R Documentation

Fuse learner with multiclass method.

Description

Fuses a base learner with a multi-class method. Creates a learner object, which can be used like any other learner object. This way learners which can only handle binary classification will be able to handle multi-class problems too. Currently only "one-vs-all" is implemented.

Usage

make.multiclass.wrapper(learner, ...)

Arguments

learner 	 [learner or string]
Learning algorithm. See learners.

... 	 [any]
Optional parameters. Not used currently.

Value

learner.

[Package mlr version 0.3.819 Index]