makeAggregation {mlr}	R Documentation

Specifiy your own aggregation of measures

Description

This is an adavanced feature of mlr. It gives access to some inner workings so the result might not be compatible with everything!

Usage

makeAggregation(id, fun)

Arguments

id
[character(1)]
Name of the aggregation method. (Preferably the same name as the generated function)

fun
[function]
A function with following signature: function(task, perf.test, perf.train, measure, group, pred)

task: task (SupervisedTask) object

perf.test: numerical vector of performance results on the test data set

perf.train. numerical vector of performance results on the train data set

measure: Measure object.

group: grouping vector

pred: Prediction object

Value

Aggregation object

See Also

aggregations, setAggregation

Examples

# computes the interquartile range on all performance values
test.iqr = makeAggregation(id="test.iqr",
  fun = function (task, perf.test, perf.train, measure, group, pred) IQR(perf.test))

[Package mlr version 1.2 Index]