makeCostMeasure {mlr}	R Documentation

Creates a measure for non-standard misclassification costs.

Description

Creates a measure for non-standard misclassification costs.

Usage

  makeCostMeasure(id = "costs", minimize = TRUE, costs,
    task, aggregate = mean)

Arguments

id
[character(1)]
Name of measure. Default is “costs”.

minimize
[logical(1)]
Should the measure be minimized? Default is TRUE. Otherwise you are effectively specifying a benefits matrix.

costs
[matrix]
Matrix of misclassification costs. Rows and columns have to be named with class labels, order does not matter. Rows indicate true classes, columns predicted classes.

task
[ClassifTask]
Classification task. Has to be passed, so validity of matrix names can be checked.

aggregate
[logical(1)]
How to aggregate costs over all cases in one test set prediction? Default is mean.

Value

[Measure].

See Also

measures, makeMeasure

[Package mlr version 1.2 Index]