makeCostMeasure {mlr}	R Documentation

Creates a measure for non-standard misclassification costs.

Description

Creates a measure for non-standard misclassification costs.

Usage

makeCostMeasure(id = "costs", minimize = TRUE, costs, task,
  combine = mean, best = NULL, worst = NULL)

Arguments

id
[character(1)]
Name of measure. Default is “costs”.

minimize
[logical(1)]
Should the measure be minimized? Otherwise you are effectively specifying a benefits matrix. Default is TRUE.

costs
[matrix]
Matrix of misclassification costs. Rows and columns have to be named with class labels, order does not matter. Rows indicate true classes, columns predicted classes.

task
[ClassifTask]
Classification task. Has to be passed, so validity of matrix names can be checked.

combine
[function]
How to combine costs over all cases for a SINGLE test set? Note this is not the same as the aggregate argument in makeMeasure You can set this as well via setAggregation, as for any measure. Default is mean.

best
[numeric(1)]
Best obtainable value for measure. Default is -Inf or Inf, depending on minimize.

worst
[numeric(1)]
Worst obtainable value for measure. Default is Inf or -Inf, depending on minimize.

Value

[Measure].

See Also

Other performance: G1, G2, acc, auc, bac, ber, cindex, db, dunn, f1, fdr, featperc, fn, fnr, fp, fpr, gmean, gpr, mae, mcc, mcp, meancosts, measures, medae, medse, mmce, mse, multiclass.auc, npv, ppv, rmse, sae, silhouette, sse, timeboth, timepredict, timetrain, tn, tnr, tp, tpr; Measure, makeMeasure; makeCustomResampledMeasure; performance

[Package mlr version 2.1 Index]