makeFeatSelWrapper {mlr}	R Documentation

Fuse learner with feature selection.

Description

Fuses a base learner with a search strategy to select variables. Creates a learner object, which can be used like any other learner object, but which internally uses selectFeatures. If the train function is called on it, the search strategy and resampling are invoked to select an optimal set of variables. Finally, a model is fitted on the complete training data with these variables and returned. See selectFeatures for more details.

After training, the optimal features (and other related information) can be retrieved with getFeatSelResult.

Usage

makeFeatSelWrapper(learner, resampling, measures, bit.names, bits.to.features,
  control, show.info = getMlrOption("show.info"))

Arguments

learner
[Learner | character(1)]
The learner. If you pass a string the learner will be created via makeLearner.

resampling
[ResampleInstance | ResampleDesc]
Resampling strategy for feature selection. If you pass a description, it is instantiated once at the beginning by default, so all points are evaluated on the same training/test sets. If you want to change that behaviour, look at FeatSelControl.

measures
[list of Measure | Measure]
Performance measures to evaluate. The first measure, aggregated by the first aggregation function is optimized, others are simply evaluated.

bit.names
[character]
Names of bits encoding the solutions. Also defines the total number of bits in the encoding. Per default these are the feature names of the task.

bits.to.features
[function(x, task)]
Function which transforms an integer-0-1 vector into a character vector of selected features. Per default a value of 1 in the ith bit selects the ith feature to be in the candidate solution.

control
[see FeatSelControl] Control object for search method. Also selects the optimization algorithm for feature selection.

show.info
[logical(1)]
Print verbose output on console? Default is set via configureMlr.

Value

[Learner].

See Also

Other featsel: FeatSelControl, FeatSelControlExhaustive, FeatSelControlGA, FeatSelControlRandom, FeatSelControlSequential, makeFeatSelControlExhaustive, makeFeatSelControlGA, makeFeatSelControlRandom, makeFeatSelControlSequential; analyzeFeatSelResult; selectFeatures

Examples

# nested resampling with feature selection (with a pretty stupid algorithm for selection)
outer = makeResampleDesc("CV", iters = 2L)
inner = makeResampleDesc("Holdout")
ctrl = makeFeatSelControlRandom(maxit = 3)
lrn = makeFeatSelWrapper("classif.ksvm", resampling = inner, control = ctrl)

## Loading required package: kernlab

# we also extract the selected features for all iteration here
r = resample(lrn, iris.task, outer, extract = getFeatSelResult)

## [Resample] cross-validation iter: 1

## Using automatic sigma estimation (sigest) for RBF or laplace kernel

## [selectFeatures] 1: 1 bits: mmce.test.mean= 0.2

## Using automatic sigma estimation (sigest) for RBF or laplace kernel

## [selectFeatures] 1: 4 bits: mmce.test.mean=0.04

## Using automatic sigma estimation (sigest) for RBF or laplace kernel

## [selectFeatures] 1: 3 bits: mmce.test.mean=0.08

## Using automatic sigma estimation (sigest) for RBF or laplace kernel

## [Resample] cross-validation iter: 2

## Using automatic sigma estimation (sigest) for RBF or laplace kernel

## [selectFeatures] 1: 2 bits: mmce.test.mean=0.32

## Using automatic sigma estimation (sigest) for RBF or laplace kernel

## [selectFeatures] 1: 2 bits: mmce.test.mean=0.08

## Using automatic sigma estimation (sigest) for RBF or laplace kernel

## [selectFeatures] 1: 1 bits: mmce.test.mean=0.08

## Using automatic sigma estimation (sigest) for RBF or laplace kernel

## [Resample] Result: mmce.test.mean=0.0467

[Package mlr version 1.2 Index]