makeFilterWrapper {mlr}	R Documentation

Fuse learner with filter method.

Description

Fuses a base learner with a filter method. Creates a learner object, which can be used like any other learner object. Internally Uses filterFeatures before every model fit.

Usage

  makeFilterWrapper(learner,
    fw.method = "random.forest.importance", fw.perc = 1)

Arguments

learner
[Learner]
The learner.

fw.method
[character(1)]
Filter method. Available are: linear.correlation, rank.correlation, information.gain, gain.ratio, symmetrical.uncertainty, chi.squared, random.forest.importance, relief, oneR Default is random.forest.importance.

fw.perc
[numeric(1)]
Percentage of highest ranking features to select after filtering. Default is 1 (=100 percent).

Details

Look at package FSelector for details on the filter algorithms.

After training, the selected features can be retrieved with getTuneResult.

Value

[Learner].

Examples

task = makeClassifTask(data=iris, target="Species")
lrn = makeLearner("classif.lda")
inner = makeResampleDesc("Holdout")
outer = makeResampleDesc("CV", iters = 2)
lrn = makeFilterWrapper(lrn, fw.perc=0.5)
mod = train(lrn, task)

## Lade nötiges Paket: FSelector

print(getFilteredFeatures(mod))

## [1] "Petal.Length" "Petal.Width"

r = resample(lrn, task, outer, extract = function(model) {
getFilteredFeatures(model)
})

## [Resample] cross-validation iter: 1
## [Resample] cross-validation iter: 2
## [Resample] Result: mmce.test.mean=0.0533

print(r$extract)

## [[1]]
## [1] "Petal.Length" "Petal.Width" 
## 
## [[2]]
## [1] "Petal.Width"  "Petal.Length"

[Package mlr version 1.2 Index]