makeImputeWrapper {mlr}	R Documentation

Fuse learner with an imputation method.

Description

Fuses a base learner with an imputation method. Creates a learner object, which can be used like any other learner object. Internally uses impute before training the learner and reimpute before predicting.

Usage

makeImputeWrapper(learner, classes = list(), cols = list(),
  dummy.cols = character(0L), dummy.type = "factor",
  impute.new.levels = TRUE, recode.factor.levels = TRUE)

Arguments

learner
[Learner | character(1)]
The learner. If you pass a string the learner will be created via makeLearner.

classes
[named list]
Named list containing imputation techniques for classes of columns. E.g. list(numeric = imputeMedian()).

cols
[named list]
Named list containing names of imputation methods to impute missing values in the data column referenced by the list element's name. Overwrites imputation set via classes.

dummy.cols
[character]
Column names to create dummy columns (containing binary missing indicator) for. Default is character(0).

dummy.type
[character(1)]
How dummy columns are encoded. Either as 0/1 with type “numeric” or as “factor”. Default is “factor”.

impute.new.levels
[logical(1)]
If new, unencountered factor level occur during reimputation, should these be handled as NAs and then be imputed the same way? Default is TRUE.

recode.factor.levels
[logical(1)]
Recode factor levels after reimputation, so they match the respective element of lvls (in the description object) and therefore match the levels of the feature factor in the training data after imputation?. Default is TRUE.

Value

[Learner].

See Also

Other impute: imputations, imputeConstant, imputeHist, imputeLearner, imputeMax, imputeMean, imputeMedian, imputeMin, imputeMode, imputeNormal, imputeUniform; impute; makeImputeMethod; reimpute

[Package mlr version 2.1 Index]