makeMeasure {mlr}	R Documentation

Construct performance measure.

Description

A measure object encapsulates a function to evaluate the performance of a prediction. Information about already implemented measures can be obtained here: measures.

A learner is trained on a a training set d1, results in a model m, predicts another set d2 (which may be a different one or the training set), resulting in the prediction. The performance measure can now be defined using all of the information of the original task, the fitted model and the prediction.

Object slots:

id [character(1)]

See argument.

minimize [logical(1)]

See argument.

properties [character]

See argument.

allowed.pred.types [character]

See argument.

req.pred [logical(1)]

Is prediction object required in calculation?

req.task [logical(1)]

Is task object required in calculation?.

req.model [logical(1)]

Is model object required in calculation?

fun [function]

See argument.

extra.args [list]

See argument.

aggr [Aggregation]

See argument.

.

best [numeric(1)]

See argument.

worst [numeric(1)]

See argument.

Usage

makeMeasure(id, minimize, properties = character(0L),
  allowed.pred.types = character(0L), fun, extra.args = list(),
  aggr = test.mean, best = NULL, worst = NULL)

Arguments

id
[character(1)]
Name of measure.

minimize
[logical(1)]
Should the measure be minimized? Default is in TRUE.

properties
[character]
Set of measure properties. Some standard property names include:

classif

Is the measure applicable for classification?

classif.multi

Is the measure applicable for multi-class classification?

regr

Is the measure applicable for regression?

surv

Is the measure applicable for survival?

costsens

Is the measure applicable for cost-sensitve learning?

Default is character(0).

allowed.pred.types
[character]
Which prediction types are allowed for this measure? Subset of “response”,“prob”, “se”. Default is character(0).

fun
[function(task, model, pred, extra.args)]
Calculates performance value.

extra.args
[list]
List of extra arguments which will always be passed to fun. Default is empty list.

aggr
[Aggregation]
Aggregation funtion, which is used to aggregate the values measured on test / training sets of the measure to a single value. Default is test.mean.

best
[numeric(1)]
Best obtainable value for measure. Default is -Inf or Inf, depending on minimize.

worst
[numeric(1)]
Worst obtainable value for measure. Default is Inf or -Inf, depending on minimize.

Value

[Measure].

See Also

Other performance: acc, auc, ber, cindex, f1, fdr, featperc, fn, fnr, fp, fpr, gmean, gpr, mae, mcc, mcp, meancosts, measures, medae, medse, mmce, mse, multiclass.auc, npv, ppv, rmse, sae, sse, timeboth, timepredict, timetrain, tn, tnr, tp, tpr; makeCostMeasure; makeCustomResampledMeasure; performance

Examples

f = function(task, model, pred, extra.args)
  sum((pred$data$response - pred$data$truth)^2)
makeMeasure(id = "my.sse", minimize = TRUE, properties = c("regr", "response"), fun = f)

## Performance measure: my.sse
## Properties: regr,response
## Minimize: TRUE
## Best: -Inf; Worst: Inf
## Aggregated by: test.mean

[Package mlr version 2.0 Index]