makeMeasure {mlr}	R Documentation

Construct performance measure.

Description

A measure object encapsulates a function to evaluate the performance of a prediction. Information about already implemented measures can be obtained here: measures.

Usage

  makeMeasure(id, minimize, classif = FALSE, regr = FALSE,
    only.binary = FALSE,
    allowed.pred.types = character(0L), fun,
    extra.args = list(), aggr = test.mean)

Arguments

id
[character(1)]
Name of measure.

minimize
[logical(1)]
Should the measure be minimized? Default is in TRUE.

classif
[logical(1)]
Is the measure applicable for classification? Default is FALSE.

regr
[logical(1)]
Is the measure applicable for regression? Default is FALSE.

only.binary
[logical(1)]
Is the measure only applicable to binary classification? Only reasonable if classif is TRUE. Default is FALSE.

allowed.pred.types
[character]
Which prediction types are allowed for this measure? Subset of “response”,“prob”, “se”. Default is character(0).

fun
[function(task, model, pred, extra.args)]
Calculates performance value.

extra.args
[list]
List of extra arguments which will always be passed to fun. Default is empty list.

aggr
[Aggregation]
Aggregation funtion, which is used to aggregate the values measured on test / training sets of the measure to a single value. Default is test.mean.

Details

A learner is trained on a a training set d1, results in a model m, predicts another set d2 (which may be a different one or the training set), resulting in the prediction. The performance measure can now be defined using all of the information of the original task, the fitted model and the prediction.

Object slots:

id [character(1)]

See argument.

minimize [logical(1)]

See argument.

classif [logical(1)]

See argument.

regr [logical(1)]

See argument.

only.binary [logical(1)]

See argument.

allowed.pred.types [character]

See argument.

req.pred [logical(1)]

Is prediction object required in calculation?

req.task [logical(1)]

Is task object required in calculation?.

req.model [logical(1)]

Is model object required in calculation?

fun [function]

See argument.

extra.args [list]

See argument.

aggr [Aggregation]

See argument.

.

Value

[Measure].

Examples

f <- function(task, model, pred, extra.args)
  sum((pred$data$response - pred$data$truth)^2)
makeMeasure(id="my.sse", minimize=TRUE, regr=TRUE, allowed.pred.types="response", fun=f)

## Performance measure: my.sse
## Minimize: TRUE
## Aggregated by: test.mean

[Package mlr version 1.2 Index]