makeModelMultiplexerParamSet {mlr}	R Documentation

Creates a parameter set for model multiplexer tuning.

Description

Handy way to create the param set with too much typing.

The following is done automatically:

The selected.learner param is created

Parameter names are prefixed.

The requires field of each param is set. This makes all parameters subordinate to selected.learner

Usage

makeModelMultiplexerParamSet(multiplexer, ..., .check = TRUE)

Arguments

multiplexer
[ModelMultiplexer]
The muliplexer learner.

...
[ParamSet | Param]
(a) First option: Named param sets. Names must correspond to base learners. You only need to enter the parameters you want to tune without reference to the selected.learner field in any way.
(b) Second option. Just the params you would enter in the param sets. Even shorterto create. Only works when it can be uniquely identified to which learner each of your passed parameters belongs.

.check
[logical]
Check that for each param in ... one param in found in the base learners. Default is TRUE

Value

[ParamSet].

See Also

Other multiplexer: ModelMultiplexer, makeModelMultiplexer

Other tune: ModelMultiplexer, makeModelMultiplexer; TuneControl, TuneControlCMAES, TuneControlGrid, TuneControlIrace, TuneControlOptim, TuneControlRandom, makeTuneControlCMAES, makeTuneControlGrid, makeTuneControlIrace, makeTuneControlOptim, makeTuneControlRandom; getTuneResult; makeTuneWrapper; tuneParams; tuneThreshold

Examples

# See makeModelMultiplexer

[Package mlr version 1.2 Index]