Instantiates a resampling strategy object.

This class encapsulates training and test sets generated from the data set for a number of iterations. It mainly stores a set of integer vectors indicating the training and test examples for each iteration.

makeResampleInstance(desc, task, size, ...)

Arguments

desc
(ResampleDesc | character(1))
Resampling description object or name of resampling strategy. In the latter case makeResampleDesc will be called internally on the string.

task
(Task)
Data of task to resample from. Prefer to pass this instead of size.

size
(integer)
Size of the data set to resample. Can be used instead of task.

...
(any)
Passed down to makeResampleDesc in case you passed a string in desc. Otherwise ignored.

Value

(ResampleInstance).

Details

Object slots:

desc (ResampleDesc)

See argument.

size (integer(1))

See argument.

train.inds (list of integer)

List of of training indices for all iterations.

test.inds (list of integer)

List of of test indices for all iterations.

group (factor)

Optional grouping of resampling iterations. This encodes whether specific iterations 'belong together' (e.g. repeated CV), and it can later be used to aggregate performance values accordingly. Default is 'factor()'.

See also

Examples

rdesc = makeResampleDesc("Bootstrap", iters = 10)
rin = makeResampleInstance(rdesc, task = iris.task)

rdesc = makeResampleDesc("CV", iters = 50)
rin = makeResampleInstance(rdesc, size = nrow(iris))

rin = makeResampleInstance("CV", iters = 10, task = iris.task)