makeResampleInstance {mlr}	R Documentation

Instantiates a resampling strategy object.

Description

This class encapsulates training and test sets generated from the data set for a number of iterations. It mainly stores a set of integer vectors indicating the training and test examples for each iteration.

Usage

makeResampleInstance(desc, task, size)

Arguments

desc
[ResampleDesc]
Resampling description object.

task
[SupervisedTask]
Data of task to resample from. Prefer to pass this instead of size.

size
[integer]
Size of the data set to resample. Can be used instead of task.

Details

Object slots:

desc [ResampleDesc]

See argument.

size [integer(1)]

See argument.

train.inds [list of integer]

List of of training indices for all iterations.

test.inds [list of integer]

List of of test indices for all iterations.

group [factor]

Optional grouping of resampling iterations. This encodes whether specfic iterations 'belong together' (e.g. repeated CV), and it can later be used to aggregate performance values accordingly. Default is 'factor()'.

Value

[ResampleInstance].

See Also

makeResampleDesc, resample

Examples

task <- makeClassifTask(data = iris, target = "Species")
rdesc <- makeResampleDesc("Bootstrap", iters = 10)
rin <- makeResampleInstance(rdesc, task = task)

# Alternatively provide the size argument instead of a task object
rdesc <- makeResampleDesc("CV", iters = 50)
rin <- makeResampleInstance(rdesc, size = nrow(iris))

[Package mlr version 1.2 Index]