makeSMOTEWrapper {mlr}	R Documentation

Fuse learner with SMOTE oversampling for imbalancy correction in binary classification.

Description

Creates a learner object, which can be used like any other learner object. Internally uses smote before every model fit.

Note that observation weights do not influence the sampling and are simply passed down to the next learner.

Usage

makeSMOTEWrapper(learner, sw.rate, sw.nn = 5L)

Arguments

learner
[Learner | character(1)]
The learner. If you pass a string the learner will be created via makeLearner.

sw.rate
[numeric(1)]
Factor to oversample the smaller class. Must be between 1 and Inf, where 1 means no oversampling and 2 would mean doubling the class size.

sw.nn
[integer(1)]
Number of nearest neighbors to consider. Default is 5.

Value

[Learner].

[Package mlr version 2.0 Index]