makeUndersampleWrapper {mlr}	R Documentation

Fuse learner with simple ove/underrsampling for imbalancy correction in binary classification.

Description

Creates a learner object, which can be used like any other learner object. Internally uses oversample or undersample before every model fit.

Note that observation weights do not influence the sampling and are simply passed down to the next learner.

Usage

makeUndersampleWrapper(learner, usw.rate = 1)

makeOversampleWrapper(learner, osw.rate = 1)

Arguments

learner
[Learner | character(1)]
The learner. If you pass a string the learner will be created via makeLearner.

usw.rate
[numeric(1)]
Factor to downsample the bigger class. Must be between 0 and 1, where 1 means no downsampling, 0.5 implies reduction to 50 percent and 0 would imply reduction to 0 observations. Default is 1.

osw.rate
[numeric(1)]
Factor to oversample the smaller class. Must be between 1 and Inf, where 1 means no oversampling and 2 would mean doubling the class size. Default is 1.

Value

[Learner].

See Also

Other imbalancy: makeOverBaggingWrapper; oversample, undersample; smote

[Package mlr version 2.1 Index]