makeWeightedClassesWrapper {mlr}	R Documentation

Wraps a classifier for weighted fitting where each class receives a weight.

Description

Creates a wrapper, which can be used like any other learner object.

Fitting is performed in a weighted fashion where each observation receives a weight, depending on the class it belongs to, see wcw.weight. This might help to mitigate problems caused by imbalanced class distributions.

This weighted fitting can be achieved in two ways:

a) The learner already has a parameter for class weighting, so one weight can directly be defined per class. Example: “classif.ksvm” and parameter class.weights. In this case we don't really do anything fancy. We convert wcw.weight a bit, but basically simply bind its value to the class weighting param. The wrapper in this case simply offers a convenient, consistent fashion for class weighting - and tuning! See example below.

b) The learner does not have a direct parameter to support class weighting, but supports observation weights, so hasProperties(learner, 'weights') is TRUE. This means that an individual, arbitrary weight can be set per observation during training. We set this weight depending on the class internally in the wrapper. Basically we introduce something like a new “class.weights” parameter for the learner via observation weights.

Usage

makeWeightedClassesWrapper(learner, wcw.param = NULL, wcw.weight = 1)

Arguments

learner
[Learner | character(1)]
The classification learner. If you pass a string the learner will be created via makeLearner.

wcw.param
[character(1)]
Name of already existing learner param which allows class weighting. Or NULL if such a param does not exist. Must during training accept a named vector of class weights, where length equals the number of classes. Default is NULL.

wcw.weight
[numeric]
Weight for each class. Must be a vector of the same number of elements as classes are in task, and must also be in the same order as the class levels are in task$task.desc$class.levels. For convenience, one must pass a single number in case of binary classification, which is then taken as the weight of the positive class, while the negative class receives a weight of 1. Default is 1.

Value

[Learner].

Examples

# using the direct parameter of the SVM
lrn = makeWeightedClassesWrapper("classif.ksvm", wcw.param = "class.weights", wcw.weight = 0.01)
res = holdout(lrn, sonar.task)

## [Resample] holdout iter: 1

## Using automatic sigma estimation (sigest) for RBF or laplace kernel

## [Resample] Result: mmce.test.mean=0.514

print(getConfMatrix(res$pred))

##        predicted
## true    M  R -SUM-
##   M     0 36    36
##   R     0 34     0
##   -SUM- 0 36    36

# using the observation weights of logreg
lrn = makeWeightedClassesWrapper("classif.logreg", wcw.weight = 0.01)
res = holdout(lrn, sonar.task)

## [Resample] holdout iter: 1

## Warning: glm.fit: algorithm did not converge
## Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred

## [Resample] Result: mmce.test.mean=0.386

print(getConfMatrix(res$pred))

##        predicted
## true     M  R -SUM-
##   M     21 13    13
##   R     14 22    14
##   -SUM- 14 13    27

# tuning the imbalancy param and the SVM param in one go
lrn = makeWeightedClassesWrapper("classif.ksvm", wcw.param = "class.weights")
ps = makeParamSet(
  makeNumericParam("wcw.weight", lower = 1, upper = 10),
  makeNumericParam("C", lower = -12, upper = 12, trafo = function(x) 2^x),
  makeNumericParam("sigma", lower = -12, upper = 12, trafo = function(x) 2^x)
)
ctrl = makeTuneControlRandom(maxit = 3L)
rdesc = makeResampleDesc("CV", iters = 2L, stratify = TRUE)
res = tuneParams(lrn, sonar.task, rdesc, par.set = ps, control = ctrl)

## [Tune] Started tuning learner weightedclasses.classif.ksvm for parameter set:
##               Type len Def    Constr Req Trafo
## wcw.weight numeric   -   -   1 to 10   -     -
## C          numeric   -   - -12 to 12   -     Y
## sigma      numeric   -   - -12 to 12   -     Y
## With control class: TuneControlRandom
## Imputation value: 1
## [Tune] 1: wcw.weight=2.36; C=0.916; sigma=32.3 : mmce.test.mean=0.466
## [Tune] 2: wcw.weight=6.68; C=0.973; sigma=2.43e+03 : mmce.test.mean=0.466
## [Tune] 3: wcw.weight=6.06; C=0.122; sigma=0.0155 : mmce.test.mean=0.466
## [Tune] Result: wcw.weight=6.68; C=0.973; sigma=2.43e+03 : mmce.test.mean=0.466

print(res)

## Tune result:
## Op. pars: wcw.weight=6.68; C=0.973; sigma=2.43e+03
## mmce.test.mean=0.466

print(res$opt.path)

## Optimization path
##   Dimensions: x = 3/3, y = 1
##   Length: 3
##   Add x values transformed: FALSE
##   Error messages: TRUE. Errors: 0 / 3.
##   Exec times: TRUE. Range: 0.076 - 0.08.

[Package mlr version 2.1 Index]