makeWeightedClassesWrapper {mlr}	R Documentation

Wraps a classifier for weighted fitting where each class receives a weight.

Description

Creates a wrapper, which can be used like any other learner object.

Fitting is performed in a weighted fashion where each observation receives a weight, depending on the class is belongs to, see wcw.weight. This might help to mitigate problems in imbalanced classification problems.

Usage

makeWeightedClassesWrapper(learner, wcw.weight)

Arguments

learner
[Learner | character(1)]
The classification learner. If you pass a string the learner will be created via makeLearner.

wcw.weight
[numeric]
Weight for each class. Must be a vector of the same number of elements as classes are in task, and must also be in the same order as the class levels are in task$task.desc$class.levels. For convenience, one must pass a single number in case of binary classification, which is then taken as the weight of the positive class, while the negative class receives a weight of 1.

Value

[Learner].

[Package mlr version 2.0 Index]