Wraps a classifier for weighted fitting where each class receives a weight.

Creates a wrapper, which can be used like any other learner object.

Fitting is performed in a weighted fashion where each observation receives a weight, depending on the class it belongs to, see wcw.weight. This might help to mitigate problems caused by imbalanced class distributions.

This weighted fitting can be achieved in two ways:

a) The learner already has a parameter for class weighting, so one weight can directly be defined per class. Example: “classif.ksvm” and parameter class.weights. In this case we don't really do anything fancy. We convert wcw.weight a bit, but basically simply bind its value to the class weighting param. The wrapper in this case simply offers a convenient, consistent fashion for class weighting - and tuning! See example below.

b) The learner does not have a direct parameter to support class weighting, but supports observation weights, so hasLearnerProperties(learner, 'weights') is TRUE. This means that an individual, arbitrary weight can be set per observation during training. We set this weight depending on the class internally in the wrapper. Basically we introduce something like a new “class.weights” parameter for the learner via observation weights.

Usage

makeWeightedClassesWrapper(learner, wcw.param = NULL, wcw.weight = 1)

Arguments

learner

(Learner | character(1))
The classification learner. If you pass a string the learner will be created via makeLearner.

wcw.param

(character(1))
Name of already existing learner parameter, which allows class weighting. The default (wcw.param = NULL) will use the parameter defined in the learner (class.weights.param). During training, the parameter must accept a named vector of class weights, where length equals the number of classes.

wcw.weight

(numeric)
Weight for each class. Must be a vector of the same number of elements as classes are in task, and must also be in the same order as the class levels are in getTaskDesc(task)$class.levels. For convenience, one must pass a single number in case of binary classification, which is then taken as the weight of the positive class, while the negative class receives a weight of 1. Default is 1.

Value

Learner.

See also

Examples

# \donttest{
set.seed(123)
# using the direct parameter of the SVM (which is already defined in the learner)
lrn = makeWeightedClassesWrapper("classif.ksvm", wcw.weight = 0.01)
res = holdout(lrn, sonar.task)
#> Resampling: holdout
#> Measures:             mmce      
#> [Resample] iter 1:    0.5428571 
#> 
#> Aggregated Result: mmce.test.mean=0.5428571
#> 
print(calculateConfusionMatrix(res$pred))
#>         predicted
#> true     M  R -err.-
#>   M      0 38     38
#>   R      0 32      0
#>   -err.- 0 38     38

# using the observation weights of logreg
lrn = makeWeightedClassesWrapper("classif.logreg", wcw.weight = 0.01)
res = holdout(lrn, sonar.task)
#> Resampling: holdout
#> Measures:             mmce      
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> [Resample] iter 1:    0.3285714 
#> 
#> Aggregated Result: mmce.test.mean=0.3285714
#> 
print(calculateConfusionMatrix(res$pred))
#>         predicted
#> true      M  R -err.-
#>   M      28  7      7
#>   R      16 19     16
#>   -err.- 16  7     23

# tuning the imbalancy param and the SVM param in one go
lrn = makeWeightedClassesWrapper("classif.ksvm", wcw.param = "class.weights")
ps = makeParamSet(
  makeNumericParam("wcw.weight", lower = 1, upper = 10),
  makeNumericParam("C", lower = -12, upper = 12, trafo = function(x) 2^x),
  makeNumericParam("sigma", lower = -12, upper = 12, trafo = function(x) 2^x)
)
ctrl = makeTuneControlRandom(maxit = 3L)
rdesc = makeResampleDesc("CV", iters = 2L, stratify = TRUE)
res = tuneParams(lrn, sonar.task, rdesc, par.set = ps, control = ctrl)
#> [Tune] Started tuning learner weightedclasses.classif.ksvm for parameter set:
#>               Type len Def    Constr Req Tunable Trafo
#> wcw.weight numeric   -   -   1 to 10   -    TRUE     -
#> C          numeric   -   - -12 to 12   -    TRUE     Y
#> sigma      numeric   -   - -12 to 12   -    TRUE     Y
#> With control class: TuneControlRandom
#> Imputation value: 1
#> [Tune-x] 1: wcw.weight=1.11; C=441; sigma=0.0013
#> [Tune-y] 1: mmce.test.mean=0.2644231; time: 0.0 min
#> [Tune-x] 2: wcw.weight=3.81; C=0.336; sigma=6.86
#> [Tune-y] 2: mmce.test.mean=0.4663462; time: 0.0 min
#> [Tune-x] 3: wcw.weight=4.05; C=3.78; sigma=241
#> [Tune-y] 3: mmce.test.mean=0.4663462; time: 0.0 min
#> [Tune] Result: wcw.weight=1.11; C=441; sigma=0.0013 : mmce.test.mean=0.2644231
print(res)
#> Tune result:
#> Op. pars: wcw.weight=1.11; C=441; sigma=0.0013
#> mmce.test.mean=0.2644231
# print(res$opt.path)
# }