measures {mlr}	R Documentation

Performance measures.

Description

A performance measure is evaluated after a single train/predict step and returns a single number to assess the quality of the prediction (or maybe only the model, think AIC). The measure itself knows whether it wants to be minimized or maximized and for what tasks it is applicable. See below for a list of already implemented measures. If you want a measure for a misclassification cost matrix, look at makeCostMeasure. If you want to implement your own measure, look at makeMeasure.

Classification:

mmce
Mean misclassification error.

acc
Accuracy.

ber
Balanced error rate. Mean of misclassification error rates on all individual classes.

tp
True positives.

tpr
True positive rate, also called hit rate or recall.

fp
False positives, also called false alarms.

fpr
False positive rate, also called false alarm rate or fall-out.

tn
True negatives, also called correct rejections.

tnr
True negative rate. Also called specificity.

fn
False negatives, also called misses.

fnr
False negative rate.

ppv
Positive predictive value, also called precision.

npv
Negative predictive value.

fdr
False discovery rate.

f1
F1 measure.

mcc
Matthews correlation coefficient.

gmean
G-mean, geometric mean of recall and specificity.

gpr
Geometric mean of precision and recall.

auc
Area under the curve.

multiclass.auc
Area under the curve for multiclass problems. Calls pROC::multiclass.roc.

Only mmce, acc, multiclass.auc and ber can be used for multiclass problems.

Regression:

sse
Sum of squared errors

mse
Mean of squared errors

medse
Median of squared errors

sae
Sum of absolute errors

mae
Mean of absolute errors

medae
Median of absolute errors

rmse
Root mean square error

Survival:

cindex
Concordance index

Cost-sensitive:

meancosts
Mean costs of the predicted choices.

mcp
Misclassification penalty, i.e. average difference between costs of oracle and model prediction.

General:

timetrain
Time of fitting the model

timepredict
Time of predicting test set

timeboth
timetrain + trainpredict

featperc
Percentage of original features used for model, useful for feature selection.

Usage

none

none

none

none

none

none

none

none

none

none

none

none

none

none

none

none

none

none

none

none

none

none

none

none

none

none

none

none

none

none

none

none

none

meancosts

mcp

Format

none

See Also

Other performance: Measure, makeMeasure; makeCostMeasure; makeCustomResampledMeasure; performance

[Package mlr version 2.0 Index]