normalizeFeatures {mlr}	R Documentation

Normalize features

Description

Normalize features by different methods. Internally normalize is used. Non numerical features will be left untouched and passed to the result.

Usage

normalizeFeatures(obj, target = character(0L), method = "standardize",
  exclude = character(0L), range = c(0, 1))

Arguments

obj
[data.frame | Task]
Input data.

target
[character(1)]
Name of the column(s) specifying the response. Only used when obj is a data.frame, otherwise ignored.

method
[character(1)]
Normalizing method.
Available are:
“center”: centering of each feature
“scale”: scaling of each feature
“standardize”: centering and scaling
“range”: Scale the data to a given range.

exclude
[character]
Names of the columns to exclude. The target does not have to be included here. Default is none.

range
[numeric(2)]
Range the features should be scaled to. Default is c(0,1).

Value

[data.frame | Task]. Same type as obj.

See Also

normalize

[Package mlr version 2.1 Index]