oversample {mlr}	R Documentation

Over- or undersample binary classification task to handle class imbalancy.

Description

Oversampling: From the smaller class, observations are randomly drawn with repetitions.

Undersampling: From the larger class, observations are randomly drawn without repetitions.

Usage

oversample(obj, target, rate)

undersample(obj, target, rate)

Arguments

obj
[data.frame | ClassifTask]
Input data.

target
[character(1)]
Name of the column specifying the response. Only used when obj is a data.frame, otherwise ignored.

rate
[numeric(1)]
Factor to upsample the smaller or downsample the bigger class. For undersampling: Must be between 0 and 1, where 1 means no downsampling, 0.5 implies reduction to 50 percent and 0 would imply reduction to 0 observations. For oversampling: Must be between 1 and Inf, where 1 means no oversampling and 2 would mean doubling the class size.

Value

[data.frame | ClassifTask]. Same type as obj.

See Also

Other OverUndersample: makeOversampleWrapper, makeUndersampleWrapper

[Package mlr version 1.2 Index]