performance {mlr}	R Documentation

Measure performance of prediction.

Description

Measures the quality of a prediction w.r.t. some performance measure.

Usage

performance(pred, measures, task, model)

Arguments

pred
[Prediction]
Prediction object.

measures
[Measure | list of Measure]
Performance measure(s) to evaluate.

task
[Task]
Learning task, might be requested by performance measure, usually not needed except for clustering.

model
[WrappedModel]
Model built on training data, might be requested by performance measure, usually not needed.

Value

[named numeric]. Performance value(s), named by measure(s).

See Also

Other performance: G1, G2, acc, auc, bac, ber, cindex, db, dunn, f1, fdr, featperc, fn, fnr, fp, fpr, gmean, gpr, mae, mcc, mcp, meancosts, measures, medae, medse, mmce, mse, multiclass.auc, npv, ppv, rmse, sae, silhouette, sse, timeboth, timepredict, timetrain, tn, tnr, tp, tpr; Measure, makeMeasure; makeCostMeasure; makeCustomResampledMeasure

Examples

training.set = seq(1, nrow(iris), by = 2)
test.set = seq(2, nrow(iris), by = 2)

task = makeClassifTask(data = iris, target = "Species")
lrn = makeLearner("classif.lda")
mod = train(lrn, task, subset = training.set)
pred = predict(mod, newdata = iris[test.set, ])
performance(pred, measures = mmce)

## mmce 
## 0.04

# Compute multiple performance measures at once
ms = list("mmce" = mmce, "acc" = acc, "timetrain" = timetrain)
performance(pred, measures = ms, task, mod)

##           mmce.mmce             acc.acc timetrain.timetrain 
##               0.040               0.960               0.003

[Package mlr version 2.1 Index]