performance {mlr}	R Documentation

Measure performance of prediction.

Description

Measures the quality of a prediction w.r.t. some performance measure.

Usage

  performance(pred, measures, task, model)

Arguments

pred
[Prediction]
Prediction object to evaluate.

measures
[Measure | list of Measure]
Performance measure(s) to evaluate.

task
[SupervisedTask]
Learning task, might be requested by performance measure, usually not needed.

model
[WrappedModel]
Model built on training data, might be requested by performance measure, usually not needed.

Value

A single numerical performance value.

See Also

makeMeasure, measures

Examples

training.set <- seq(1, nrow(iris), by = 2)
test.set <- seq(2, nrow(iris), by = 2)

task <- makeClassifTask(data = iris, target = "Species")
lrn <- makeLearner("classif.lda")
mod <- train(lrn, task, subset = training.set)
pred <- predict(mod, newdata = iris[test.set, ])

## Here we define the mean misclassification error (MMCE) as our performance measure
my.mmce <- function(task, model, pred, extra.args) {
  length(which(pred$data$response != pred$data$truth)) / nrow(pred$data)
}
ms <- makeMeasure(id = "misclassification.rate",
                  minimize = TRUE,
                  classif = TRUE,
                  allowed.pred.types = "response",
                  fun = my.mmce)
performance(pred, ms, task, mod)

## [1] 0.04

## Indeed the MMCE is already implemented in mlr beside other common performance measures
performance(pred, measures = mmce)

## [1] 0.04

## Compute multiple performance measures at once
ms <- list("mmce" = mmce, "acc" = acc, "timetrain" = timetrain)
sapply(ms, function(the.ms) {
  performance(pred, measures = the.ms, task, mod)
})

##      mmce       acc timetrain 
##     0.040     0.960     0.009

[Package mlr version 1.2 Index]