predict {mlr}	R Documentation

Predict new data.

Description

Predict the target variable of new data using a fitted model. If the type is set to "prob" or "decision" probabilities or decision values will be stored in the resulting object. The resulting class labels are the classes with the maximum values or thresholding can also be used.

Arguments

object 	 [wrapped.model]
Wrapped model, trained from a learn task.

task 	 [learn.task]
Specifies learning task. If this is passed, data from this task is predicted.

subset 	 [integer]
Index vector to subset the data in the task to use for prediction.

newdata 	 [data.frame]
New observations which should be predicted. Alternatively pass this instead of task.

type 	 [string]
Classification: "response" | "prob" | "decision", specifying the type to predict. Default is "response". "decision" is experimental. Ignored for regression.

threshold 	 [numeric]
Threshold to produce class labels if type is not "response". Currently only supported for binary classification and type="prob", where it represents the required predicted probability for the positive class, so that a positive class is predicted as "response". Default is 0.5 for type="prob". Ignored for regression.

Value

prediction.

See Also

train

[Package mlr version 0.3.732 Index]