predict.WrappedModel {mlr}	R Documentation

Predict new data.

Description

Predict the target variable of new data using a fitted model. What is stored exactly in the [Prediction] object depends on the predict.type setting of the Learner.

Usage

  ## S3 method for class 'WrappedModel'
 predict(object, task, newdata,
    subset, ...)

Arguments

object
[WrappedModel]
Wrapped model, result of train.

task
[SupervisedTask]
The task. If this is passed, data from this task is predicted.

newdata
[data.frame]
New observations which should be predicted. Pass this alternatively instead of task.

subset
[integer]
Index vector to subset task or newdata. Default is all data.

...
[any]
Currently ignored.

Value

[Prediction].

Note

To extract probabilities use getProbabilities.

Examples

## split iris data in training and test set
n <- nrow(iris)
mixed.set <- sample(1:n)
training.set <- mixed.set[1:(n/2)]
test.set <- mixed.set[(n/2 + 1):n]

## use linear discriminant analysis as learner for classification task
task <- makeClassifTask(data = iris, target = "Species")
learner <- makeLearner("classif.lda", method = "mle")
mod <- train(learner, task, subset = training.set)

## predict class labels for test data
pred <- predict(mod, newdata = iris[test.set,])
head(pred$data)

##        truth   response
## 1 versicolor versicolor
## 2  virginica  virginica
## 3  virginica  virginica
## 4  virginica  virginica
## 5     setosa     setosa
## 6 versicolor versicolor

## predict now probabiliies instead of class labels
learner <- makeLearner("classif.lda", method = "mle", predict.type = "prob")
mod <- train(learner, task, subset = training.set)
pred <- predict(mod, newdata = iris[test.set, ])
head(pred$data)

##          truth prob.setosa prob.versicolor prob.virginica   response
## 90  versicolor   4.407e-19       9.999e-01      1.206e-04 versicolor
## 102  virginica   1.058e-35       6.921e-04      9.993e-01  virginica
## 138  virginica   6.954e-33       2.522e-03      9.975e-01  virginica
## 106  virginica   1.338e-44       1.329e-06      1.000e+00  virginica
## 9       setosa   1.000e+00       5.297e-14      4.295e-32     setosa
## 100 versicolor   1.646e-17       9.999e-01      8.241e-05 versicolor

[Package mlr version 1.2 Index]