predict.WrappedModel {mlr}	R Documentation

Predict new data.

Description

Predict the target variable of new data using a fitted model. What is stored exactly in the [Prediction] object depends on the predict.type setting of the Learner. If predict.type was set to “prob” probability thresholding can be done calling the setThreshold function on the prediction object.

Usage

## S3 method for class 'WrappedModel'
predict(object, task, newdata, subset, ...)

Arguments

object
[WrappedModel]
Wrapped model, result of train.

task
[SupervisedTask]
The task. If this is passed, data from this task is predicted.

newdata
[data.frame]
New observations which should be predicted. Pass this alternatively instead of task.

subset
[integer]
Index vector to subset task or newdata. Default is all data.

...
[any]
Currently ignored.

Value

[Prediction].

Note

To extract probabilities use getProbabilities.

Examples

# train and predict
train.set = seq(1, 150, 2)
test.set = seq(2, 150, 2)
model = train("classif.lda", iris.task, subset = train.set)
p = predict(model, newdata = iris, subset = test.set)
print(p)

## Prediction:
## predict.type: response
## threshold: 
## time: 0.00
##    truth response
## 1 setosa   setosa
## 2 setosa   setosa
## 3 setosa   setosa
## 4 setosa   setosa
## 5 setosa   setosa
## 6 setosa   setosa

predict(model, task = iris.task, subset = test.set)

## Prediction:
## predict.type: response
## threshold: 
## time: 0.00
##   id  truth response
## 1  2 setosa   setosa
## 2  4 setosa   setosa
## 3  6 setosa   setosa
## 4  8 setosa   setosa
## 5 10 setosa   setosa
## 6 12 setosa   setosa

# predict now probabiliies instead of class labels
lrn = makeLearner("classif.lda", predict.type = "prob")
model = train(lrn, iris.task, subset = train.set)
p = predict(model, task = iris.task, subset = test.set)
print(p)

## Prediction:
## predict.type: prob
## threshold: setosa=0.33,versicolor=0.33,virginica=0.33
## time: 0.00
##    id  truth prob.setosa prob.versicolor prob.virginica response
## 2   2 setosa           1       1.080e-16      5.496e-36   setosa
## 4   4 setosa           1       3.942e-16      4.986e-35   setosa
## 6   6 setosa           1       1.101e-20      4.582e-40   setosa
## 8   8 setosa           1       1.281e-19      1.404e-39   setosa
## 10 10 setosa           1       3.465e-18      3.359e-38   setosa
## 12 12 setosa           1       1.929e-18      6.401e-38   setosa

getProbabilities(p)

##        setosa versicolor virginica
## 2   1.000e+00  1.080e-16 5.496e-36
## 4   1.000e+00  3.942e-16 4.986e-35
## 6   1.000e+00  1.101e-20 4.582e-40
## 8   1.000e+00  1.281e-19 1.404e-39
## 10  1.000e+00  3.465e-18 3.359e-38
## 12  1.000e+00  1.929e-18 6.401e-38
## 14  1.000e+00  9.175e-19 6.829e-39
## 16  1.000e+00  4.689e-27 2.746e-48
## 18  1.000e+00  4.645e-20 8.129e-40
## 20  1.000e+00  6.694e-22 5.219e-42
## 22  1.000e+00  1.067e-19 7.786e-39
## 24  1.000e+00  1.576e-13 1.372e-30
## 26  1.000e+00  1.645e-15 2.044e-34
## 28  1.000e+00  6.146e-21 2.497e-41
## 30  1.000e+00  1.659e-16 1.809e-35
## 32  1.000e+00  7.487e-18 9.890e-37
## 34  1.000e+00  2.671e-28 1.080e-50
## 36  1.000e+00  4.807e-20 2.339e-40
## 38  1.000e+00  3.813e-23 2.052e-44
## 40  1.000e+00  7.466e-20 6.277e-40
## 42  1.000e+00  3.675e-09 4.537e-26
## 44  1.000e+00  2.377e-14 3.535e-31
## 46  1.000e+00  4.173e-15 1.631e-33
## 48  1.000e+00  1.089e-17 4.864e-37
## 50  1.000e+00  1.774e-19 1.730e-39
## 52  4.334e-19  9.998e-01 2.451e-04
## 54  1.331e-22  9.997e-01 2.715e-04
## 56  1.996e-21  9.997e-01 2.860e-04
## 58  3.839e-14  1.000e+00 7.107e-08
## 60  1.501e-20  9.997e-01 3.122e-04
## 62  7.761e-20  9.996e-01 4.090e-04
## 64  2.081e-22  9.988e-01 1.171e-03
## 66  3.562e-17  1.000e+00 1.673e-05
## 68  3.771e-15  1.000e+00 1.680e-07
## 70  2.966e-17  1.000e+00 1.333e-06
## 72  6.044e-17  1.000e+00 5.494e-06
## 74  1.492e-20  9.999e-01 5.355e-05
## 76  2.932e-18  1.000e+00 3.462e-05
## 78  2.103e-26  8.507e-01 1.493e-01
## 80  5.539e-12  1.000e+00 1.079e-08
## 82  1.436e-15  1.000e+00 1.470e-07
## 84  2.971e-30  4.245e-01 5.755e-01
## 86  1.111e-19  9.983e-01 1.650e-03
## 88  1.466e-23  9.997e-01 3.307e-04
## 90  6.669e-21  9.999e-01 1.077e-04
## 92  7.536e-21  9.996e-01 4.137e-04
## 94  9.307e-15  1.000e+00 8.661e-08
## 96  3.008e-16  1.000e+00 3.390e-06
## 98  5.496e-18  1.000e+00 1.507e-05
## 100 1.363e-18  1.000e+00 2.825e-05
## 102 4.405e-37  2.115e-03 9.979e-01
## 104 6.525e-36  6.511e-03 9.935e-01
## 106 2.777e-46  4.861e-06 1.000e+00
## 108 3.911e-39  1.607e-03 9.984e-01
## 110 4.550e-45  4.040e-07 1.000e+00
## 112 6.618e-37  3.250e-03 9.968e-01
## 114 1.073e-40  1.951e-04 9.998e-01
## 116 3.838e-40  2.724e-05 1.000e+00
## 118 1.308e-40  2.435e-05 1.000e+00
## 120 1.969e-32  4.345e-01 5.655e-01
## 122 6.210e-37  1.069e-03 9.989e-01
## 124 2.371e-31  1.300e-01 8.700e-01
## 126 1.835e-33  2.731e-02 9.727e-01
## 128 5.723e-29  2.605e-01 7.395e-01
## 130 9.803e-30  5.519e-01 4.481e-01
## 132 8.666e-33  8.092e-03 9.919e-01
## 134 5.292e-27  9.386e-01 6.143e-02
## 136 2.222e-45  3.291e-06 1.000e+00
## 138 1.630e-32  3.699e-02 9.630e-01
## 140 3.685e-36  1.254e-03 9.987e-01
## 142 1.587e-37  2.050e-04 9.998e-01
## 144 1.670e-44  2.438e-06 1.000e+00
## 146 3.097e-40  4.264e-05 1.000e+00
## 148 1.441e-34  5.121e-03 9.949e-01
## 150 1.747e-31  6.121e-02 9.388e-01

[Package mlr version 1.2 Index]