predict.WrappedModel {mlr}	R Documentation

Predict new data.

Description

Predict the target variable of new data using a fitted model. What is stored exactly in the [Prediction] object depends on the predict.type setting of the Learner.

Usage

  ## S3 method for class 'WrappedModel'
 predict(object, task, newdata,
    subset, ...)

Arguments

object
[WrappedModel]
Wrapped model, result of train.

task
[SupervisedTask]
The task. If this is passed, data from this task is predicted.

newdata
[data.frame]
New observations which should be predicted. Pass this alternatively instead of task.

subset
[integer]
Index vector to subset task or newdata. Default is all data.

...
[any]
Currently ignored.

Value

[Prediction].

Note

To extract probabilities use getProbabilities.

Examples

## split iris data in training and test set
n <- nrow(iris)
mixed.set <- sample(1:n)
training.set <- mixed.set[1:(n/2)]
test.set <- mixed.set[(n/2 + 1):n]

## use linear discriminant analysis as learner for classification task
task <- makeClassifTask(data = iris, target = "Species")
learner <- makeLearner("classif.lda", method = "mle")
mod <- train(learner, task, subset = training.set)

## predict class labels for test data
pred <- predict(mod, newdata = iris[test.set,])
head(pred$data)

##        truth   response
## 1 versicolor versicolor
## 2 versicolor versicolor
## 3 versicolor versicolor
## 4     setosa     setosa
## 5 versicolor versicolor
## 6  virginica  virginica

## predict now probabiliies instead of class labels
learner <- makeLearner("classif.lda", method = "mle", predict.type = "prob")
mod <- train(learner, task, subset = training.set)
pred <- predict(mod, newdata = iris[test.set, ])
head(pred$data)

##          truth prob.setosa prob.versicolor prob.virginica   response
## 86  versicolor   6.752e-21       9.963e-01      3.687e-03 versicolor
## 53  versicolor   6.101e-23       9.941e-01      5.928e-03 versicolor
## 63  versicolor   7.413e-20       1.000e+00      4.183e-06 versicolor
## 35      setosa   1.000e+00       6.942e-18      8.434e-37     setosa
## 74  versicolor   1.263e-24       9.967e-01      3.305e-03 versicolor
## 121  virginica   3.210e-41       4.551e-05      1.000e+00  virginica

[Package mlr version 1.2 Index]