predictLearner {mlr}	R Documentation

Predict new data with an R learner.

Description

Mainly for internal use. Predict new data with a fitted model. You have to implement this method if you want to add another learner to this package.

Usage

  predictLearner(.learner, .model, .newdata, ...)

Arguments

.learner
[RLearner]
Wrapped learner.

.model
[WrappedModel]
Model produced by training.

.newdata
[data.frame]
New data to predict. Does not include target column.

...
[any]
Additional parameters, which need to be passed to the underlying predict function.

Details

You implementation must adhere to the following: The model must be fitted on the subset of .task given by .subset. All parameters in ... must be passed to the underlying training function.

Value

For classification: Either a factor for type “response” or a matrix for type “prob”. In the latter case the columns must be named with the class labels. For regressions: Either a numeric for type “response” or a matrix with two columns for type “se”. In the latter case first column is the estimated response (mean value) and the second column the estimated standard errors.

[Package mlr version 1.2 Index]