removeConstantFeatures {mlr}	R Documentation

Remove constant features from a data set.

Description

Constant features can lead to errors in some models and obviously provide no information in the training set that can be learned from. With the argument “perc”, there is a possibility to also remove features for which less than “perc” percent of the observations differ from the mode value.

Usage

removeConstantFeatures(x, perc = 0, dont.rm = character(0),
  na.mode = "na.rm", show.info = TRUE)

Arguments

x
[data.frame | SupervisedTask]
The data set or task.

perc
[numeric(1)]
The percentage of a feature values in [0, 1) that must differ from the mode value. Default is 0, which means only constant features with exactly one observed level are removed.

dont.rm
[character]
Names of the columns which must not be deleted. Default is no columns. Note that for a task the target column is always added to these so it is never removed.

na.mode
[character(1)]

na.rm: NAs will be ignored in the percentage calculation.

single: NAs will be seen as a single, extra level in the percentage calculation.

distinct: NAs will be seen as multiple extra levels in the percentage calculation.

show.info
[logical(1)]
Verbose output on console? Default is TRUE.

Value

[data.frame | SupervisedTask].

[Package mlr version 1.2 Index]