resample {mlr}	R Documentation

Fit models according to a resampling strategy.

Description

Given a resampling strategy, which defines sets of training and test indices, fits the selected learner using the training sets and performs predictions for the training/test sets. (This depends on what you selected in the resampling strategy, see parameter predict in makeResampleDesc.) Then performance measures are calculated and aggregated. You are able to return all fitted models (parameter models) or extract specific parts of the models (parameter extract) as returning all of them completely might be memory intensive.

Usage

  resample(learner, task, resampling, measures, weights,
    models = FALSE, extract = function(model) { },
    show.info = TRUE)

Arguments

learner
[Learner]
The learner.

task
[SupervisedTask]
The task.

resampling
[ResampleDesc or ResampleInstance]
Resampling strategy. If a description is passed, it is instantiated automatically.

measures
[Measure | list of Measure]
Performance measure(s) to evaluate.

weights
[numeric]
Optional, non-negative case weight vector to be used during fitting. If given, must be of same length as observations in task and in corresponding order. Overwrites weights specified in the task ([SupervisedTask]). By default missing which means no weights are used unless specified in the task.

models
[logical(1)]
Should all fitted models be returned? Default is FALSE.

extract
[function(model)]
Function used to extract information from a fitted model during resampling. Is applied to every WrappedModel resulting from calls to train during resampling. Default is to extract nothing.

show.info
[logical(1)]
Should a few informative lines about the current resampling iteration and the result be logged to the R console? Default is TRUE.

Details

For construction of the resampling strategies use the factory methods makeResampleDesc and makeResampleInstance.

Value

List of:

measures.test [data.frame]
Rows correspond to test sets in resampling iterations, columns to performance measures.

measures.train [data.frame]
Rows correspond to training sets in resampling iterations, columns to performance measures.

aggr [named numeric]
Vector of aggregated performance values. Names are coded like this <measure>.<aggregation>.

pred [ResamplePrediction]
Container for all predictions during resampling.

models [list of WrappedModel]
List of fitted models or NULL.

extract [list]
List of extracted parts from fitted models or NULL.

See Also

makeResampleDesc, makeResampleInstance

Examples

task <- makeClassifTask(data = iris, target = "Species")
rdesc <- makeResampleDesc("Bootstrap", iters = 10)
rin <- makeResampleInstance(rdesc, task = task)
r1 <- resample(makeLearner("classif.qda"), task, rin)

## [Resample] OOB bootstrapping iter: 1
## [Resample] OOB bootstrapping iter: 2
## [Resample] OOB bootstrapping iter: 3
## [Resample] OOB bootstrapping iter: 4
## [Resample] OOB bootstrapping iter: 5
## [Resample] OOB bootstrapping iter: 6
## [Resample] OOB bootstrapping iter: 7
## [Resample] OOB bootstrapping iter: 8
## [Resample] OOB bootstrapping iter: 9
## [Resample] OOB bootstrapping iter: 10
## [Resample] Result: mmce.test.mean=0.0334

print(r1$measures.test)

##    iter    mmce
## 1     1 0.01754
## 2     2 0.05085
## 3     3 0.05455
## 4     4 0.08929
## 5     5 0.01786
## 6     6 0.01818
## 7     7 0.00000
## 8     8 0.03390
## 9     9 0.03333
## 10   10 0.01887

print(r1$aggr)

## mmce.test.mean 
##        0.03344

r2 <- resample(makeLearner("classif.rpart"), task, rin)

## [Resample] OOB bootstrapping iter: 1
## [Resample] OOB bootstrapping iter: 2
## [Resample] OOB bootstrapping iter: 3
## [Resample] OOB bootstrapping iter: 4
## [Resample] OOB bootstrapping iter: 5
## [Resample] OOB bootstrapping iter: 6
## [Resample] OOB bootstrapping iter: 7
## [Resample] OOB bootstrapping iter: 8
## [Resample] OOB bootstrapping iter: 9
## [Resample] OOB bootstrapping iter: 10
## [Resample] Result: mmce.test.mean=0.0677

print(r2$measures.test)

##    iter    mmce
## 1     1 0.08772
## 2     2 0.06780
## 3     3 0.05455
## 4     4 0.07143
## 5     5 0.07143
## 6     6 0.07273
## 7     7 0.07692
## 8     8 0.06780
## 9     9 0.05000
## 10   10 0.05660

print(r2$aggr)

## mmce.test.mean 
##         0.0677

[Package mlr version 1.2 Index]