crossval {mlr}	R Documentation

Fit models according to a resampling strategy.

Description

resample: Given a resampling strategy, which defines sets of training and test indices, fits the selected learner using the training sets and performs predictions for the training/test sets. This depends on what you selected in the resampling strategy, see parameter predict in makeResampleDesc.

Then performance measures are calculated on all respective data sets and aggregated.

You are able to return all fitted models (parameter models) or extract specific parts of the models (parameter extract) as returning all of them completely might be memory intensive.

For construction of the resampling strategies use the factory methods makeResampleDesc and makeResampleInstance.

The remaining functions on this page are convenience wrappers for the various existing resampling strategies.

Usage

crossval(learner, task, iters = 10L, stratify = FALSE, measures,
  models = FALSE, ..., show.info = TRUE)

repcv(learner, task, folds = 10L, reps = 10L, stratify = FALSE, measures,
  models = FALSE, ..., show.info = TRUE)

holdout(learner, task, split = 2/3, stratify = FALSE, measures,
  models = FALSE, ..., show.info = TRUE)

subsample(learner, task, iters = 30, split = 2/3, stratify = FALSE,
  measures, models = FALSE, ..., show.info = TRUE)

bootstrapOOB(learner, task, iters = 30, stratify = FALSE, measures,
  models = FALSE, ..., show.info = TRUE)

bootstrapB632(learner, task, iters = 30, stratify = FALSE, measures,
  models = FALSE, ..., show.info = TRUE)

bootstrapB632plus(learner, task, iters = 30, stratify = FALSE, measures,
  models = FALSE, ..., show.info = TRUE)

resample(learner, task, resampling, measures, weights = NULL,
  models = FALSE, extract, show.info = getMlrOption("show.info"))

Arguments

learner
[Learner | character(1)]
The learner. If you pass a string the learner will be created via makeLearner.

task
[SupervisedTask]
The task.

resampling
[ResampleDesc or ResampleInstance]
Resampling strategy. If a description is passed, it is instantiated automatically.

iters
[integer(1)]
See ResampleDesc.

folds
[integer(1)]
See ResampleDesc.

reps
[integer(1)]
See ResampleDesc.

split
[numeric(1)]
See ResampleDesc.

stratify
[logical(1)]
See ResampleDesc.

measures
[Measure | list of Measure]
Performance measure(s) to evaluate.

weights
[numeric]
Optional, non-negative case weight vector to be used during fitting. If given, must be of same length as observations in task and in corresponding order. Overwrites weights specified in the task. By default NULL which means no weights are used unless specified in the task.

models
[logical(1)]
Should all fitted models be returned? Default is FALSE.

extract
[function]
Function used to extract information from a fitted model during resampling. Is applied to every WrappedModel resulting from calls to train during resampling. Default is to extract nothing.

...
[any]
Further hyperparameters passed to learner.

show.info
[logical(1)]
Print verbose output on console? Default is set via configureMlr.

Value

List of:

measures.test [data.frame]
Gives you access to performance measurements on the individual test sets. Rows correspond to sets in resampling iterations, columns to performance measures.

measures.train [data.frame]
Gives you access to performance measurements on the individual training sets. Rows correspond to sets in resampling iterations, columns to performance measures. Usually not available, only if specifically requested, see general description above.

aggr [numeric]
Named vector of aggregated performance values. Names are coded like this <measure>.<aggregation>.

pred [ResamplePrediction]
Container for all predictions during resampling.

models [list of WrappedModel]
List of fitted models or NULL.

extract [list]
List of extracted parts from fitted models or NULL.

See Also

Other resample: ResampleDesc, makeResampleDesc; ResampleInstance, makeResampleInstance; ResamplePrediction

Examples

task = makeClassifTask(data = iris, target = "Species")
rdesc = makeResampleDesc("CV", iters = 2)
r = resample(makeLearner("classif.qda"), task, rdesc)

## [Resample] cross-validation iter: 1
## [Resample] cross-validation iter: 2
## [Resample] Result: mmce.test.mean=0.0467

print(r$aggr)

## mmce.test.mean 
##        0.04667

print(r$measures.test)

##   iter    mmce
## 1    1 0.04000
## 2    2 0.05333

print(r$pred)

## Resampled Prediction for:
## Resample description: cross-validation with 2 iterations.
## Predict: test
## Stratification: FALSE
## predict.type: response
## threshold: 
## time (mean): 0.00
##   id  truth response iter  set
## 1  3 setosa   setosa    1 test
## 2  4 setosa   setosa    1 test
## 3  5 setosa   setosa    1 test
## 4  6 setosa   setosa    1 test
## 5 10 setosa   setosa    1 test
## 6 12 setosa   setosa    1 test

[Package mlr version 2.0 Index]