resample.desc-class {mlr}	R Documentation

resample.desc

Description

Base class for description of resampling algorithms. A description of a resampling algorithm contains all necessary information to provide a resampling.instance, when given the size of the data set.

Details

resample.desc-class: Getter.

instance.class [character]

S4 class name of the corresponding resample.instance

name [character]

Name of this resampling algorithm

iters [numeric]

Number of iterations. Note that this the complete number of generated train/test sets, so for a 10 times repeated 5fold cross-validation it would be 50.

has.groups [boolean]

Is special grouping used for predictions of a iteration in order to aggregate them differently?

Extends

object

Methods

[

to.string

make.res.instance

make.res.instance

[Package mlr version 0.3.1095 Index]