resample.desc-class {mlr}	R Documentation

resample.desc

Description

Base class for description of resampling algorithms. A description of a resampling algorithm contains all necessary information to provide a resampling.instance, when given the size of the data set.

Details

Getter.

instance.class [character]

S4 class name of the corresponding resample.instance

name [character]

Name of this resampling algorithm

iters [numeric]

Number of iterations

Extends

object

Methods

to.string

make.res.instance

make.res.instance

[Package mlr version 0.3.777 Index]