selectFeatures {mlr}	R Documentation

Feature selection by wrapper approach.

Description

Optimizes the features for a classification or regression problem by choosing a variable selection wrapper approach. Allows for different optimization methods, such as forward search or a genetic algorithm. You can select such an algorithm (and its settings) by passing a corresponding control object. For a complete list of implemented algorithms look at the subclasses of [FeatSelControl].

Usage

  selectFeatures(learner, task, resampling, control,
    measures, bit.names, bits.to.features,
    show.info = TRUE)

Arguments

learner
[Learner]
The learner.

task
[SupervisedTask]
The task.

resampling
[ResampleInstance | ResampleDesc]
Resampling strategy to feature sets. If you pass a description, it is instantiated once at the beginning by default, so all points are evaluated on the same training/test sets. If you want to change that behaviour, look at FeatSelControl.

control
[see FeatSelControl] Control object for search method. Also selects the optimization algorithm for feature selection.

measures
[list of Measure]
Performance measures to evaluate. The first measure, aggregated by the first aggregation function is optimized during selection, others are simply evaluated.

bit.names
[character]
Names of bits encoding the solutions. Also defines the total number of bits in the encoding. Per default these are the feature names of the task.

bits.to.features
[function(x, task)]
Function which transforms an integer-0-1 vector into a character vector of selected features. Per default a value of 1 in the ith bit selects the ith feature to be in the candidate solution.

show.info
[logical(1)]
A logical value, indicating whether information should be printed. The default is TRUE.

Details

All algorithms operate on a 0-1-bit encoding of candidate solutions. Per default a single bit corresponds to a single feature, but you are able to change this by using the arguments bit.names and bits.to.features. Thus allowing you to switch on whole groups of features with a single bit.

Value

[FeatSelResult].

Examples

task <- makeClassifTask(data=iris, target="Species")
lrn <- makeLearner("classif.rpart")
rdesc <- makeResampleDesc("Holdout")

## Now create control-objects for each of the possible feature selection algorithms:
ctrlSeq <- makeFeatSelControlSequential(method="sfs", maxit=NA)
ctrlGA <- makeFeatSelControlGA(maxit=5, max.features=NA, crossover.rate=0.5,
  mutation.rate=0.1, mu=10, lambda=5)
ctrlRand <- makeFeatSelControlRandom(maxit=10, max.features=NA, prob=0.5)
ctrlExh <- makeFeatSelControlExhaustive(maxit=NA, max.features=NA)

## Let's run the feature selction algorithm:

sfSeq <- selectFeatures(lrn, task, rdesc, control=ctrlSeq)

## [selectFeatures] 1: 0 bits: mmce.test.mean=0.66
## [selectFeatures] 2: 1 bits: mmce.test.mean=0.28
## [selectFeatures] 2: 1 bits: mmce.test.mean= 0.4
## [selectFeatures] 2: 1 bits: mmce.test.mean=0.08
## [selectFeatures] 2: 1 bits: mmce.test.mean=0.02
## [selectFeatures] 3: 2 bits: mmce.test.mean=0.02
## [selectFeatures] 3: 2 bits: mmce.test.mean=0.02
## [selectFeatures] 3: 2 bits: mmce.test.mean=0.08

sfSeq

## FeatSel result:
## Features (1): Petal.Width
## =0.02

sfGA <- selectFeatures(lrn, task, rdesc, control=ctrlGA)

## [selectFeatures] 1: 2 bits: mmce.test.mean=0.04
## [selectFeatures] 1: 3 bits: mmce.test.mean=0.06
## [selectFeatures] 1: 1 bits: mmce.test.mean=0.44
## [selectFeatures] 1: 0 bits: mmce.test.mean=0.52
## [selectFeatures] 1: 1 bits: mmce.test.mean=0.06
## [selectFeatures] 1: 1 bits: mmce.test.mean=0.04
## [selectFeatures] 1: 2 bits: mmce.test.mean=0.04
## [selectFeatures] 1: 2 bits: mmce.test.mean=0.26
## [selectFeatures] 1: 2 bits: mmce.test.mean=0.06
## [selectFeatures] 1: 2 bits: mmce.test.mean=0.06
## [selectFeatures] 1: 1 bits: mmce.test.mean=0.28
## [selectFeatures] 1: 2 bits: mmce.test.mean=0.06
## [selectFeatures] 1: 2 bits: mmce.test.mean=0.26
## [selectFeatures] 1: 2 bits: mmce.test.mean=0.04
## [selectFeatures] 1: 1 bits: mmce.test.mean=0.04
## [selectFeatures] 2: 3 bits: mmce.test.mean=0.06
## [selectFeatures] 2: 2 bits: mmce.test.mean=0.06
## [selectFeatures] 2: 2 bits: mmce.test.mean=0.06
## [selectFeatures] 2: 1 bits: mmce.test.mean=0.44
## [selectFeatures] 2: 3 bits: mmce.test.mean=0.06
## [selectFeatures] 3: 2 bits: mmce.test.mean=0.06
## [selectFeatures] 3: 3 bits: mmce.test.mean=0.06
## [selectFeatures] 3: 1 bits: mmce.test.mean=0.04
## [selectFeatures] 3: 1 bits: mmce.test.mean=0.04
## [selectFeatures] 3: 3 bits: mmce.test.mean=0.06
## [selectFeatures] 4: 2 bits: mmce.test.mean=0.04
## [selectFeatures] 4: 2 bits: mmce.test.mean=0.06
## [selectFeatures] 4: 1 bits: mmce.test.mean=0.04
## [selectFeatures] 4: 0 bits: mmce.test.mean=0.62
## [selectFeatures] 4: 2 bits: mmce.test.mean=0.04
## [selectFeatures] 5: 2 bits: mmce.test.mean=0.04
## [selectFeatures] 5: 1 bits: mmce.test.mean=0.04
## [selectFeatures] 5: 2 bits: mmce.test.mean=0.04
## [selectFeatures] 5: 1 bits: mmce.test.mean=0.04
## [selectFeatures] 5: 1 bits: mmce.test.mean=0.04

sfGA

## FeatSel result:
## Features (2): Sepal.Width, Petal.Length
## =0.04

sfRand <- selectFeatures(lrn, task, rdesc, control=ctrlRand)

## [selectFeatures] 1: 3 bits: mmce.test.mean=0.04
## [selectFeatures] 1: 2 bits: mmce.test.mean=0.04
## [selectFeatures] 1: 1 bits: mmce.test.mean=0.24
## [selectFeatures] 1: 1 bits: mmce.test.mean= 0.6
## [selectFeatures] 1: 3 bits: mmce.test.mean=0.04
## [selectFeatures] 1: 1 bits: mmce.test.mean= 0.6
## [selectFeatures] 1: 2 bits: mmce.test.mean=0.04
## [selectFeatures] 1: 2 bits: mmce.test.mean=0.04
## [selectFeatures] 1: 3 bits: mmce.test.mean=0.04
## [selectFeatures] 1: 2 bits: mmce.test.mean=0.02

sfRand

## FeatSel result:
## Features (2): Sepal.Width, Petal.Width
## =0.02

sfExh <- selectFeatures(lrn, task, rdesc, control=ctrlExh)

## [selectFeatures] 1: 0 bits: mmce.test.mean=0.74
## [selectFeatures] 1: 1 bits: mmce.test.mean= 0.3
## [selectFeatures] 1: 1 bits: mmce.test.mean=0.42
## [selectFeatures] 1: 1 bits: mmce.test.mean=0.02
## [selectFeatures] 1: 1 bits: mmce.test.mean=0.06
## [selectFeatures] 1: 2 bits: mmce.test.mean= 0.3
## [selectFeatures] 1: 2 bits: mmce.test.mean=0.02
## [selectFeatures] 1: 2 bits: mmce.test.mean=0.06
## [selectFeatures] 1: 2 bits: mmce.test.mean=0.02
## [selectFeatures] 1: 2 bits: mmce.test.mean=0.06
## [selectFeatures] 1: 2 bits: mmce.test.mean=0.06
## [selectFeatures] 1: 3 bits: mmce.test.mean=0.02
## [selectFeatures] 1: 3 bits: mmce.test.mean=0.06
## [selectFeatures] 1: 3 bits: mmce.test.mean=0.06
## [selectFeatures] 1: 3 bits: mmce.test.mean=0.06
## [selectFeatures] 1: 4 bits: mmce.test.mean=0.06

sfExh

## FeatSel result:
## Features (1): Petal.Length
## =0.02

[Package mlr version 1.2 Index]