setThreshold {mlr}	R Documentation

Set threshold of prediction object.

Description

Set threshold of prediction object for classification. Creates corresponding discrete class response for the newly set threshold. For binary classification: The positive class is predicted if the probability value exceeds the threshold. For multiclass: Probabilities are divided by corresponding thresholds and the class with maximum resulting value is selected. The result of both are equivalent if in the multi-threshold case the values are greater than 0 and sum to 1.

Usage

  setThreshold(pred, threshold)

Arguments

pred
[Prediction]
The prediction.

threshold
[numeric]
Threshold to produce class labels. Has to be a named vector, where names correspond to class labels. Only if pred is a prediction object resulting from binary classification it can be a single numerical threshold for the positive class.

Value

[Prediction] with changed threshold and corresponding response.

See Also

predict.WrappedModel

Examples

## create task and train learner (LDA)
task <- makeClassifTask(data = iris, target = "Species")
lrn <- makeLearner("classif.lda", predict.type = "prob")
mod <- train(lrn, task)

## predict probabilities and compute performance
pred <- predict(mod, newdata = iris)
performance(pred, measures = mmce)

## [1] 0.02

head(as.data.frame(pred))

##    truth prob.setosa prob.versicolor prob.virginica response
## 1 setosa           1       3.896e-22      2.611e-42   setosa
## 2 setosa           1       7.218e-18      5.042e-37   setosa
## 3 setosa           1       1.464e-19      4.676e-39   setosa
## 4 setosa           1       1.269e-16      3.567e-35   setosa
## 5 setosa           1       1.637e-22      1.083e-42   setosa
## 6 setosa           1       3.883e-21      4.567e-40   setosa

## adjust threshold and predict probabilities again
threshold <- c(setosa = 0.4, versicolor = 0.3, virginica = 0.3)
pred <- setThreshold(pred, threshold = threshold)
performance(pred, measures = mmce)

## [1] 0.02

head(as.data.frame(pred))

##    truth prob.setosa prob.versicolor prob.virginica response
## 1 setosa           1       3.896e-22      2.611e-42   setosa
## 2 setosa           1       7.218e-18      5.042e-37   setosa
## 3 setosa           1       1.464e-19      4.676e-39   setosa
## 4 setosa           1       1.269e-16      3.567e-35   setosa
## 5 setosa           1       1.637e-22      1.083e-42   setosa
## 6 setosa           1       3.883e-21      4.567e-40   setosa

[Package mlr version 1.2 Index]