smote {mlr}	R Documentation

Synthetic Minority Oversampling Technique to handle class imbalancy in binary classification.

Description

In each iteration, samples one minority class element x1, then one of x1's nearest neighbors: x2. Both points are now interpolated / convex-combined, resulting in a new virtual data point x3 for the minority class.

The method handles factor features, too. The gower distance is used for nearest neighbor calculation, see daisy. For interpolation, the new factor level for x3 is sampled from the two given levels of x1 and x2 per feature.

Usage

smote(task, rate, nn = 5L)

Arguments

task
[Task]
The task.

rate
[numeric(1)]
Factor to upsample the smaller class. Must be between 1 and Inf, where 1 means no oversampling and 2 would mean doubling the class size.

nn
[integer(1)]
Number of nearest neighbors to consider. Default is 5.

Value

[Task].

See Also

Other imbalancy: makeOverBaggingWrapper; makeOversampleWrapper, makeUndersampleWrapper; oversample, undersample

[Package mlr version 2.1 Index]