Train a learning algorithm.

Given a Task, creates a model for the learning machine which can be used for predictions on new data.

train(learner, task, subset, weights = NULL)

Arguments

learner
[Learner | character(1)] The learner. If you pass a string the learner will be created via makeLearner.

task
[Task] The task.

subset
[integer | logical] Selected cases. Either a logical or an index vector. By default all observations are used.

weights
[numeric] Optional, non-negative case weight vector to be used during fitting. If given, must be of same length as subset and in corresponding order. By default NULL which means no weights are used unless specified in the task (Task). Weights from the task will be overwritten.

Value

[WrappedModel].

See also

predict.WrappedModel

Examples

training.set = sample(seq_len(nrow(iris)), nrow(iris) / 2)

## use linear discriminant analysis to classify iris data
task = makeClassifTask(data = iris, target = "Species")
learner = makeLearner("classif.lda", method = "mle")
mod = train(learner, task, subset = training.set)
print(mod)

#> Model for learner.id=classif.lda; learner.class=classif.lda
#> Trained on: task.id = iris; obs = 75; features = 4
#> Hyperparameters: method=mle

## use random forest to classify iris data
task = makeClassifTask(data = iris, target = "Species")
learner = makeLearner("classif.rpart", minsplit = 7, predict.type = "prob")
mod = train(learner, task, subset = training.set)
print(mod)

#> Model for learner.id=classif.rpart; learner.class=classif.rpart
#> Trained on: task.id = iris; obs = 75; features = 4
#> Hyperparameters: xval=0,minsplit=7