train {mlr}	R Documentation

Train a learning algorithm.

Description

Given a SupervisedTask, creates a model for the learning machine which can be used for predictions on new data.

Usage

train(learner, task, subset, weights = NULL)

Arguments

learner
[Learner | character(1)]
The learner. If you pass a string the learner will be created via makeLearner.

task
[SupervisedTask]
The task.

subset
[integer]
An index vector specifying the training cases to be used for fitting. By default the complete data set is used.

weights
[numeric]
Optional, non-negative case weight vector to be used during fitting. If given, must be of same length as subset and in corresponding order. By default NULL which means no weights are used unless specified in the task ([SupervisedTask]). Weights from the task will be overwritten.

Value

[WrappedModel].

See Also

predict.WrappedModel

Examples

training.set = sample(1:nrow(iris), nrow(iris) / 2)

## use linear discriminant analysis to classify iris data
task = makeClassifTask(data = iris, target = "Species")
learner = makeLearner("classif.lda", method = "mle")
mod = train(learner, task, subset = training.set)
print(mod)

## Model for id = classif.lda class = classif.lda
## Trained on obs: 75
## Used features: 4
## Hyperparameters: method=mle

## use random forest to classify iris data
task = makeClassifTask(data = iris, target = "Species")
learner = makeLearner("classif.rpart", minsplit = 7, predict.type = "prob")
mod = train(learner, task, subset = training.set)
print(mod)

## Model for id = classif.rpart class = classif.rpart
## Trained on obs: 75
## Used features: 4
## Hyperparameters: xval=0,minsplit=7

[Package mlr version 2.0 Index]