tuneParams {mlr}	R Documentation

Hyperparameter tuning.

Description

Optimizes the hyperparameters of a learner for a classification or regression problem. Allows for different optimization methods, such as grid search, evolutionary strategies, etc. You can select such an algorithm (and its settings) by passing a corresponding control object. For a complete list of implemented algorithms look at TuneControl.

Usage

  tuneParams(learner, task, resampling, measures, par.set,
    control, show.info = TRUE)

Arguments

learner
[Learner]
The learner.

task
[SupervisedTask]
The task.

resampling
[ResampleInstance | ResampleDesc]
Resampling strategy to evaluate points in hyperparameter space. If you pass a description, it is instantiated once at the beginning by default, so all points are evaluated on the same training/test sets. If you want to change that behaviour, look at TuneControl.

par.set
[ParamSet]
Collection of parameters and their constraints for optimization.

control
[TuneControl]
Control object for search method. Also selects the optimization algorithm for tuning.

measures
[list of Measure]
Performance measures to evaluate. The first measure, aggregated by the first aggregation function is optimized during tuning, others are simply evaluated.

show.info
[logical(1)]
Show info message after each hyperparameter evaluation? Default is TRUE.

Details

Note that if tranformations are associated with the parameters, the returned result will contain transformed values in the optimal result and the path.

Value

[TuneResult].

[Package mlr version 1.2 Index]