tuneParams {mlr}	R Documentation

Hyperparameter tuning.

Description

Optimizes the hyperparameters of a learner. Allows for different optimization methods, such as grid search, evolutionary strategies, iterated F-race, etc. You can select such an algorithm (and its settings) by passing a corresponding control object. For a complete list of implemented algorithms look at TuneControl.

Multi-criteria tuning can be done with tuneParamsMultiCrit.

Usage

tuneParams(learner, task, resampling, measures, par.set, control,
  show.info = getMlrOption("show.info"))

Arguments

learner
[Learner | character(1)]
The learner. If you pass a string the learner will be created via makeLearner.

task
[Task]
The task.

resampling
[ResampleInstance | ResampleDesc]
Resampling strategy to evaluate points in hyperparameter space. If you pass a description, it is instantiated once at the beginning by default, so all points are evaluated on the same training/test sets. If you want to change that behavior, look at TuneControl.

measures
[list of Measure | Measure]
Performance measures to evaluate. The first measure, aggregated by the first aggregation function is optimized, others are simply evaluated.

par.set
[ParamSet]
Collection of parameters and their constraints for optimization.

control
[TuneControl]
Control object for search method. Also selects the optimization algorithm for tuning.

show.info
[logical(1)]
Print verbose output on console? Default is set via configureMlr.

Value

[TuneResult].

See Also

Other tune: ModelMultiplexer, makeModelMultiplexer; TuneControl, TuneControlCMAES, TuneControlGenSA, TuneControlGrid, TuneControlIrace, TuneControlRandom, makeTuneControlCMAES, makeTuneControlGenSA, makeTuneControlGrid, makeTuneControlIrace, makeTuneControlRandom; getTuneResult; makeModelMultiplexerParamSet; makeTuneWrapper; tuneThreshold

Examples

# a grid search for an SVM (with a tiny number of points...)
# note how easily we can optimize on a log-scale
ps = makeParamSet(
  makeNumericParam("C", lower = -12, upper = 12, trafo = function(x) 2^x),
  makeNumericParam("sigma", lower = -12, upper = 12, trafo = function(x) 2^x)
)
ctrl = makeTuneControlGrid(resolution = 2L)
rdesc = makeResampleDesc("CV", iters = 2L)
res = tuneParams("classif.ksvm", iris.task, rdesc, par.set = ps, control = ctrl)

## [Tune] Started tuning learner classif.ksvm for parameter set:
##          Type len Def    Constr Req Trafo
## C     numeric   -   - -12 to 12   -     Y
## sigma numeric   -   - -12 to 12   -     Y
## With control class: TuneControlGrid
## Imputation value: 1
## [Tune] 1: C=0.000244; sigma=0.000244 : mmce.test.mean=0.72
## [Tune] 2: C=4.1e+03; sigma=0.000244 : mmce.test.mean=0.04
## [Tune] 3: C=0.000244; sigma=4.1e+03 : mmce.test.mean=0.72
## [Tune] 4: C=4.1e+03; sigma=4.1e+03 : mmce.test.mean=0.707
## [Tune] Result: C=4.1e+03; sigma=0.000244 : mmce.test.mean=0.04

print(res)

## Tune result:
## Op. pars: C=4.1e+03; sigma=0.000244
## mmce.test.mean=0.04

print(as.data.frame(res$opt.path))

##     C sigma mmce.test.mean dob eol error.message exec.time
## 1 -12   -12         0.7200   1  NA          <NA>     0.038
## 2  12   -12         0.0400   2  NA          <NA>     0.034
## 3 -12    12         0.7200   3  NA          <NA>     0.036
## 4  12    12         0.7067   4  NA          <NA>     0.036

print(as.data.frame(trafoOptPath(res$opt.path)))

##           C     sigma mmce.test.mean dob eol
## 1 2.441e-04 2.441e-04         0.7200   1  NA
## 2 4.096e+03 2.441e-04         0.0400   2  NA
## 3 2.441e-04 4.096e+03         0.7200   3  NA
## 4 4.096e+03 4.096e+03         0.7067   4  NA

## Not run: 
##D # we optimize the SVM over 3 kernels simultanously
##D # note how we use dependent params (requires = ...) and iterated F-racing here
##D ps = makeParamSet(
##D   makeNumericParam("C", lower = -12, upper = 12, trafo = function(x) 2^x),
##D   makeDiscreteParam("kernel", values = c("vanilladot", "polydot", "rbfdot")),
##D   makeNumericParam("sigma", lower = -12, upper = 12, trafo = function(x) 2^x,
##D     requires = quote(kernel == "rbfdot")),
##D   makeIntegerParam("degree", lower = 2L, upper = 5L,
##D     requires = quote(kernel == "polydot"))
##D )
##D print(ps)
##D ctrl = makeTuneControlIrace(maxExperiments = 200L)
##D rdesc = makeResampleDesc("Holdout")
##D res = tuneParams("classif.ksvm", iris.task, rdesc, par.set = ps, control = ctrl)
##D print(res)
##D print(head(as.data.frame(res$opt.path)))
## End(Not run)

[Package mlr version 2.1 Index]