tuneThreshold {mlr}	R Documentation

Tune prediction threshold.

Description

Optimizes the threshold of prediction based on probabilities. Uses optimizeSubInts for 2class problems and cma_es for multiclass problems.

Usage

tuneThreshold(pred, measure, task, model, nsub = 20L, control = list())

Arguments

pred
[Prediction]
Prediction object.

measure
[Measure]
Performance measure to optimize.

task
[SupervisedTask]
Learning task. Rarely neeeded, only when required for the performance measure.

model
[WrappedModel]
Fitted model. Rarely neeeded, only when required for the performance measure.

nsub
[integer(1)]
Passed to optimizeSubInts for 2class problems. Default is 20.

control
[list]
Control object for cma_es when used. Default is empty list.

Value

[list]. A named list with with the following components: th is the optimal threshold, perf the performance value.

See Also

Other tune: ModelMultiplexer, makeModelMultiplexer; TuneControl, TuneControlCMAES, TuneControlGrid, TuneControlIrace, TuneControlOptim, TuneControlRandom, makeTuneControlCMAES, makeTuneControlGenSA, makeTuneControlGrid, makeTuneControlIrace, makeTuneControlRandom; getTuneResult; makeModelMultiplexerParamSet; makeTuneWrapper; tuneParams

[Package mlr version 2.0 Index]