wrapped.model-class {mlr}	R Documentation

Induced model of learner.

Description

Result from train. It internally stores the underlying fitted model, the IDs of the subset used for training, variables used for training and information about second-level optimization like tuned hyperparameters or selected variables.

Details

wrapped.model-class: Getter.

learner [learner]

Learner that was used to fit the model.

learner model [any]

Underlying model from used R package.

subset [integer]

Subset used for training.

vars [character]

Variables used for training.

hyper.pars [list]

List of fixed hyperparameters and respective values for this model.

hyper.names [character]

Names of used hyperparameters.

hyper.types [character]

For which step in the model building process the respective hyperparameters used? Named character vector.

fail [NULL | string]

Generally NULL but if the training failed, the error message of the underlying train function.

opt [path.element]

Optimum of second-level optimization.

path [list of path.elements]

Path of second-level optimization.

tuned.par [list]

If tuning was performed, best found set of hyperparameters.

tuned.perf [numeric]

If tuning was performed, performance of best found set of hyperparameters.

sel.vars [character]

If variable selection was performed, best found set of variables.

sel.perf [numeric]

If variable selection was performed, performance of best found set of variables.

Extends

object

Methods

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

predict

signature(object = "wrapped.model"): Predict the target variable of new data using a fitted model. If the type is set to "prob" or "decision" probabilities or decision values will be stored in the resulting object. The resulting class labels are the classes with the maximum values or thresholding can also be used.

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

to.string

[

signature( = "wrapped.model"): Getter.

[Package mlr version 0.3.777 Index]