wrapped.model-class {mlr}	R Documentation

Induced model of learner.

Description

Result from train. It internally stores the underlying fitted model, the IDs of the subset used for thraining, variables used for training and information about second-level optimization like tuned hyperparameters or selected variables.

Details

wrapped.model-class: Getter.

wrapped.learner [wrapped.learner]

Wrapped learner that was used to fit the model.

learner model [any]

Undelying model from used R package.

subset [integer]

Subset used for training.

vars [character]

Variables used for training.

parset [list]

Hyperparameters used for training.

fail [NULL | string]

Generally NULL but if the training failed, the error message of the underlying train function.

opt [path.element]

Optimum of second-level optimization.

path [list of path.elements]

Path of second-level optimization.

tuned.par [list]

If tuning was performed, best found set of hyperparameters.

tuned.perf [numeric]

If tuning was performed, performance of best found set of hyperparameters.

sel.vars [character]

If variable selection was performed, best found set of variables.

sel.perf [numeric]

If variable selection was performed, performance of best found set of variables.

Extends

object

Methods

predict.learner

predict.learner

predict.learner

predict.learner

predict.learner

predict.learner

predict.learner

predict.learner

predict.learner

predict.learner

predict.learner

predict.learner

predict.learner

predict.learner

predict.learner

predict.learner

predict.learner

predict.learner

predict.learner

predict.learner

predict

signature(object = "wrapped.model"): Predict the target variable of new data using a fitted model.

predict.learner

predict.learner

predict.learner

predict.learner

predict.learner

to.string

[

signature( = "wrapped.model"): Getter.

[Package mlr version 0.3.528 Index]